Improved precision of Log; Fixed benchmarks for Log and Exp
