+++ approx.go
@@ -1,4102 +1,4103 @@
 package fixed
 
 var logApprox = []int64{
-	/*log(0.00024)*/ -34069,
-	/*log(0.00049)*/ -31230,
-	/*log(0.00073)*/ -29569,
-	/*log(0.00098)*/ -28391,
-	/*log(0.00122)*/ -27477,
-	/*log(0.00146)*/ -26730,
-	/*log(0.00171)*/ -26099,
-	/*log(0.00195)*/ -25552,
-	/*log(0.00220)*/ -25069,
-	/*log(0.00244)*/ -24638,
-	/*log(0.00269)*/ -24247,
-	/*log(0.00293)*/ -23891,
-	/*log(0.00317)*/ -23563,
-	/*log(0.00342)*/ -23259,
-	/*log(0.00366)*/ -22977,
-	/*log(0.00391)*/ -22713,
-	/*log(0.00415)*/ -22464,
-	/*log(0.00439)*/ -22230,
-	/*log(0.00464)*/ -22009,
-	/*log(0.00488)*/ -21799,
-	/*log(0.00513)*/ -21599,
-	/*log(0.00537)*/ -21408,
-	/*log(0.00562)*/ -21226,
-	/*log(0.00586)*/ -21052,
-	/*log(0.00610)*/ -20885,
-	/*log(0.00635)*/ -20724,
-	/*log(0.00659)*/ -20569,
-	/*log(0.00684)*/ -20420,
-	/*log(0.00708)*/ -20277,
-	/*log(0.00732)*/ -20138,
-	/*log(0.00757)*/ -20003,
-	/*log(0.00781)*/ -19873,
-	/*log(0.00806)*/ -19747,
-	/*log(0.00830)*/ -19625,
-	/*log(0.00854)*/ -19506,
-	/*log(0.00879)*/ -19391,
-	/*log(0.00903)*/ -19279,
-	/*log(0.00928)*/ -19170,
-	/*log(0.00952)*/ -19063,
-	/*log(0.00977)*/ -18959,
-	/*log(0.01001)*/ -18858,
-	/*log(0.01025)*/ -18760,
-	/*log(0.01050)*/ -18663,
-	/*log(0.01074)*/ -18569,
-	/*log(0.01099)*/ -18477,
-	/*log(0.01123)*/ -18387,
-	/*log(0.01147)*/ -18299,
-	/*log(0.01172)*/ -18213,
-	/*log(0.01196)*/ -18128,
-	/*log(0.01221)*/ -18045,
-	/*log(0.01245)*/ -17964,
-	/*log(0.01270)*/ -17885,
-	/*log(0.01294)*/ -17807,
-	/*log(0.01318)*/ -17730,
-	/*log(0.01343)*/ -17655,
-	/*log(0.01367)*/ -17581,
-	/*log(0.01392)*/ -17509,
-	/*log(0.01416)*/ -17437,
-	/*log(0.01440)*/ -17367,
-	/*log(0.01465)*/ -17299,
-	/*log(0.01489)*/ -17231,
-	/*log(0.01514)*/ -17164,
-	/*log(0.01538)*/ -17099,
-	/*log(0.01562)*/ -17034,
-	/*log(0.01587)*/ -16971,
-	/*log(0.01611)*/ -16908,
-	/*log(0.01636)*/ -16847,
-	/*log(0.01660)*/ -16786,
-	/*log(0.01685)*/ -16726,
-	/*log(0.01709)*/ -16667,
-	/*log(0.01733)*/ -16609,
-	/*log(0.01758)*/ -16552,
-	/*log(0.01782)*/ -16495,
-	/*log(0.01807)*/ -16440,
-	/*log(0.01831)*/ -16385,
-	/*log(0.01855)*/ -16330,
-	/*log(0.01880)*/ -16277,
-	/*log(0.01904)*/ -16224,
-	/*log(0.01929)*/ -16172,
-	/*log(0.01953)*/ -16120,
-	/*log(0.01978)*/ -16069,
-	/*log(0.02002)*/ -16019,
-	/*log(0.02026)*/ -15969,
-	/*log(0.02051)*/ -15920,
-	/*log(0.02075)*/ -15872,
-	/*log(0.02100)*/ -15824,
-	/*log(0.02124)*/ -15777,
-	/*log(0.02148)*/ -15730,
-	/*log(0.02173)*/ -15684,
-	/*log(0.02197)*/ -15638,
-	/*log(0.02222)*/ -15593,
-	/*log(0.02246)*/ -15548,
-	/*log(0.02271)*/ -15504,
-	/*log(0.02295)*/ -15460,
-	/*log(0.02319)*/ -15416,
-	/*log(0.02344)*/ -15374,
-	/*log(0.02368)*/ -15331,
-	/*log(0.02393)*/ -15289,
-	/*log(0.02417)*/ -15247,
-	/*log(0.02441)*/ -15206,
-	/*log(0.02466)*/ -15166,
-	/*log(0.02490)*/ -15125,
-	/*log(0.02515)*/ -15085,
-	/*log(0.02539)*/ -15046,
-	/*log(0.02563)*/ -15006,
-	/*log(0.02588)*/ -14968,
-	/*log(0.02612)*/ -14929,
-	/*log(0.02637)*/ -14891,
-	/*log(0.02661)*/ -14853,
-	/*log(0.02686)*/ -14816,
-	/*log(0.02710)*/ -14779,
-	/*log(0.02734)*/ -14742,
-	/*log(0.02759)*/ -14706,
-	/*log(0.02783)*/ -14670,
-	/*log(0.02808)*/ -14634,
-	/*log(0.02832)*/ -14598,
-	/*log(0.02856)*/ -14563,
-	/*log(0.02881)*/ -14528,
-	/*log(0.02905)*/ -14494,
-	/*log(0.02930)*/ -14460,
-	/*log(0.02954)*/ -14426,
-	/*log(0.02979)*/ -14392,
-	/*log(0.03003)*/ -14358,
-	/*log(0.03027)*/ -14325,
-	/*log(0.03052)*/ -14292,
-	/*log(0.03076)*/ -14260,
-	/*log(0.03101)*/ -14227,
-	/*log(0.03125)*/ -14195,
-	/*log(0.03149)*/ -14163,
-	/*log(0.03174)*/ -14132,
-	/*log(0.03198)*/ -14100,
-	/*log(0.03223)*/ -14069,
-	/*log(0.03247)*/ -14038,
-	/*log(0.03271)*/ -14008,
-	/*log(0.03296)*/ -13977,
-	/*log(0.03320)*/ -13947,
-	/*log(0.03345)*/ -13917,
-	/*log(0.03369)*/ -13887,
-	/*log(0.03394)*/ -13857,
-	/*log(0.03418)*/ -13828,
-	/*log(0.03442)*/ -13799,
-	/*log(0.03467)*/ -13770,
-	/*log(0.03491)*/ -13741,
-	/*log(0.03516)*/ -13713,
-	/*log(0.03540)*/ -13684,
-	/*log(0.03564)*/ -13656,
-	/*log(0.03589)*/ -13628,
-	/*log(0.03613)*/ -13600,
-	/*log(0.03638)*/ -13573,
-	/*log(0.03662)*/ -13546,
-	/*log(0.03687)*/ -13518,
-	/*log(0.03711)*/ -13491,
-	/*log(0.03735)*/ -13464,
-	/*log(0.03760)*/ -13438,
-	/*log(0.03784)*/ -13411,
-	/*log(0.03809)*/ -13385,
-	/*log(0.03833)*/ -13359,
-	/*log(0.03857)*/ -13333,
-	/*log(0.03882)*/ -13307,
-	/*log(0.03906)*/ -13281,
-	/*log(0.03931)*/ -13256,
-	/*log(0.03955)*/ -13230,
-	/*log(0.03979)*/ -13205,
-	/*log(0.04004)*/ -13180,
-	/*log(0.04028)*/ -13155,
-	/*log(0.04053)*/ -13130,
-	/*log(0.04077)*/ -13106,
-	/*log(0.04102)*/ -13081,
-	/*log(0.04126)*/ -13057,
-	/*log(0.04150)*/ -13033,
-	/*log(0.04175)*/ -13009,
-	/*log(0.04199)*/ -12985,
-	/*log(0.04224)*/ -12961,
-	/*log(0.04248)*/ -12938,
-	/*log(0.04272)*/ -12914,
-	/*log(0.04297)*/ -12891,
-	/*log(0.04321)*/ -12868,
-	/*log(0.04346)*/ -12844,
-	/*log(0.04370)*/ -12822,
-	/*log(0.04395)*/ -12799,
-	/*log(0.04419)*/ -12776,
-	/*log(0.04443)*/ -12753,
-	/*log(0.04468)*/ -12731,
-	/*log(0.04492)*/ -12709,
-	/*log(0.04517)*/ -12686,
-	/*log(0.04541)*/ -12664,
-	/*log(0.04565)*/ -12642,
-	/*log(0.04590)*/ -12621,
-	/*log(0.04614)*/ -12599,
-	/*log(0.04639)*/ -12577,
-	/*log(0.04663)*/ -12556,
-	/*log(0.04688)*/ -12534,
-	/*log(0.04712)*/ -12513,
-	/*log(0.04736)*/ -12492,
-	/*log(0.04761)*/ -12471,
-	/*log(0.04785)*/ -12450,
-	/*log(0.04810)*/ -12429,
-	/*log(0.04834)*/ -12408,
-	/*log(0.04858)*/ -12388,
-	/*log(0.04883)*/ -12367,
-	/*log(0.04907)*/ -12347,
-	/*log(0.04932)*/ -12326,
-	/*log(0.04956)*/ -12306,
-	/*log(0.04980)*/ -12286,
-	/*log(0.05005)*/ -12266,
-	/*log(0.05029)*/ -12246,
-	/*log(0.05054)*/ -12226,
-	/*log(0.05078)*/ -12207,
-	/*log(0.05103)*/ -12187,
-	/*log(0.05127)*/ -12167,
-	/*log(0.05151)*/ -12148,
-	/*log(0.05176)*/ -12128,
-	/*log(0.05200)*/ -12109,
-	/*log(0.05225)*/ -12090,
-	/*log(0.05249)*/ -12071,
-	/*log(0.05273)*/ -12052,
-	/*log(0.05298)*/ -12033,
-	/*log(0.05322)*/ -12014,
-	/*log(0.05347)*/ -11995,
-	/*log(0.05371)*/ -11977,
-	/*log(0.05396)*/ -11958,
-	/*log(0.05420)*/ -11940,
-	/*log(0.05444)*/ -11921,
-	/*log(0.05469)*/ -11903,
-	/*log(0.05493)*/ -11885,
-	/*log(0.05518)*/ -11867,
-	/*log(0.05542)*/ -11848,
-	/*log(0.05566)*/ -11830,
-	/*log(0.05591)*/ -11813,
-	/*log(0.05615)*/ -11795,
-	/*log(0.05640)*/ -11777,
-	/*log(0.05664)*/ -11759,
-	/*log(0.05688)*/ -11742,
-	/*log(0.05713)*/ -11724,
-	/*log(0.05737)*/ -11707,
-	/*log(0.05762)*/ -11689,
-	/*log(0.05786)*/ -11672,
-	/*log(0.05811)*/ -11655,
-	/*log(0.05835)*/ -11637,
-	/*log(0.05859)*/ -11620,
-	/*log(0.05884)*/ -11603,
-	/*log(0.05908)*/ -11586,
-	/*log(0.05933)*/ -11569,
-	/*log(0.05957)*/ -11553,
-	/*log(0.05981)*/ -11536,
-	/*log(0.06006)*/ -11519,
-	/*log(0.06030)*/ -11503,
-	/*log(0.06055)*/ -11486,
-	/*log(0.06079)*/ -11470,
-	/*log(0.06104)*/ -11453,
-	/*log(0.06128)*/ -11437,
-	/*log(0.06152)*/ -11421,
-	/*log(0.06177)*/ -11404,
-	/*log(0.06201)*/ -11388,
-	/*log(0.06226)*/ -11372,
-	/*log(0.06250)*/ -11356,
-	/*log(0.06274)*/ -11340,
-	/*log(0.06299)*/ -11324,
-	/*log(0.06323)*/ -11308,
-	/*log(0.06348)*/ -11293,
-	/*log(0.06372)*/ -11277,
-	/*log(0.06396)*/ -11261,
-	/*log(0.06421)*/ -11246,
-	/*log(0.06445)*/ -11230,
-	/*log(0.06470)*/ -11214,
-	/*log(0.06494)*/ -11199,
-	/*log(0.06519)*/ -11184,
-	/*log(0.06543)*/ -11168,
-	/*log(0.06567)*/ -11153,
-	/*log(0.06592)*/ -11138,
-	/*log(0.06616)*/ -11123,
-	/*log(0.06641)*/ -11108,
-	/*log(0.06665)*/ -11093,
-	/*log(0.06689)*/ -11078,
-	/*log(0.06714)*/ -11063,
-	/*log(0.06738)*/ -11048,
-	/*log(0.06763)*/ -11033,
-	/*log(0.06787)*/ -11018,
-	/*log(0.06812)*/ -11004,
-	/*log(0.06836)*/ -10989,
-	/*log(0.06860)*/ -10974,
-	/*log(0.06885)*/ -10960,
-	/*log(0.06909)*/ -10945,
-	/*log(0.06934)*/ -10931,
-	/*log(0.06958)*/ -10916,
-	/*log(0.06982)*/ -10902,
-	/*log(0.07007)*/ -10888,
-	/*log(0.07031)*/ -10874,
-	/*log(0.07056)*/ -10859,
-	/*log(0.07080)*/ -10845,
-	/*log(0.07104)*/ -10831,
-	/*log(0.07129)*/ -10817,
-	/*log(0.07153)*/ -10803,
-	/*log(0.07178)*/ -10789,
-	/*log(0.07202)*/ -10775,
-	/*log(0.07227)*/ -10761,
-	/*log(0.07251)*/ -10748,
-	/*log(0.07275)*/ -10734,
-	/*log(0.07300)*/ -10720,
-	/*log(0.07324)*/ -10706,
-	/*log(0.07349)*/ -10693,
-	/*log(0.07373)*/ -10679,
-	/*log(0.07397)*/ -10666,
-	/*log(0.07422)*/ -10652,
-	/*log(0.07446)*/ -10639,
-	/*log(0.07471)*/ -10625,
-	/*log(0.07495)*/ -10612,
-	/*log(0.07520)*/ -10599,
-	/*log(0.07544)*/ -10585,
-	/*log(0.07568)*/ -10572,
-	/*log(0.07593)*/ -10559,
-	/*log(0.07617)*/ -10546,
-	/*log(0.07642)*/ -10533,
-	/*log(0.07666)*/ -10520,
-	/*log(0.07690)*/ -10507,
-	/*log(0.07715)*/ -10494,
-	/*log(0.07739)*/ -10481,
-	/*log(0.07764)*/ -10468,
-	/*log(0.07788)*/ -10455,
-	/*log(0.07812)*/ -10442,
-	/*log(0.07837)*/ -10429,
-	/*log(0.07861)*/ -10417,
-	/*log(0.07886)*/ -10404,
-	/*log(0.07910)*/ -10391,
-	/*log(0.07935)*/ -10379,
-	/*log(0.07959)*/ -10366,
-	/*log(0.07983)*/ -10353,
-	/*log(0.08008)*/ -10341,
-	/*log(0.08032)*/ -10328,
-	/*log(0.08057)*/ -10316,
-	/*log(0.08081)*/ -10304,
-	/*log(0.08105)*/ -10291,
-	/*log(0.08130)*/ -10279,
-	/*log(0.08154)*/ -10267,
-	/*log(0.08179)*/ -10254,
-	/*log(0.08203)*/ -10242,
-	/*log(0.08228)*/ -10230,
-	/*log(0.08252)*/ -10218,
-	/*log(0.08276)*/ -10206,
-	/*log(0.08301)*/ -10194,
-	/*log(0.08325)*/ -10182,
-	/*log(0.08350)*/ -10170,
-	/*log(0.08374)*/ -10158,
-	/*log(0.08398)*/ -10146,
-	/*log(0.08423)*/ -10134,
-	/*log(0.08447)*/ -10122,
-	/*log(0.08472)*/ -10110,
-	/*log(0.08496)*/ -10098,
-	/*log(0.08521)*/ -10087,
-	/*log(0.08545)*/ -10075,
-	/*log(0.08569)*/ -10063,
-	/*log(0.08594)*/ -10052,
-	/*log(0.08618)*/ -10040,
-	/*log(0.08643)*/ -10028,
-	/*log(0.08667)*/ -10017,
-	/*log(0.08691)*/ -10005,
-	/*log(0.08716)*/ -9994,
-	/*log(0.08740)*/ -9982,
-	/*log(0.08765)*/ -9971,
-	/*log(0.08789)*/ -9960,
-	/*log(0.08813)*/ -9948,
-	/*log(0.08838)*/ -9937,
-	/*log(0.08862)*/ -9926,
-	/*log(0.08887)*/ -9914,
-	/*log(0.08911)*/ -9903,
-	/*log(0.08936)*/ -9892,
-	/*log(0.08960)*/ -9881,
-	/*log(0.08984)*/ -9870,
-	/*log(0.09009)*/ -9858,
-	/*log(0.09033)*/ -9847,
-	/*log(0.09058)*/ -9836,
-	/*log(0.09082)*/ -9825,
-	/*log(0.09106)*/ -9814,
-	/*log(0.09131)*/ -9803,
-	/*log(0.09155)*/ -9792,
-	/*log(0.09180)*/ -9781,
-	/*log(0.09204)*/ -9771,
-	/*log(0.09229)*/ -9760,
-	/*log(0.09253)*/ -9749,
-	/*log(0.09277)*/ -9738,
-	/*log(0.09302)*/ -9727,
-	/*log(0.09326)*/ -9717,
-	/*log(0.09351)*/ -9706,
-	/*log(0.09375)*/ -9695,
-	/*log(0.09399)*/ -9685,
-	/*log(0.09424)*/ -9674,
-	/*log(0.09448)*/ -9663,
-	/*log(0.09473)*/ -9653,
-	/*log(0.09497)*/ -9642,
-	/*log(0.09521)*/ -9632,
-	/*log(0.09546)*/ -9621,
-	/*log(0.09570)*/ -9611,
-	/*log(0.09595)*/ -9600,
-	/*log(0.09619)*/ -9590,
-	/*log(0.09644)*/ -9580,
-	/*log(0.09668)*/ -9569,
-	/*log(0.09692)*/ -9559,
-	/*log(0.09717)*/ -9549,
-	/*log(0.09741)*/ -9538,
-	/*log(0.09766)*/ -9528,
-	/*log(0.09790)*/ -9518,
-	/*log(0.09814)*/ -9508,
-	/*log(0.09839)*/ -9497,
-	/*log(0.09863)*/ -9487,
-	/*log(0.09888)*/ -9477,
-	/*log(0.09912)*/ -9467,
-	/*log(0.09937)*/ -9457,
-	/*log(0.09961)*/ -9447,
-	/*log(0.09985)*/ -9437,
-	/*log(0.10010)*/ -9427,
-	/*log(0.10034)*/ -9417,
-	/*log(0.10059)*/ -9407,
-	/*log(0.10083)*/ -9397,
-	/*log(0.10107)*/ -9387,
-	/*log(0.10132)*/ -9377,
-	/*log(0.10156)*/ -9367,
-	/*log(0.10181)*/ -9358,
-	/*log(0.10205)*/ -9348,
-	/*log(0.10229)*/ -9338,
-	/*log(0.10254)*/ -9328,
-	/*log(0.10278)*/ -9318,
-	/*log(0.10303)*/ -9309,
-	/*log(0.10327)*/ -9299,
-	/*log(0.10352)*/ -9289,
-	/*log(0.10376)*/ -9280,
-	/*log(0.10400)*/ -9270,
-	/*log(0.10425)*/ -9260,
-	/*log(0.10449)*/ -9251,
-	/*log(0.10474)*/ -9241,
-	/*log(0.10498)*/ -9232,
-	/*log(0.10522)*/ -9222,
-	/*log(0.10547)*/ -9213,
-	/*log(0.10571)*/ -9203,
-	/*log(0.10596)*/ -9194,
-	/*log(0.10620)*/ -9184,
-	/*log(0.10645)*/ -9175,
-	/*log(0.10669)*/ -9166,
-	/*log(0.10693)*/ -9156,
-	/*log(0.10718)*/ -9147,
-	/*log(0.10742)*/ -9138,
-	/*log(0.10767)*/ -9128,
-	/*log(0.10791)*/ -9119,
-	/*log(0.10815)*/ -9110,
-	/*log(0.10840)*/ -9101,
-	/*log(0.10864)*/ -9091,
-	/*log(0.10889)*/ -9082,
-	/*log(0.10913)*/ -9073,
-	/*log(0.10938)*/ -9064,
-	/*log(0.10962)*/ -9055,
-	/*log(0.10986)*/ -9046,
-	/*log(0.11011)*/ -9036,
-	/*log(0.11035)*/ -9027,
-	/*log(0.11060)*/ -9018,
-	/*log(0.11084)*/ -9009,
-	/*log(0.11108)*/ -9000,
-	/*log(0.11133)*/ -8991,
-	/*log(0.11157)*/ -8982,
-	/*log(0.11182)*/ -8973,
-	/*log(0.11206)*/ -8964,
-	/*log(0.11230)*/ -8956,
-	/*log(0.11255)*/ -8947,
-	/*log(0.11279)*/ -8938,
-	/*log(0.11304)*/ -8929,
-	/*log(0.11328)*/ -8920,
-	/*log(0.11353)*/ -8911,
-	/*log(0.11377)*/ -8902,
-	/*log(0.11401)*/ -8894,
-	/*log(0.11426)*/ -8885,
-	/*log(0.11450)*/ -8876,
-	/*log(0.11475)*/ -8867,
-	/*log(0.11499)*/ -8859,
-	/*log(0.11523)*/ -8850,
-	/*log(0.11548)*/ -8841,
-	/*log(0.11572)*/ -8833,
-	/*log(0.11597)*/ -8824,
-	/*log(0.11621)*/ -8816,
-	/*log(0.11646)*/ -8807,
-	/*log(0.11670)*/ -8798,
-	/*log(0.11694)*/ -8790,
-	/*log(0.11719)*/ -8781,
-	/*log(0.11743)*/ -8773,
-	/*log(0.11768)*/ -8764,
-	/*log(0.11792)*/ -8756,
-	/*log(0.11816)*/ -8747,
-	/*log(0.11841)*/ -8739,
-	/*log(0.11865)*/ -8730,
-	/*log(0.11890)*/ -8722,
-	/*log(0.11914)*/ -8714,
-	/*log(0.11938)*/ -8705,
-	/*log(0.11963)*/ -8697,
-	/*log(0.11987)*/ -8688,
-	/*log(0.12012)*/ -8680,
-	/*log(0.12036)*/ -8672,
-	/*log(0.12061)*/ -8663,
-	/*log(0.12085)*/ -8655,
-	/*log(0.12109)*/ -8647,
-	/*log(0.12134)*/ -8639,
-	/*log(0.12158)*/ -8630,
-	/*log(0.12183)*/ -8622,
-	/*log(0.12207)*/ -8614,
-	/*log(0.12231)*/ -8606,
-	/*log(0.12256)*/ -8598,
-	/*log(0.12280)*/ -8590,
-	/*log(0.12305)*/ -8581,
-	/*log(0.12329)*/ -8573,
-	/*log(0.12354)*/ -8565,
-	/*log(0.12378)*/ -8557,
-	/*log(0.12402)*/ -8549,
-	/*log(0.12427)*/ -8541,
-	/*log(0.12451)*/ -8533,
-	/*log(0.12476)*/ -8525,
-	/*log(0.12500)*/ -8517,
-	/*log(0.12524)*/ -8509,
-	/*log(0.12549)*/ -8501,
-	/*log(0.12573)*/ -8493,
-	/*log(0.12598)*/ -8485,
-	/*log(0.12622)*/ -8477,
-	/*log(0.12646)*/ -8469,
-	/*log(0.12671)*/ -8461,
-	/*log(0.12695)*/ -8453,
-	/*log(0.12720)*/ -8446,
-	/*log(0.12744)*/ -8438,
-	/*log(0.12769)*/ -8430,
-	/*log(0.12793)*/ -8422,
-	/*log(0.12817)*/ -8414,
-	/*log(0.12842)*/ -8406,
-	/*log(0.12866)*/ -8399,
-	/*log(0.12891)*/ -8391,
-	/*log(0.12915)*/ -8383,
-	/*log(0.12939)*/ -8375,
-	/*log(0.12964)*/ -8368,
-	/*log(0.12988)*/ -8360,
-	/*log(0.13013)*/ -8352,
-	/*log(0.13037)*/ -8345,
-	/*log(0.13062)*/ -8337,
-	/*log(0.13086)*/ -8329,
-	/*log(0.13110)*/ -8322,
-	/*log(0.13135)*/ -8314,
-	/*log(0.13159)*/ -8306,
-	/*log(0.13184)*/ -8299,
-	/*log(0.13208)*/ -8291,
-	/*log(0.13232)*/ -8284,
-	/*log(0.13257)*/ -8276,
-	/*log(0.13281)*/ -8269,
-	/*log(0.13306)*/ -8261,
-	/*log(0.13330)*/ -8254,
-	/*log(0.13354)*/ -8246,
-	/*log(0.13379)*/ -8239,
-	/*log(0.13403)*/ -8231,
-	/*log(0.13428)*/ -8224,
-	/*log(0.13452)*/ -8216,
-	/*log(0.13477)*/ -8209,
-	/*log(0.13501)*/ -8201,
-	/*log(0.13525)*/ -8194,
-	/*log(0.13550)*/ -8187,
-	/*log(0.13574)*/ -8179,
-	/*log(0.13599)*/ -8172,
-	/*log(0.13623)*/ -8164,
-	/*log(0.13647)*/ -8157,
-	/*log(0.13672)*/ -8150,
-	/*log(0.13696)*/ -8143,
-	/*log(0.13721)*/ -8135,
-	/*log(0.13745)*/ -8128,
-	/*log(0.13770)*/ -8121,
-	/*log(0.13794)*/ -8113,
-	/*log(0.13818)*/ -8106,
-	/*log(0.13843)*/ -8099,
-	/*log(0.13867)*/ -8092,
-	/*log(0.13892)*/ -8085,
-	/*log(0.13916)*/ -8077,
-	/*log(0.13940)*/ -8070,
-	/*log(0.13965)*/ -8063,
-	/*log(0.13989)*/ -8056,
-	/*log(0.14014)*/ -8049,
-	/*log(0.14038)*/ -8042,
-	/*log(0.14062)*/ -8034,
-	/*log(0.14087)*/ -8027,
-	/*log(0.14111)*/ -8020,
-	/*log(0.14136)*/ -8013,
-	/*log(0.14160)*/ -8006,
-	/*log(0.14185)*/ -7999,
-	/*log(0.14209)*/ -7992,
-	/*log(0.14233)*/ -7985,
-	/*log(0.14258)*/ -7978,
-	/*log(0.14282)*/ -7971,
-	/*log(0.14307)*/ -7964,
-	/*log(0.14331)*/ -7957,
-	/*log(0.14355)*/ -7950,
-	/*log(0.14380)*/ -7943,
-	/*log(0.14404)*/ -7936,
-	/*log(0.14429)*/ -7929,
-	/*log(0.14453)*/ -7922,
-	/*log(0.14478)*/ -7915,
-	/*log(0.14502)*/ -7908,
-	/*log(0.14526)*/ -7902,
-	/*log(0.14551)*/ -7895,
-	/*log(0.14575)*/ -7888,
-	/*log(0.14600)*/ -7881,
-	/*log(0.14624)*/ -7874,
-	/*log(0.14648)*/ -7867,
-	/*log(0.14673)*/ -7860,
-	/*log(0.14697)*/ -7854,
-	/*log(0.14722)*/ -7847,
-	/*log(0.14746)*/ -7840,
-	/*log(0.14771)*/ -7833,
-	/*log(0.14795)*/ -7826,
-	/*log(0.14819)*/ -7820,
-	/*log(0.14844)*/ -7813,
-	/*log(0.14868)*/ -7806,
-	/*log(0.14893)*/ -7800,
-	/*log(0.14917)*/ -7793,
-	/*log(0.14941)*/ -7786,
-	/*log(0.14966)*/ -7779,
-	/*log(0.14990)*/ -7773,
-	/*log(0.15015)*/ -7766,
-	/*log(0.15039)*/ -7759,
-	/*log(0.15063)*/ -7753,
-	/*log(0.15088)*/ -7746,
-	/*log(0.15112)*/ -7740,
-	/*log(0.15137)*/ -7733,
-	/*log(0.15161)*/ -7726,
-	/*log(0.15186)*/ -7720,
-	/*log(0.15210)*/ -7713,
-	/*log(0.15234)*/ -7707,
-	/*log(0.15259)*/ -7700,
-	/*log(0.15283)*/ -7693,
-	/*log(0.15308)*/ -7687,
-	/*log(0.15332)*/ -7680,
-	/*log(0.15356)*/ -7674,
-	/*log(0.15381)*/ -7667,
-	/*log(0.15405)*/ -7661,
-	/*log(0.15430)*/ -7654,
-	/*log(0.15454)*/ -7648,
-	/*log(0.15479)*/ -7641,
-	/*log(0.15503)*/ -7635,
-	/*log(0.15527)*/ -7629,
-	/*log(0.15552)*/ -7622,
-	/*log(0.15576)*/ -7616,
-	/*log(0.15601)*/ -7609,
-	/*log(0.15625)*/ -7603,
-	/*log(0.15649)*/ -7597,
-	/*log(0.15674)*/ -7590,
-	/*log(0.15698)*/ -7584,
-	/*log(0.15723)*/ -7577,
-	/*log(0.15747)*/ -7571,
-	/*log(0.15771)*/ -7565,
-	/*log(0.15796)*/ -7558,
-	/*log(0.15820)*/ -7552,
-	/*log(0.15845)*/ -7546,
-	/*log(0.15869)*/ -7539,
-	/*log(0.15894)*/ -7533,
-	/*log(0.15918)*/ -7527,
-	/*log(0.15942)*/ -7521,
-	/*log(0.15967)*/ -7514,
-	/*log(0.15991)*/ -7508,
-	/*log(0.16016)*/ -7502,
-	/*log(0.16040)*/ -7496,
-	/*log(0.16064)*/ -7489,
-	/*log(0.16089)*/ -7483,
-	/*log(0.16113)*/ -7477,
-	/*log(0.16138)*/ -7471,
-	/*log(0.16162)*/ -7464,
-	/*log(0.16187)*/ -7458,
-	/*log(0.16211)*/ -7452,
-	/*log(0.16235)*/ -7446,
-	/*log(0.16260)*/ -7440,
-	/*log(0.16284)*/ -7434,
-	/*log(0.16309)*/ -7428,
-	/*log(0.16333)*/ -7421,
-	/*log(0.16357)*/ -7415,
-	/*log(0.16382)*/ -7409,
-	/*log(0.16406)*/ -7403,
-	/*log(0.16431)*/ -7397,
-	/*log(0.16455)*/ -7391,
-	/*log(0.16479)*/ -7385,
-	/*log(0.16504)*/ -7379,
-	/*log(0.16528)*/ -7373,
-	/*log(0.16553)*/ -7367,
-	/*log(0.16577)*/ -7361,
-	/*log(0.16602)*/ -7355,
-	/*log(0.16626)*/ -7349,
-	/*log(0.16650)*/ -7343,
-	/*log(0.16675)*/ -7337,
-	/*log(0.16699)*/ -7331,
-	/*log(0.16724)*/ -7325,
-	/*log(0.16748)*/ -7319,
-	/*log(0.16772)*/ -7313,
-	/*log(0.16797)*/ -7307,
-	/*log(0.16821)*/ -7301,
-	/*log(0.16846)*/ -7295,
-	/*log(0.16870)*/ -7289,
-	/*log(0.16895)*/ -7283,
-	/*log(0.16919)*/ -7277,
-	/*log(0.16943)*/ -7271,
-	/*log(0.16968)*/ -7265,
-	/*log(0.16992)*/ -7259,
-	/*log(0.17017)*/ -7253,
-	/*log(0.17041)*/ -7248,
-	/*log(0.17065)*/ -7242,
-	/*log(0.17090)*/ -7236,
-	/*log(0.17114)*/ -7230,
-	/*log(0.17139)*/ -7224,
-	/*log(0.17163)*/ -7218,
-	/*log(0.17188)*/ -7213,
-	/*log(0.17212)*/ -7207,
-	/*log(0.17236)*/ -7201,
-	/*log(0.17261)*/ -7195,
-	/*log(0.17285)*/ -7189,
-	/*log(0.17310)*/ -7184,
-	/*log(0.17334)*/ -7178,
-	/*log(0.17358)*/ -7172,
-	/*log(0.17383)*/ -7166,
-	/*log(0.17407)*/ -7160,
-	/*log(0.17432)*/ -7155,
-	/*log(0.17456)*/ -7149,
-	/*log(0.17480)*/ -7143,
-	/*log(0.17505)*/ -7138,
-	/*log(0.17529)*/ -7132,
-	/*log(0.17554)*/ -7126,
-	/*log(0.17578)*/ -7120,
-	/*log(0.17603)*/ -7115,
-	/*log(0.17627)*/ -7109,
-	/*log(0.17651)*/ -7103,
-	/*log(0.17676)*/ -7098,
-	/*log(0.17700)*/ -7092,
-	/*log(0.17725)*/ -7086,
-	/*log(0.17749)*/ -7081,
-	/*log(0.17773)*/ -7075,
-	/*log(0.17798)*/ -7070,
-	/*log(0.17822)*/ -7064,
-	/*log(0.17847)*/ -7058,
-	/*log(0.17871)*/ -7053,
-	/*log(0.17896)*/ -7047,
-	/*log(0.17920)*/ -7042,
-	/*log(0.17944)*/ -7036,
-	/*log(0.17969)*/ -7030,
-	/*log(0.17993)*/ -7025,
-	/*log(0.18018)*/ -7019,
-	/*log(0.18042)*/ -7014,
-	/*log(0.18066)*/ -7008,
-	/*log(0.18091)*/ -7003,
-	/*log(0.18115)*/ -6997,
-	/*log(0.18140)*/ -6992,
-	/*log(0.18164)*/ -6986,
-	/*log(0.18188)*/ -6981,
-	/*log(0.18213)*/ -6975,
-	/*log(0.18237)*/ -6970,
-	/*log(0.18262)*/ -6964,
-	/*log(0.18286)*/ -6959,
-	/*log(0.18311)*/ -6953,
-	/*log(0.18335)*/ -6948,
-	/*log(0.18359)*/ -6942,
-	/*log(0.18384)*/ -6937,
-	/*log(0.18408)*/ -6931,
-	/*log(0.18433)*/ -6926,
-	/*log(0.18457)*/ -6921,
-	/*log(0.18481)*/ -6915,
-	/*log(0.18506)*/ -6910,
-	/*log(0.18530)*/ -6904,
-	/*log(0.18555)*/ -6899,
-	/*log(0.18579)*/ -6894,
-	/*log(0.18604)*/ -6888,
-	/*log(0.18628)*/ -6883,
-	/*log(0.18652)*/ -6877,
-	/*log(0.18677)*/ -6872,
-	/*log(0.18701)*/ -6867,
-	/*log(0.18726)*/ -6861,
-	/*log(0.18750)*/ -6856,
-	/*log(0.18774)*/ -6851,
-	/*log(0.18799)*/ -6845,
-	/*log(0.18823)*/ -6840,
-	/*log(0.18848)*/ -6835,
-	/*log(0.18872)*/ -6830,
-	/*log(0.18896)*/ -6824,
-	/*log(0.18921)*/ -6819,
-	/*log(0.18945)*/ -6814,
-	/*log(0.18970)*/ -6808,
-	/*log(0.18994)*/ -6803,
-	/*log(0.19019)*/ -6798,
-	/*log(0.19043)*/ -6793,
-	/*log(0.19067)*/ -6787,
-	/*log(0.19092)*/ -6782,
-	/*log(0.19116)*/ -6777,
-	/*log(0.19141)*/ -6772,
-	/*log(0.19165)*/ -6766,
-	/*log(0.19189)*/ -6761,
-	/*log(0.19214)*/ -6756,
-	/*log(0.19238)*/ -6751,
-	/*log(0.19263)*/ -6746,
-	/*log(0.19287)*/ -6740,
-	/*log(0.19312)*/ -6735,
-	/*log(0.19336)*/ -6730,
-	/*log(0.19360)*/ -6725,
-	/*log(0.19385)*/ -6720,
-	/*log(0.19409)*/ -6715,
-	/*log(0.19434)*/ -6709,
-	/*log(0.19458)*/ -6704,
-	/*log(0.19482)*/ -6699,
-	/*log(0.19507)*/ -6694,
-	/*log(0.19531)*/ -6689,
-	/*log(0.19556)*/ -6684,
-	/*log(0.19580)*/ -6679,
-	/*log(0.19604)*/ -6674,
-	/*log(0.19629)*/ -6668,
-	/*log(0.19653)*/ -6663,
-	/*log(0.19678)*/ -6658,
-	/*log(0.19702)*/ -6653,
-	/*log(0.19727)*/ -6648,
-	/*log(0.19751)*/ -6643,
-	/*log(0.19775)*/ -6638,
-	/*log(0.19800)*/ -6633,
-	/*log(0.19824)*/ -6628,
-	/*log(0.19849)*/ -6623,
-	/*log(0.19873)*/ -6618,
-	/*log(0.19897)*/ -6613,
-	/*log(0.19922)*/ -6608,
-	/*log(0.19946)*/ -6603,
-	/*log(0.19971)*/ -6598,
-	/*log(0.19995)*/ -6593,
-	/*log(0.20020)*/ -6588,
-	/*log(0.20044)*/ -6583,
-	/*log(0.20068)*/ -6578,
-	/*log(0.20093)*/ -6573,
-	/*log(0.20117)*/ -6568,
-	/*log(0.20142)*/ -6563,
-	/*log(0.20166)*/ -6558,
-	/*log(0.20190)*/ -6553,
-	/*log(0.20215)*/ -6548,
-	/*log(0.20239)*/ -6543,
-	/*log(0.20264)*/ -6538,
-	/*log(0.20288)*/ -6533,
-	/*log(0.20312)*/ -6528,
-	/*log(0.20337)*/ -6523,
-	/*log(0.20361)*/ -6518,
-	/*log(0.20386)*/ -6514,
-	/*log(0.20410)*/ -6509,
-	/*log(0.20435)*/ -6504,
-	/*log(0.20459)*/ -6499,
-	/*log(0.20483)*/ -6494,
-	/*log(0.20508)*/ -6489,
-	/*log(0.20532)*/ -6484,
-	/*log(0.20557)*/ -6479,
-	/*log(0.20581)*/ -6474,
-	/*log(0.20605)*/ -6470,
-	/*log(0.20630)*/ -6465,
-	/*log(0.20654)*/ -6460,
-	/*log(0.20679)*/ -6455,
-	/*log(0.20703)*/ -6450,
-	/*log(0.20728)*/ -6445,
-	/*log(0.20752)*/ -6441,
-	/*log(0.20776)*/ -6436,
-	/*log(0.20801)*/ -6431,
-	/*log(0.20825)*/ -6426,
-	/*log(0.20850)*/ -6421,
-	/*log(0.20874)*/ -6417,
-	/*log(0.20898)*/ -6412,
-	/*log(0.20923)*/ -6407,
-	/*log(0.20947)*/ -6402,
-	/*log(0.20972)*/ -6397,
-	/*log(0.20996)*/ -6393,
-	/*log(0.21021)*/ -6388,
-	/*log(0.21045)*/ -6383,
-	/*log(0.21069)*/ -6378,
-	/*log(0.21094)*/ -6374,
-	/*log(0.21118)*/ -6369,
-	/*log(0.21143)*/ -6364,
-	/*log(0.21167)*/ -6359,
-	/*log(0.21191)*/ -6355,
-	/*log(0.21216)*/ -6350,
-	/*log(0.21240)*/ -6345,
-	/*log(0.21265)*/ -6341,
-	/*log(0.21289)*/ -6336,
-	/*log(0.21313)*/ -6331,
-	/*log(0.21338)*/ -6327,
-	/*log(0.21362)*/ -6322,
-	/*log(0.21387)*/ -6317,
-	/*log(0.21411)*/ -6312,
-	/*log(0.21436)*/ -6308,
-	/*log(0.21460)*/ -6303,
-	/*log(0.21484)*/ -6299,
-	/*log(0.21509)*/ -6294,
-	/*log(0.21533)*/ -6289,
-	/*log(0.21558)*/ -6285,
-	/*log(0.21582)*/ -6280,
-	/*log(0.21606)*/ -6275,
-	/*log(0.21631)*/ -6271,
-	/*log(0.21655)*/ -6266,
-	/*log(0.21680)*/ -6261,
-	/*log(0.21704)*/ -6257,
-	/*log(0.21729)*/ -6252,
-	/*log(0.21753)*/ -6248,
-	/*log(0.21777)*/ -6243,
-	/*log(0.21802)*/ -6238,
-	/*log(0.21826)*/ -6234,
-	/*log(0.21851)*/ -6229,
-	/*log(0.21875)*/ -6225,
-	/*log(0.21899)*/ -6220,
-	/*log(0.21924)*/ -6216,
-	/*log(0.21948)*/ -6211,
-	/*log(0.21973)*/ -6206,
-	/*log(0.21997)*/ -6202,
-	/*log(0.22021)*/ -6197,
-	/*log(0.22046)*/ -6193,
-	/*log(0.22070)*/ -6188,
-	/*log(0.22095)*/ -6184,
-	/*log(0.22119)*/ -6179,
-	/*log(0.22144)*/ -6175,
-	/*log(0.22168)*/ -6170,
-	/*log(0.22192)*/ -6166,
-	/*log(0.22217)*/ -6161,
-	/*log(0.22241)*/ -6157,
-	/*log(0.22266)*/ -6152,
-	/*log(0.22290)*/ -6148,
-	/*log(0.22314)*/ -6143,
-	/*log(0.22339)*/ -6139,
-	/*log(0.22363)*/ -6134,
-	/*log(0.22388)*/ -6130,
-	/*log(0.22412)*/ -6125,
-	/*log(0.22437)*/ -6121,
-	/*log(0.22461)*/ -6116,
-	/*log(0.22485)*/ -6112,
-	/*log(0.22510)*/ -6108,
-	/*log(0.22534)*/ -6103,
-	/*log(0.22559)*/ -6099,
-	/*log(0.22583)*/ -6094,
-	/*log(0.22607)*/ -6090,
-	/*log(0.22632)*/ -6085,
-	/*log(0.22656)*/ -6081,
-	/*log(0.22681)*/ -6077,
-	/*log(0.22705)*/ -6072,
-	/*log(0.22729)*/ -6068,
-	/*log(0.22754)*/ -6063,
-	/*log(0.22778)*/ -6059,
-	/*log(0.22803)*/ -6055,
-	/*log(0.22827)*/ -6050,
-	/*log(0.22852)*/ -6046,
-	/*log(0.22876)*/ -6041,
-	/*log(0.22900)*/ -6037,
-	/*log(0.22925)*/ -6033,
-	/*log(0.22949)*/ -6028,
-	/*log(0.22974)*/ -6024,
-	/*log(0.22998)*/ -6020,
-	/*log(0.23022)*/ -6015,
-	/*log(0.23047)*/ -6011,
-	/*log(0.23071)*/ -6007,
-	/*log(0.23096)*/ -6002,
-	/*log(0.23120)*/ -5998,
-	/*log(0.23145)*/ -5994,
-	/*log(0.23169)*/ -5989,
-	/*log(0.23193)*/ -5985,
-	/*log(0.23218)*/ -5981,
-	/*log(0.23242)*/ -5976,
-	/*log(0.23267)*/ -5972,
-	/*log(0.23291)*/ -5968,
-	/*log(0.23315)*/ -5964,
-	/*log(0.23340)*/ -5959,
-	/*log(0.23364)*/ -5955,
-	/*log(0.23389)*/ -5951,
-	/*log(0.23413)*/ -5946,
-	/*log(0.23438)*/ -5942,
-	/*log(0.23462)*/ -5938,
-	/*log(0.23486)*/ -5934,
-	/*log(0.23511)*/ -5929,
-	/*log(0.23535)*/ -5925,
-	/*log(0.23560)*/ -5921,
-	/*log(0.23584)*/ -5917,
-	/*log(0.23608)*/ -5912,
-	/*log(0.23633)*/ -5908,
-	/*log(0.23657)*/ -5904,
-	/*log(0.23682)*/ -5900,
-	/*log(0.23706)*/ -5895,
-	/*log(0.23730)*/ -5891,
-	/*log(0.23755)*/ -5887,
-	/*log(0.23779)*/ -5883,
-	/*log(0.23804)*/ -5879,
-	/*log(0.23828)*/ -5874,
-	/*log(0.23853)*/ -5870,
-	/*log(0.23877)*/ -5866,
-	/*log(0.23901)*/ -5862,
-	/*log(0.23926)*/ -5858,
-	/*log(0.23950)*/ -5853,
-	/*log(0.23975)*/ -5849,
-	/*log(0.23999)*/ -5845,
-	/*log(0.24023)*/ -5841,
-	/*log(0.24048)*/ -5837,
-	/*log(0.24072)*/ -5833,
-	/*log(0.24097)*/ -5829,
-	/*log(0.24121)*/ -5824,
-	/*log(0.24146)*/ -5820,
-	/*log(0.24170)*/ -5816,
-	/*log(0.24194)*/ -5812,
-	/*log(0.24219)*/ -5808,
-	/*log(0.24243)*/ -5804,
-	/*log(0.24268)*/ -5800,
-	/*log(0.24292)*/ -5795,
-	/*log(0.24316)*/ -5791,
-	/*log(0.24341)*/ -5787,
-	/*log(0.24365)*/ -5783,
-	/*log(0.24390)*/ -5779,
-	/*log(0.24414)*/ -5775,
-	/*log(0.24438)*/ -5771,
-	/*log(0.24463)*/ -5767,
-	/*log(0.24487)*/ -5763,
-	/*log(0.24512)*/ -5759,
-	/*log(0.24536)*/ -5754,
-	/*log(0.24561)*/ -5750,
-	/*log(0.24585)*/ -5746,
-	/*log(0.24609)*/ -5742,
-	/*log(0.24634)*/ -5738,
-	/*log(0.24658)*/ -5734,
-	/*log(0.24683)*/ -5730,
-	/*log(0.24707)*/ -5726,
-	/*log(0.24731)*/ -5722,
-	/*log(0.24756)*/ -5718,
-	/*log(0.24780)*/ -5714,
-	/*log(0.24805)*/ -5710,
-	/*log(0.24829)*/ -5706,
-	/*log(0.24854)*/ -5702,
-	/*log(0.24878)*/ -5698,
-	/*log(0.24902)*/ -5694,
-	/*log(0.24927)*/ -5690,
-	/*log(0.24951)*/ -5686,
-	/*log(0.24976)*/ -5682,
-	/*log(0.25000)*/ -5678,
-	/*log(0.25024)*/ -5674,
-	/*log(0.25049)*/ -5670,
-	/*log(0.25073)*/ -5666,
-	/*log(0.25098)*/ -5662,
-	/*log(0.25122)*/ -5658,
-	/*log(0.25146)*/ -5654,
-	/*log(0.25171)*/ -5650,
-	/*log(0.25195)*/ -5646,
-	/*log(0.25220)*/ -5642,
-	/*log(0.25244)*/ -5638,
-	/*log(0.25269)*/ -5634,
-	/*log(0.25293)*/ -5630,
-	/*log(0.25317)*/ -5626,
-	/*log(0.25342)*/ -5622,
-	/*log(0.25366)*/ -5618,
-	/*log(0.25391)*/ -5614,
-	/*log(0.25415)*/ -5610,
-	/*log(0.25439)*/ -5606,
-	/*log(0.25464)*/ -5602,
-	/*log(0.25488)*/ -5599,
-	/*log(0.25513)*/ -5595,
-	/*log(0.25537)*/ -5591,
-	/*log(0.25562)*/ -5587,
-	/*log(0.25586)*/ -5583,
-	/*log(0.25610)*/ -5579,
-	/*log(0.25635)*/ -5575,
-	/*log(0.25659)*/ -5571,
-	/*log(0.25684)*/ -5567,
-	/*log(0.25708)*/ -5563,
-	/*log(0.25732)*/ -5559,
-	/*log(0.25757)*/ -5556,
-	/*log(0.25781)*/ -5552,
-	/*log(0.25806)*/ -5548,
-	/*log(0.25830)*/ -5544,
-	/*log(0.25854)*/ -5540,
-	/*log(0.25879)*/ -5536,
-	/*log(0.25903)*/ -5532,
-	/*log(0.25928)*/ -5529,
-	/*log(0.25952)*/ -5525,
-	/*log(0.25977)*/ -5521,
-	/*log(0.26001)*/ -5517,
-	/*log(0.26025)*/ -5513,
-	/*log(0.26050)*/ -5509,
-	/*log(0.26074)*/ -5505,
-	/*log(0.26099)*/ -5502,
-	/*log(0.26123)*/ -5498,
-	/*log(0.26147)*/ -5494,
-	/*log(0.26172)*/ -5490,
-	/*log(0.26196)*/ -5486,
-	/*log(0.26221)*/ -5482,
-	/*log(0.26245)*/ -5479,
-	/*log(0.26270)*/ -5475,
-	/*log(0.26294)*/ -5471,
-	/*log(0.26318)*/ -5467,
-	/*log(0.26343)*/ -5463,
-	/*log(0.26367)*/ -5460,
-	/*log(0.26392)*/ -5456,
-	/*log(0.26416)*/ -5452,
-	/*log(0.26440)*/ -5448,
-	/*log(0.26465)*/ -5445,
-	/*log(0.26489)*/ -5441,
-	/*log(0.26514)*/ -5437,
-	/*log(0.26538)*/ -5433,
-	/*log(0.26562)*/ -5429,
-	/*log(0.26587)*/ -5426,
-	/*log(0.26611)*/ -5422,
-	/*log(0.26636)*/ -5418,
-	/*log(0.26660)*/ -5414,
-	/*log(0.26685)*/ -5411,
-	/*log(0.26709)*/ -5407,
-	/*log(0.26733)*/ -5403,
-	/*log(0.26758)*/ -5399,
-	/*log(0.26782)*/ -5396,
-	/*log(0.26807)*/ -5392,
-	/*log(0.26831)*/ -5388,
-	/*log(0.26855)*/ -5385,
-	/*log(0.26880)*/ -5381,
-	/*log(0.26904)*/ -5377,
-	/*log(0.26929)*/ -5373,
-	/*log(0.26953)*/ -5370,
-	/*log(0.26978)*/ -5366,
-	/*log(0.27002)*/ -5362,
-	/*log(0.27026)*/ -5359,
-	/*log(0.27051)*/ -5355,
-	/*log(0.27075)*/ -5351,
-	/*log(0.27100)*/ -5347,
-	/*log(0.27124)*/ -5344,
-	/*log(0.27148)*/ -5340,
-	/*log(0.27173)*/ -5336,
-	/*log(0.27197)*/ -5333,
-	/*log(0.27222)*/ -5329,
-	/*log(0.27246)*/ -5325,
-	/*log(0.27271)*/ -5322,
-	/*log(0.27295)*/ -5318,
-	/*log(0.27319)*/ -5314,
-	/*log(0.27344)*/ -5311,
-	/*log(0.27368)*/ -5307,
-	/*log(0.27393)*/ -5303,
-	/*log(0.27417)*/ -5300,
-	/*log(0.27441)*/ -5296,
-	/*log(0.27466)*/ -5292,
-	/*log(0.27490)*/ -5289,
-	/*log(0.27515)*/ -5285,
-	/*log(0.27539)*/ -5282,
-	/*log(0.27563)*/ -5278,
-	/*log(0.27588)*/ -5274,
-	/*log(0.27612)*/ -5271,
-	/*log(0.27637)*/ -5267,
-	/*log(0.27661)*/ -5263,
-	/*log(0.27686)*/ -5260,
-	/*log(0.27710)*/ -5256,
-	/*log(0.27734)*/ -5253,
-	/*log(0.27759)*/ -5249,
-	/*log(0.27783)*/ -5245,
-	/*log(0.27808)*/ -5242,
-	/*log(0.27832)*/ -5238,
-	/*log(0.27856)*/ -5235,
-	/*log(0.27881)*/ -5231,
-	/*log(0.27905)*/ -5227,
-	/*log(0.27930)*/ -5224,
-	/*log(0.27954)*/ -5220,
-	/*log(0.27979)*/ -5217,
-	/*log(0.28003)*/ -5213,
-	/*log(0.28027)*/ -5210,
-	/*log(0.28052)*/ -5206,
-	/*log(0.28076)*/ -5202,
-	/*log(0.28101)*/ -5199,
-	/*log(0.28125)*/ -5195,
-	/*log(0.28149)*/ -5192,
-	/*log(0.28174)*/ -5188,
-	/*log(0.28198)*/ -5185,
-	/*log(0.28223)*/ -5181,
-	/*log(0.28247)*/ -5178,
-	/*log(0.28271)*/ -5174,
-	/*log(0.28296)*/ -5171,
-	/*log(0.28320)*/ -5167,
-	/*log(0.28345)*/ -5163,
-	/*log(0.28369)*/ -5160,
-	/*log(0.28394)*/ -5156,
-	/*log(0.28418)*/ -5153,
-	/*log(0.28442)*/ -5149,
-	/*log(0.28467)*/ -5146,
-	/*log(0.28491)*/ -5142,
-	/*log(0.28516)*/ -5139,
-	/*log(0.28540)*/ -5135,
-	/*log(0.28564)*/ -5132,
-	/*log(0.28589)*/ -5128,
-	/*log(0.28613)*/ -5125,
-	/*log(0.28638)*/ -5121,
-	/*log(0.28662)*/ -5118,
-	/*log(0.28687)*/ -5114,
-	/*log(0.28711)*/ -5111,
-	/*log(0.28735)*/ -5107,
-	/*log(0.28760)*/ -5104,
-	/*log(0.28784)*/ -5100,
-	/*log(0.28809)*/ -5097,
-	/*log(0.28833)*/ -5093,
-	/*log(0.28857)*/ -5090,
-	/*log(0.28882)*/ -5087,
-	/*log(0.28906)*/ -5083,
-	/*log(0.28931)*/ -5080,
-	/*log(0.28955)*/ -5076,
-	/*log(0.28979)*/ -5073,
-	/*log(0.29004)*/ -5069,
-	/*log(0.29028)*/ -5066,
-	/*log(0.29053)*/ -5062,
-	/*log(0.29077)*/ -5059,
-	/*log(0.29102)*/ -5056,
-	/*log(0.29126)*/ -5052,
-	/*log(0.29150)*/ -5049,
-	/*log(0.29175)*/ -5045,
-	/*log(0.29199)*/ -5042,
-	/*log(0.29224)*/ -5038,
-	/*log(0.29248)*/ -5035,
-	/*log(0.29272)*/ -5032,
-	/*log(0.29297)*/ -5028,
-	/*log(0.29321)*/ -5025,
-	/*log(0.29346)*/ -5021,
-	/*log(0.29370)*/ -5018,
-	/*log(0.29395)*/ -5014,
-	/*log(0.29419)*/ -5011,
-	/*log(0.29443)*/ -5008,
-	/*log(0.29468)*/ -5004,
-	/*log(0.29492)*/ -5001,
-	/*log(0.29517)*/ -4998,
-	/*log(0.29541)*/ -4994,
-	/*log(0.29565)*/ -4991,
-	/*log(0.29590)*/ -4987,
-	/*log(0.29614)*/ -4984,
-	/*log(0.29639)*/ -4981,
-	/*log(0.29663)*/ -4977,
-	/*log(0.29688)*/ -4974,
-	/*log(0.29712)*/ -4970,
-	/*log(0.29736)*/ -4967,
-	/*log(0.29761)*/ -4964,
-	/*log(0.29785)*/ -4960,
-	/*log(0.29810)*/ -4957,
-	/*log(0.29834)*/ -4954,
-	/*log(0.29858)*/ -4950,
-	/*log(0.29883)*/ -4947,
-	/*log(0.29907)*/ -4944,
-	/*log(0.29932)*/ -4940,
-	/*log(0.29956)*/ -4937,
-	/*log(0.29980)*/ -4934,
-	/*log(0.30005)*/ -4930,
-	/*log(0.30029)*/ -4927,
-	/*log(0.30054)*/ -4924,
-	/*log(0.30078)*/ -4920,
-	/*log(0.30103)*/ -4917,
-	/*log(0.30127)*/ -4914,
-	/*log(0.30151)*/ -4910,
-	/*log(0.30176)*/ -4907,
-	/*log(0.30200)*/ -4904,
-	/*log(0.30225)*/ -4900,
-	/*log(0.30249)*/ -4897,
-	/*log(0.30273)*/ -4894,
-	/*log(0.30298)*/ -4891,
-	/*log(0.30322)*/ -4887,
-	/*log(0.30347)*/ -4884,
-	/*log(0.30371)*/ -4881,
-	/*log(0.30396)*/ -4877,
-	/*log(0.30420)*/ -4874,
-	/*log(0.30444)*/ -4871,
-	/*log(0.30469)*/ -4867,
-	/*log(0.30493)*/ -4864,
-	/*log(0.30518)*/ -4861,
-	/*log(0.30542)*/ -4858,
-	/*log(0.30566)*/ -4854,
-	/*log(0.30591)*/ -4851,
-	/*log(0.30615)*/ -4848,
-	/*log(0.30640)*/ -4845,
-	/*log(0.30664)*/ -4841,
-	/*log(0.30688)*/ -4838,
-	/*log(0.30713)*/ -4835,
-	/*log(0.30737)*/ -4832,
-	/*log(0.30762)*/ -4828,
-	/*log(0.30786)*/ -4825,
-	/*log(0.30811)*/ -4822,
-	/*log(0.30835)*/ -4819,
-	/*log(0.30859)*/ -4815,
-	/*log(0.30884)*/ -4812,
-	/*log(0.30908)*/ -4809,
-	/*log(0.30933)*/ -4806,
-	/*log(0.30957)*/ -4802,
-	/*log(0.30981)*/ -4799,
-	/*log(0.31006)*/ -4796,
-	/*log(0.31030)*/ -4793,
-	/*log(0.31055)*/ -4789,
-	/*log(0.31079)*/ -4786,
-	/*log(0.31104)*/ -4783,
-	/*log(0.31128)*/ -4780,
-	/*log(0.31152)*/ -4777,
-	/*log(0.31177)*/ -4773,
-	/*log(0.31201)*/ -4770,
-	/*log(0.31226)*/ -4767,
-	/*log(0.31250)*/ -4764,
-	/*log(0.31274)*/ -4761,
-	/*log(0.31299)*/ -4757,
-	/*log(0.31323)*/ -4754,
-	/*log(0.31348)*/ -4751,
-	/*log(0.31372)*/ -4748,
-	/*log(0.31396)*/ -4745,
-	/*log(0.31421)*/ -4741,
-	/*log(0.31445)*/ -4738,
-	/*log(0.31470)*/ -4735,
-	/*log(0.31494)*/ -4732,
-	/*log(0.31519)*/ -4729,
-	/*log(0.31543)*/ -4726,
-	/*log(0.31567)*/ -4722,
-	/*log(0.31592)*/ -4719,
-	/*log(0.31616)*/ -4716,
-	/*log(0.31641)*/ -4713,
-	/*log(0.31665)*/ -4710,
-	/*log(0.31689)*/ -4707,
-	/*log(0.31714)*/ -4703,
-	/*log(0.31738)*/ -4700,
-	/*log(0.31763)*/ -4697,
-	/*log(0.31787)*/ -4694,
-	/*log(0.31812)*/ -4691,
-	/*log(0.31836)*/ -4688,
-	/*log(0.31860)*/ -4685,
-	/*log(0.31885)*/ -4681,
-	/*log(0.31909)*/ -4678,
-	/*log(0.31934)*/ -4675,
-	/*log(0.31958)*/ -4672,
-	/*log(0.31982)*/ -4669,
-	/*log(0.32007)*/ -4666,
-	/*log(0.32031)*/ -4663,
-	/*log(0.32056)*/ -4660,
-	/*log(0.32080)*/ -4656,
-	/*log(0.32104)*/ -4653,
-	/*log(0.32129)*/ -4650,
-	/*log(0.32153)*/ -4647,
-	/*log(0.32178)*/ -4644,
-	/*log(0.32202)*/ -4641,
-	/*log(0.32227)*/ -4638,
-	/*log(0.32251)*/ -4635,
-	/*log(0.32275)*/ -4632,
-	/*log(0.32300)*/ -4628,
-	/*log(0.32324)*/ -4625,
-	/*log(0.32349)*/ -4622,
-	/*log(0.32373)*/ -4619,
-	/*log(0.32397)*/ -4616,
-	/*log(0.32422)*/ -4613,
-	/*log(0.32446)*/ -4610,
-	/*log(0.32471)*/ -4607,
-	/*log(0.32495)*/ -4604,
-	/*log(0.32520)*/ -4601,
-	/*log(0.32544)*/ -4598,
-	/*log(0.32568)*/ -4595,
-	/*log(0.32593)*/ -4591,
-	/*log(0.32617)*/ -4588,
-	/*log(0.32642)*/ -4585,
-	/*log(0.32666)*/ -4582,
-	/*log(0.32690)*/ -4579,
-	/*log(0.32715)*/ -4576,
-	/*log(0.32739)*/ -4573,
-	/*log(0.32764)*/ -4570,
-	/*log(0.32788)*/ -4567,
-	/*log(0.32812)*/ -4564,
-	/*log(0.32837)*/ -4561,
-	/*log(0.32861)*/ -4558,
-	/*log(0.32886)*/ -4555,
-	/*log(0.32910)*/ -4552,
-	/*log(0.32935)*/ -4549,
-	/*log(0.32959)*/ -4546,
-	/*log(0.32983)*/ -4543,
-	/*log(0.33008)*/ -4540,
-	/*log(0.33032)*/ -4537,
-	/*log(0.33057)*/ -4534,
-	/*log(0.33081)*/ -4531,
-	/*log(0.33105)*/ -4528,
-	/*log(0.33130)*/ -4524,
-	/*log(0.33154)*/ -4521,
-	/*log(0.33179)*/ -4518,
-	/*log(0.33203)*/ -4515,
-	/*log(0.33228)*/ -4512,
-	/*log(0.33252)*/ -4509,
-	/*log(0.33276)*/ -4506,
-	/*log(0.33301)*/ -4503,
-	/*log(0.33325)*/ -4500,
-	/*log(0.33350)*/ -4497,
-	/*log(0.33374)*/ -4494,
-	/*log(0.33398)*/ -4491,
-	/*log(0.33423)*/ -4488,
-	/*log(0.33447)*/ -4485,
-	/*log(0.33472)*/ -4482,
-	/*log(0.33496)*/ -4479,
-	/*log(0.33521)*/ -4476,
-	/*log(0.33545)*/ -4473,
-	/*log(0.33569)*/ -4471,
-	/*log(0.33594)*/ -4468,
-	/*log(0.33618)*/ -4465,
-	/*log(0.33643)*/ -4462,
-	/*log(0.33667)*/ -4459,
-	/*log(0.33691)*/ -4456,
-	/*log(0.33716)*/ -4453,
-	/*log(0.33740)*/ -4450,
-	/*log(0.33765)*/ -4447,
-	/*log(0.33789)*/ -4444,
-	/*log(0.33813)*/ -4441,
-	/*log(0.33838)*/ -4438,
-	/*log(0.33862)*/ -4435,
-	/*log(0.33887)*/ -4432,
-	/*log(0.33911)*/ -4429,
-	/*log(0.33936)*/ -4426,
-	/*log(0.33960)*/ -4423,
-	/*log(0.33984)*/ -4420,
-	/*log(0.34009)*/ -4417,
-	/*log(0.34033)*/ -4414,
-	/*log(0.34058)*/ -4411,
-	/*log(0.34082)*/ -4408,
-	/*log(0.34106)*/ -4406,
-	/*log(0.34131)*/ -4403,
-	/*log(0.34155)*/ -4400,
-	/*log(0.34180)*/ -4397,
-	/*log(0.34204)*/ -4394,
-	/*log(0.34229)*/ -4391,
-	/*log(0.34253)*/ -4388,
-	/*log(0.34277)*/ -4385,
-	/*log(0.34302)*/ -4382,
-	/*log(0.34326)*/ -4379,
-	/*log(0.34351)*/ -4376,
-	/*log(0.34375)*/ -4373,
-	/*log(0.34399)*/ -4370,
-	/*log(0.34424)*/ -4368,
-	/*log(0.34448)*/ -4365,
-	/*log(0.34473)*/ -4362,
-	/*log(0.34497)*/ -4359,
-	/*log(0.34521)*/ -4356,
-	/*log(0.34546)*/ -4353,
-	/*log(0.34570)*/ -4350,
-	/*log(0.34595)*/ -4347,
-	/*log(0.34619)*/ -4344,
-	/*log(0.34644)*/ -4341,
-	/*log(0.34668)*/ -4339,
-	/*log(0.34692)*/ -4336,
-	/*log(0.34717)*/ -4333,
-	/*log(0.34741)*/ -4330,
-	/*log(0.34766)*/ -4327,
-	/*log(0.34790)*/ -4324,
-	/*log(0.34814)*/ -4321,
-	/*log(0.34839)*/ -4318,
-	/*log(0.34863)*/ -4316,
-	/*log(0.34888)*/ -4313,
-	/*log(0.34912)*/ -4310,
-	/*log(0.34937)*/ -4307,
-	/*log(0.34961)*/ -4304,
-	/*log(0.34985)*/ -4301,
-	/*log(0.35010)*/ -4298,
-	/*log(0.35034)*/ -4296,
-	/*log(0.35059)*/ -4293,
-	/*log(0.35083)*/ -4290,
-	/*log(0.35107)*/ -4287,
-	/*log(0.35132)*/ -4284,
-	/*log(0.35156)*/ -4281,
-	/*log(0.35181)*/ -4278,
-	/*log(0.35205)*/ -4276,
-	/*log(0.35229)*/ -4273,
-	/*log(0.35254)*/ -4270,
-	/*log(0.35278)*/ -4267,
-	/*log(0.35303)*/ -4264,
-	/*log(0.35327)*/ -4261,
-	/*log(0.35352)*/ -4259,
-	/*log(0.35376)*/ -4256,
-	/*log(0.35400)*/ -4253,
-	/*log(0.35425)*/ -4250,
-	/*log(0.35449)*/ -4247,
-	/*log(0.35474)*/ -4245,
-	/*log(0.35498)*/ -4242,
-	/*log(0.35522)*/ -4239,
-	/*log(0.35547)*/ -4236,
-	/*log(0.35571)*/ -4233,
-	/*log(0.35596)*/ -4230,
-	/*log(0.35620)*/ -4228,
-	/*log(0.35645)*/ -4225,
-	/*log(0.35669)*/ -4222,
-	/*log(0.35693)*/ -4219,
-	/*log(0.35718)*/ -4216,
-	/*log(0.35742)*/ -4214,
-	/*log(0.35767)*/ -4211,
-	/*log(0.35791)*/ -4208,
-	/*log(0.35815)*/ -4205,
-	/*log(0.35840)*/ -4202,
-	/*log(0.35864)*/ -4200,
-	/*log(0.35889)*/ -4197,
-	/*log(0.35913)*/ -4194,
-	/*log(0.35938)*/ -4191,
-	/*log(0.35962)*/ -4189,
-	/*log(0.35986)*/ -4186,
-	/*log(0.36011)*/ -4183,
-	/*log(0.36035)*/ -4180,
-	/*log(0.36060)*/ -4177,
-	/*log(0.36084)*/ -4175,
-	/*log(0.36108)*/ -4172,
-	/*log(0.36133)*/ -4169,
-	/*log(0.36157)*/ -4166,
-	/*log(0.36182)*/ -4164,
-	/*log(0.36206)*/ -4161,
-	/*log(0.36230)*/ -4158,
-	/*log(0.36255)*/ -4155,
-	/*log(0.36279)*/ -4153,
-	/*log(0.36304)*/ -4150,
-	/*log(0.36328)*/ -4147,
-	/*log(0.36353)*/ -4144,
-	/*log(0.36377)*/ -4142,
-	/*log(0.36401)*/ -4139,
-	/*log(0.36426)*/ -4136,
-	/*log(0.36450)*/ -4133,
-	/*log(0.36475)*/ -4131,
-	/*log(0.36499)*/ -4128,
-	/*log(0.36523)*/ -4125,
-	/*log(0.36548)*/ -4122,
-	/*log(0.36572)*/ -4120,
-	/*log(0.36597)*/ -4117,
-	/*log(0.36621)*/ -4114,
-	/*log(0.36646)*/ -4111,
-	/*log(0.36670)*/ -4109,
-	/*log(0.36694)*/ -4106,
-	/*log(0.36719)*/ -4103,
-	/*log(0.36743)*/ -4100,
-	/*log(0.36768)*/ -4098,
-	/*log(0.36792)*/ -4095,
-	/*log(0.36816)*/ -4092,
-	/*log(0.36841)*/ -4090,
-	/*log(0.36865)*/ -4087,
-	/*log(0.36890)*/ -4084,
-	/*log(0.36914)*/ -4081,
-	/*log(0.36938)*/ -4079,
-	/*log(0.36963)*/ -4076,
-	/*log(0.36987)*/ -4073,
-	/*log(0.37012)*/ -4071,
-	/*log(0.37036)*/ -4068,
-	/*log(0.37061)*/ -4065,
-	/*log(0.37085)*/ -4063,
-	/*log(0.37109)*/ -4060,
-	/*log(0.37134)*/ -4057,
-	/*log(0.37158)*/ -4054,
-	/*log(0.37183)*/ -4052,
-	/*log(0.37207)*/ -4049,
-	/*log(0.37231)*/ -4046,
-	/*log(0.37256)*/ -4044,
-	/*log(0.37280)*/ -4041,
-	/*log(0.37305)*/ -4038,
-	/*log(0.37329)*/ -4036,
-	/*log(0.37354)*/ -4033,
-	/*log(0.37378)*/ -4030,
-	/*log(0.37402)*/ -4028,
-	/*log(0.37427)*/ -4025,
-	/*log(0.37451)*/ -4022,
-	/*log(0.37476)*/ -4020,
-	/*log(0.37500)*/ -4017,
-	/*log(0.37524)*/ -4014,
-	/*log(0.37549)*/ -4012,
-	/*log(0.37573)*/ -4009,
-	/*log(0.37598)*/ -4006,
-	/*log(0.37622)*/ -4004,
-	/*log(0.37646)*/ -4001,
-	/*log(0.37671)*/ -3998,
-	/*log(0.37695)*/ -3996,
-	/*log(0.37720)*/ -3993,
-	/*log(0.37744)*/ -3990,
-	/*log(0.37769)*/ -3988,
-	/*log(0.37793)*/ -3985,
-	/*log(0.37817)*/ -3982,
-	/*log(0.37842)*/ -3980,
-	/*log(0.37866)*/ -3977,
-	/*log(0.37891)*/ -3975,
-	/*log(0.37915)*/ -3972,
-	/*log(0.37939)*/ -3969,
-	/*log(0.37964)*/ -3967,
-	/*log(0.37988)*/ -3964,
-	/*log(0.38013)*/ -3961,
-	/*log(0.38037)*/ -3959,
-	/*log(0.38062)*/ -3956,
-	/*log(0.38086)*/ -3953,
-	/*log(0.38110)*/ -3951,
-	/*log(0.38135)*/ -3948,
-	/*log(0.38159)*/ -3946,
-	/*log(0.38184)*/ -3943,
-	/*log(0.38208)*/ -3940,
-	/*log(0.38232)*/ -3938,
-	/*log(0.38257)*/ -3935,
-	/*log(0.38281)*/ -3933,
-	/*log(0.38306)*/ -3930,
-	/*log(0.38330)*/ -3927,
-	/*log(0.38354)*/ -3925,
-	/*log(0.38379)*/ -3922,
-	/*log(0.38403)*/ -3919,
-	/*log(0.38428)*/ -3917,
-	/*log(0.38452)*/ -3914,
-	/*log(0.38477)*/ -3912,
-	/*log(0.38501)*/ -3909,
-	/*log(0.38525)*/ -3906,
-	/*log(0.38550)*/ -3904,
-	/*log(0.38574)*/ -3901,
-	/*log(0.38599)*/ -3899,
-	/*log(0.38623)*/ -3896,
-	/*log(0.38647)*/ -3894,
-	/*log(0.38672)*/ -3891,
-	/*log(0.38696)*/ -3888,
-	/*log(0.38721)*/ -3886,
-	/*log(0.38745)*/ -3883,
-	/*log(0.38770)*/ -3881,
-	/*log(0.38794)*/ -3878,
-	/*log(0.38818)*/ -3875,
-	/*log(0.38843)*/ -3873,
-	/*log(0.38867)*/ -3870,
-	/*log(0.38892)*/ -3868,
-	/*log(0.38916)*/ -3865,
-	/*log(0.38940)*/ -3863,
-	/*log(0.38965)*/ -3860,
-	/*log(0.38989)*/ -3857,
-	/*log(0.39014)*/ -3855,
-	/*log(0.39038)*/ -3852,
-	/*log(0.39062)*/ -3850,
-	/*log(0.39087)*/ -3847,
-	/*log(0.39111)*/ -3845,
-	/*log(0.39136)*/ -3842,
-	/*log(0.39160)*/ -3840,
-	/*log(0.39185)*/ -3837,
-	/*log(0.39209)*/ -3834,
-	/*log(0.39233)*/ -3832,
-	/*log(0.39258)*/ -3829,
-	/*log(0.39282)*/ -3827,
-	/*log(0.39307)*/ -3824,
-	/*log(0.39331)*/ -3822,
-	/*log(0.39355)*/ -3819,
-	/*log(0.39380)*/ -3817,
-	/*log(0.39404)*/ -3814,
-	/*log(0.39429)*/ -3812,
-	/*log(0.39453)*/ -3809,
-	/*log(0.39478)*/ -3806,
-	/*log(0.39502)*/ -3804,
-	/*log(0.39526)*/ -3801,
-	/*log(0.39551)*/ -3799,
-	/*log(0.39575)*/ -3796,
-	/*log(0.39600)*/ -3794,
-	/*log(0.39624)*/ -3791,
-	/*log(0.39648)*/ -3789,
-	/*log(0.39673)*/ -3786,
-	/*log(0.39697)*/ -3784,
-	/*log(0.39722)*/ -3781,
-	/*log(0.39746)*/ -3779,
-	/*log(0.39771)*/ -3776,
-	/*log(0.39795)*/ -3774,
-	/*log(0.39819)*/ -3771,
-	/*log(0.39844)*/ -3769,
-	/*log(0.39868)*/ -3766,
-	/*log(0.39893)*/ -3764,
-	/*log(0.39917)*/ -3761,
-	/*log(0.39941)*/ -3759,
-	/*log(0.39966)*/ -3756,
-	/*log(0.39990)*/ -3754,
-	/*log(0.40015)*/ -3751,
-	/*log(0.40039)*/ -3749,
-	/*log(0.40063)*/ -3746,
-	/*log(0.40088)*/ -3744,
-	/*log(0.40112)*/ -3741,
-	/*log(0.40137)*/ -3739,
-	/*log(0.40161)*/ -3736,
-	/*log(0.40186)*/ -3734,
-	/*log(0.40210)*/ -3731,
-	/*log(0.40234)*/ -3729,
-	/*log(0.40259)*/ -3726,
-	/*log(0.40283)*/ -3724,
-	/*log(0.40308)*/ -3721,
-	/*log(0.40332)*/ -3719,
-	/*log(0.40356)*/ -3716,
-	/*log(0.40381)*/ -3714,
-	/*log(0.40405)*/ -3711,
-	/*log(0.40430)*/ -3709,
-	/*log(0.40454)*/ -3706,
-	/*log(0.40479)*/ -3704,
-	/*log(0.40503)*/ -3701,
-	/*log(0.40527)*/ -3699,
-	/*log(0.40552)*/ -3697,
-	/*log(0.40576)*/ -3694,
-	/*log(0.40601)*/ -3692,
-	/*log(0.40625)*/ -3689,
-	/*log(0.40649)*/ -3687,
-	/*log(0.40674)*/ -3684,
-	/*log(0.40698)*/ -3682,
-	/*log(0.40723)*/ -3679,
-	/*log(0.40747)*/ -3677,
-	/*log(0.40771)*/ -3674,
-	/*log(0.40796)*/ -3672,
-	/*log(0.40820)*/ -3669,
-	/*log(0.40845)*/ -3667,
-	/*log(0.40869)*/ -3665,
-	/*log(0.40894)*/ -3662,
-	/*log(0.40918)*/ -3660,
-	/*log(0.40942)*/ -3657,
-	/*log(0.40967)*/ -3655,
-	/*log(0.40991)*/ -3652,
-	/*log(0.41016)*/ -3650,
-	/*log(0.41040)*/ -3647,
-	/*log(0.41064)*/ -3645,
-	/*log(0.41089)*/ -3643,
-	/*log(0.41113)*/ -3640,
-	/*log(0.41138)*/ -3638,
-	/*log(0.41162)*/ -3635,
-	/*log(0.41187)*/ -3633,
-	/*log(0.41211)*/ -3630,
-	/*log(0.41235)*/ -3628,
-	/*log(0.41260)*/ -3626,
-	/*log(0.41284)*/ -3623,
-	/*log(0.41309)*/ -3621,
-	/*log(0.41333)*/ -3618,
-	/*log(0.41357)*/ -3616,
-	/*log(0.41382)*/ -3614,
-	/*log(0.41406)*/ -3611,
-	/*log(0.41431)*/ -3609,
-	/*log(0.41455)*/ -3606,
-	/*log(0.41479)*/ -3604,
-	/*log(0.41504)*/ -3601,
-	/*log(0.41528)*/ -3599,
-	/*log(0.41553)*/ -3597,
-	/*log(0.41577)*/ -3594,
-	/*log(0.41602)*/ -3592,
-	/*log(0.41626)*/ -3589,
-	/*log(0.41650)*/ -3587,
-	/*log(0.41675)*/ -3585,
-	/*log(0.41699)*/ -3582,
-	/*log(0.41724)*/ -3580,
-	/*log(0.41748)*/ -3577,
-	/*log(0.41772)*/ -3575,
-	/*log(0.41797)*/ -3573,
-	/*log(0.41821)*/ -3570,
-	/*log(0.41846)*/ -3568,
-	/*log(0.41870)*/ -3565,
-	/*log(0.41895)*/ -3563,
-	/*log(0.41919)*/ -3561,
-	/*log(0.41943)*/ -3558,
-	/*log(0.41968)*/ -3556,
-	/*log(0.41992)*/ -3554,
-	/*log(0.42017)*/ -3551,
-	/*log(0.42041)*/ -3549,
-	/*log(0.42065)*/ -3546,
-	/*log(0.42090)*/ -3544,
-	/*log(0.42114)*/ -3542,
-	/*log(0.42139)*/ -3539,
-	/*log(0.42163)*/ -3537,
-	/*log(0.42188)*/ -3535,
-	/*log(0.42212)*/ -3532,
-	/*log(0.42236)*/ -3530,
-	/*log(0.42261)*/ -3527,
-	/*log(0.42285)*/ -3525,
-	/*log(0.42310)*/ -3523,
-	/*log(0.42334)*/ -3520,
-	/*log(0.42358)*/ -3518,
-	/*log(0.42383)*/ -3516,
-	/*log(0.42407)*/ -3513,
-	/*log(0.42432)*/ -3511,
-	/*log(0.42456)*/ -3509,
-	/*log(0.42480)*/ -3506,
-	/*log(0.42505)*/ -3504,
-	/*log(0.42529)*/ -3501,
-	/*log(0.42554)*/ -3499,
-	/*log(0.42578)*/ -3497,
-	/*log(0.42603)*/ -3494,
-	/*log(0.42627)*/ -3492,
-	/*log(0.42651)*/ -3490,
-	/*log(0.42676)*/ -3487,
-	/*log(0.42700)*/ -3485,
-	/*log(0.42725)*/ -3483,
-	/*log(0.42749)*/ -3480,
-	/*log(0.42773)*/ -3478,
-	/*log(0.42798)*/ -3476,
-	/*log(0.42822)*/ -3473,
-	/*log(0.42847)*/ -3471,
-	/*log(0.42871)*/ -3469,
-	/*log(0.42896)*/ -3466,
-	/*log(0.42920)*/ -3464,
-	/*log(0.42944)*/ -3462,
-	/*log(0.42969)*/ -3459,
-	/*log(0.42993)*/ -3457,
-	/*log(0.43018)*/ -3455,
-	/*log(0.43042)*/ -3452,
-	/*log(0.43066)*/ -3450,
-	/*log(0.43091)*/ -3448,
-	/*log(0.43115)*/ -3445,
-	/*log(0.43140)*/ -3443,
-	/*log(0.43164)*/ -3441,
-	/*log(0.43188)*/ -3438,
-	/*log(0.43213)*/ -3436,
-	/*log(0.43237)*/ -3434,
-	/*log(0.43262)*/ -3432,
-	/*log(0.43286)*/ -3429,
-	/*log(0.43311)*/ -3427,
-	/*log(0.43335)*/ -3425,
-	/*log(0.43359)*/ -3422,
-	/*log(0.43384)*/ -3420,
-	/*log(0.43408)*/ -3418,
-	/*log(0.43433)*/ -3415,
-	/*log(0.43457)*/ -3413,
-	/*log(0.43481)*/ -3411,
-	/*log(0.43506)*/ -3408,
-	/*log(0.43530)*/ -3406,
-	/*log(0.43555)*/ -3404,
-	/*log(0.43579)*/ -3402,
-	/*log(0.43604)*/ -3399,
-	/*log(0.43628)*/ -3397,
-	/*log(0.43652)*/ -3395,
-	/*log(0.43677)*/ -3392,
-	/*log(0.43701)*/ -3390,
-	/*log(0.43726)*/ -3388,
-	/*log(0.43750)*/ -3386,
-	/*log(0.43774)*/ -3383,
-	/*log(0.43799)*/ -3381,
-	/*log(0.43823)*/ -3379,
-	/*log(0.43848)*/ -3376,
-	/*log(0.43872)*/ -3374,
-	/*log(0.43896)*/ -3372,
-	/*log(0.43921)*/ -3370,
-	/*log(0.43945)*/ -3367,
-	/*log(0.43970)*/ -3365,
-	/*log(0.43994)*/ -3363,
-	/*log(0.44019)*/ -3361,
-	/*log(0.44043)*/ -3358,
-	/*log(0.44067)*/ -3356,
-	/*log(0.44092)*/ -3354,
-	/*log(0.44116)*/ -3351,
-	/*log(0.44141)*/ -3349,
-	/*log(0.44165)*/ -3347,
-	/*log(0.44189)*/ -3345,
-	/*log(0.44214)*/ -3342,
-	/*log(0.44238)*/ -3340,
-	/*log(0.44263)*/ -3338,
-	/*log(0.44287)*/ -3336,
-	/*log(0.44312)*/ -3333,
-	/*log(0.44336)*/ -3331,
-	/*log(0.44360)*/ -3329,
-	/*log(0.44385)*/ -3327,
-	/*log(0.44409)*/ -3324,
-	/*log(0.44434)*/ -3322,
-	/*log(0.44458)*/ -3320,
-	/*log(0.44482)*/ -3318,
-	/*log(0.44507)*/ -3315,
-	/*log(0.44531)*/ -3313,
-	/*log(0.44556)*/ -3311,
-	/*log(0.44580)*/ -3309,
-	/*log(0.44604)*/ -3306,
-	/*log(0.44629)*/ -3304,
-	/*log(0.44653)*/ -3302,
-	/*log(0.44678)*/ -3300,
-	/*log(0.44702)*/ -3297,
-	/*log(0.44727)*/ -3295,
-	/*log(0.44751)*/ -3293,
-	/*log(0.44775)*/ -3291,
-	/*log(0.44800)*/ -3288,
-	/*log(0.44824)*/ -3286,
-	/*log(0.44849)*/ -3284,
-	/*log(0.44873)*/ -3282,
-	/*log(0.44897)*/ -3280,
-	/*log(0.44922)*/ -3277,
-	/*log(0.44946)*/ -3275,
-	/*log(0.44971)*/ -3273,
-	/*log(0.44995)*/ -3271,
-	/*log(0.45020)*/ -3268,
-	/*log(0.45044)*/ -3266,
-	/*log(0.45068)*/ -3264,
-	/*log(0.45093)*/ -3262,
-	/*log(0.45117)*/ -3260,
-	/*log(0.45142)*/ -3257,
-	/*log(0.45166)*/ -3255,
-	/*log(0.45190)*/ -3253,
-	/*log(0.45215)*/ -3251,
-	/*log(0.45239)*/ -3248,
-	/*log(0.45264)*/ -3246,
-	/*log(0.45288)*/ -3244,
-	/*log(0.45312)*/ -3242,
-	/*log(0.45337)*/ -3240,
-	/*log(0.45361)*/ -3237,
-	/*log(0.45386)*/ -3235,
-	/*log(0.45410)*/ -3233,
-	/*log(0.45435)*/ -3231,
-	/*log(0.45459)*/ -3229,
-	/*log(0.45483)*/ -3226,
-	/*log(0.45508)*/ -3224,
-	/*log(0.45532)*/ -3222,
-	/*log(0.45557)*/ -3220,
-	/*log(0.45581)*/ -3218,
-	/*log(0.45605)*/ -3215,
-	/*log(0.45630)*/ -3213,
-	/*log(0.45654)*/ -3211,
-	/*log(0.45679)*/ -3209,
-	/*log(0.45703)*/ -3207,
-	/*log(0.45728)*/ -3204,
-	/*log(0.45752)*/ -3202,
-	/*log(0.45776)*/ -3200,
-	/*log(0.45801)*/ -3198,
-	/*log(0.45825)*/ -3196,
-	/*log(0.45850)*/ -3194,
-	/*log(0.45874)*/ -3191,
-	/*log(0.45898)*/ -3189,
-	/*log(0.45923)*/ -3187,
-	/*log(0.45947)*/ -3185,
-	/*log(0.45972)*/ -3183,
-	/*log(0.45996)*/ -3181,
-	/*log(0.46021)*/ -3178,
-	/*log(0.46045)*/ -3176,
-	/*log(0.46069)*/ -3174,
-	/*log(0.46094)*/ -3172,
-	/*log(0.46118)*/ -3170,
-	/*log(0.46143)*/ -3167,
-	/*log(0.46167)*/ -3165,
-	/*log(0.46191)*/ -3163,
-	/*log(0.46216)*/ -3161,
-	/*log(0.46240)*/ -3159,
-	/*log(0.46265)*/ -3157,
-	/*log(0.46289)*/ -3155,
-	/*log(0.46313)*/ -3152,
-	/*log(0.46338)*/ -3150,
-	/*log(0.46362)*/ -3148,
-	/*log(0.46387)*/ -3146,
-	/*log(0.46411)*/ -3144,
-	/*log(0.46436)*/ -3142,
-	/*log(0.46460)*/ -3139,
-	/*log(0.46484)*/ -3137,
-	/*log(0.46509)*/ -3135,
-	/*log(0.46533)*/ -3133,
-	/*log(0.46558)*/ -3131,
-	/*log(0.46582)*/ -3129,
-	/*log(0.46606)*/ -3127,
-	/*log(0.46631)*/ -3124,
-	/*log(0.46655)*/ -3122,
-	/*log(0.46680)*/ -3120,
-	/*log(0.46704)*/ -3118,
-	/*log(0.46729)*/ -3116,
-	/*log(0.46753)*/ -3114,
-	/*log(0.46777)*/ -3112,
-	/*log(0.46802)*/ -3109,
-	/*log(0.46826)*/ -3107,
-	/*log(0.46851)*/ -3105,
-	/*log(0.46875)*/ -3103,
-	/*log(0.46899)*/ -3101,
-	/*log(0.46924)*/ -3099,
-	/*log(0.46948)*/ -3097,
-	/*log(0.46973)*/ -3094,
-	/*log(0.46997)*/ -3092,
-	/*log(0.47021)*/ -3090,
-	/*log(0.47046)*/ -3088,
-	/*log(0.47070)*/ -3086,
-	/*log(0.47095)*/ -3084,
-	/*log(0.47119)*/ -3082,
-	/*log(0.47144)*/ -3080,
-	/*log(0.47168)*/ -3077,
-	/*log(0.47192)*/ -3075,
-	/*log(0.47217)*/ -3073,
-	/*log(0.47241)*/ -3071,
-	/*log(0.47266)*/ -3069,
-	/*log(0.47290)*/ -3067,
-	/*log(0.47314)*/ -3065,
-	/*log(0.47339)*/ -3063,
-	/*log(0.47363)*/ -3061,
-	/*log(0.47388)*/ -3058,
-	/*log(0.47412)*/ -3056,
-	/*log(0.47437)*/ -3054,
-	/*log(0.47461)*/ -3052,
-	/*log(0.47485)*/ -3050,
-	/*log(0.47510)*/ -3048,
-	/*log(0.47534)*/ -3046,
-	/*log(0.47559)*/ -3044,
-	/*log(0.47583)*/ -3042,
-	/*log(0.47607)*/ -3039,
-	/*log(0.47632)*/ -3037,
-	/*log(0.47656)*/ -3035,
-	/*log(0.47681)*/ -3033,
-	/*log(0.47705)*/ -3031,
-	/*log(0.47729)*/ -3029,
-	/*log(0.47754)*/ -3027,
-	/*log(0.47778)*/ -3025,
-	/*log(0.47803)*/ -3023,
-	/*log(0.47827)*/ -3021,
-	/*log(0.47852)*/ -3019,
-	/*log(0.47876)*/ -3016,
-	/*log(0.47900)*/ -3014,
-	/*log(0.47925)*/ -3012,
-	/*log(0.47949)*/ -3010,
-	/*log(0.47974)*/ -3008,
-	/*log(0.47998)*/ -3006,
-	/*log(0.48022)*/ -3004,
-	/*log(0.48047)*/ -3002,
-	/*log(0.48071)*/ -3000,
-	/*log(0.48096)*/ -2998,
-	/*log(0.48120)*/ -2996,
-	/*log(0.48145)*/ -2994,
-	/*log(0.48169)*/ -2991,
-	/*log(0.48193)*/ -2989,
-	/*log(0.48218)*/ -2987,
-	/*log(0.48242)*/ -2985,
-	/*log(0.48267)*/ -2983,
-	/*log(0.48291)*/ -2981,
-	/*log(0.48315)*/ -2979,
-	/*log(0.48340)*/ -2977,
-	/*log(0.48364)*/ -2975,
-	/*log(0.48389)*/ -2973,
-	/*log(0.48413)*/ -2971,
-	/*log(0.48438)*/ -2969,
-	/*log(0.48462)*/ -2967,
-	/*log(0.48486)*/ -2965,
-	/*log(0.48511)*/ -2962,
-	/*log(0.48535)*/ -2960,
-	/*log(0.48560)*/ -2958,
-	/*log(0.48584)*/ -2956,
-	/*log(0.48608)*/ -2954,
-	/*log(0.48633)*/ -2952,
-	/*log(0.48657)*/ -2950,
-	/*log(0.48682)*/ -2948,
-	/*log(0.48706)*/ -2946,
-	/*log(0.48730)*/ -2944,
-	/*log(0.48755)*/ -2942,
-	/*log(0.48779)*/ -2940,
-	/*log(0.48804)*/ -2938,
-	/*log(0.48828)*/ -2936,
-	/*log(0.48853)*/ -2934,
-	/*log(0.48877)*/ -2932,
-	/*log(0.48901)*/ -2930,
-	/*log(0.48926)*/ -2928,
-	/*log(0.48950)*/ -2926,
-	/*log(0.48975)*/ -2924,
-	/*log(0.48999)*/ -2921,
-	/*log(0.49023)*/ -2919,
-	/*log(0.49048)*/ -2917,
-	/*log(0.49072)*/ -2915,
-	/*log(0.49097)*/ -2913,
-	/*log(0.49121)*/ -2911,
-	/*log(0.49146)*/ -2909,
-	/*log(0.49170)*/ -2907,
-	/*log(0.49194)*/ -2905,
-	/*log(0.49219)*/ -2903,
-	/*log(0.49243)*/ -2901,
-	/*log(0.49268)*/ -2899,
-	/*log(0.49292)*/ -2897,
-	/*log(0.49316)*/ -2895,
-	/*log(0.49341)*/ -2893,
-	/*log(0.49365)*/ -2891,
-	/*log(0.49390)*/ -2889,
-	/*log(0.49414)*/ -2887,
-	/*log(0.49438)*/ -2885,
-	/*log(0.49463)*/ -2883,
-	/*log(0.49487)*/ -2881,
-	/*log(0.49512)*/ -2879,
-	/*log(0.49536)*/ -2877,
-	/*log(0.49561)*/ -2875,
-	/*log(0.49585)*/ -2873,
-	/*log(0.49609)*/ -2871,
-	/*log(0.49634)*/ -2869,
-	/*log(0.49658)*/ -2867,
-	/*log(0.49683)*/ -2865,
-	/*log(0.49707)*/ -2863,
-	/*log(0.49731)*/ -2861,
-	/*log(0.49756)*/ -2859,
-	/*log(0.49780)*/ -2857,
-	/*log(0.49805)*/ -2855,
-	/*log(0.49829)*/ -2853,
-	/*log(0.49854)*/ -2851,
-	/*log(0.49878)*/ -2849,
-	/*log(0.49902)*/ -2847,
-	/*log(0.49927)*/ -2845,
-	/*log(0.49951)*/ -2843,
-	/*log(0.49976)*/ -2841,
-	/*log(0.50000)*/ -2839,
-	/*log(0.50024)*/ -2837,
-	/*log(0.50049)*/ -2835,
-	/*log(0.50073)*/ -2833,
-	/*log(0.50098)*/ -2831,
-	/*log(0.50122)*/ -2829,
-	/*log(0.50146)*/ -2827,
-	/*log(0.50171)*/ -2825,
-	/*log(0.50195)*/ -2823,
-	/*log(0.50220)*/ -2821,
-	/*log(0.50244)*/ -2819,
-	/*log(0.50269)*/ -2817,
-	/*log(0.50293)*/ -2815,
-	/*log(0.50317)*/ -2813,
-	/*log(0.50342)*/ -2811,
-	/*log(0.50366)*/ -2809,
-	/*log(0.50391)*/ -2807,
-	/*log(0.50415)*/ -2805,
-	/*log(0.50439)*/ -2803,
-	/*log(0.50464)*/ -2801,
-	/*log(0.50488)*/ -2799,
-	/*log(0.50513)*/ -2797,
-	/*log(0.50537)*/ -2795,
-	/*log(0.50562)*/ -2793,
-	/*log(0.50586)*/ -2791,
-	/*log(0.50610)*/ -2789,
-	/*log(0.50635)*/ -2787,
-	/*log(0.50659)*/ -2785,
-	/*log(0.50684)*/ -2783,
-	/*log(0.50708)*/ -2781,
-	/*log(0.50732)*/ -2779,
-	/*log(0.50757)*/ -2777,
-	/*log(0.50781)*/ -2775,
-	/*log(0.50806)*/ -2773,
-	/*log(0.50830)*/ -2771,
-	/*log(0.50854)*/ -2769,
-	/*log(0.50879)*/ -2767,
-	/*log(0.50903)*/ -2765,
-	/*log(0.50928)*/ -2763,
-	/*log(0.50952)*/ -2761,
-	/*log(0.50977)*/ -2759,
-	/*log(0.51001)*/ -2757,
-	/*log(0.51025)*/ -2755,
-	/*log(0.51050)*/ -2754,
-	/*log(0.51074)*/ -2752,
-	/*log(0.51099)*/ -2750,
-	/*log(0.51123)*/ -2748,
-	/*log(0.51147)*/ -2746,
-	/*log(0.51172)*/ -2744,
-	/*log(0.51196)*/ -2742,
-	/*log(0.51221)*/ -2740,
-	/*log(0.51245)*/ -2738,
-	/*log(0.51270)*/ -2736,
-	/*log(0.51294)*/ -2734,
-	/*log(0.51318)*/ -2732,
-	/*log(0.51343)*/ -2730,
-	/*log(0.51367)*/ -2728,
-	/*log(0.51392)*/ -2726,
-	/*log(0.51416)*/ -2724,
-	/*log(0.51440)*/ -2722,
-	/*log(0.51465)*/ -2720,
-	/*log(0.51489)*/ -2718,
-	/*log(0.51514)*/ -2716,
-	/*log(0.51538)*/ -2715,
-	/*log(0.51562)*/ -2713,
-	/*log(0.51587)*/ -2711,
-	/*log(0.51611)*/ -2709,
-	/*log(0.51636)*/ -2707,
-	/*log(0.51660)*/ -2705,
-	/*log(0.51685)*/ -2703,
-	/*log(0.51709)*/ -2701,
-	/*log(0.51733)*/ -2699,
-	/*log(0.51758)*/ -2697,
-	/*log(0.51782)*/ -2695,
-	/*log(0.51807)*/ -2693,
-	/*log(0.51831)*/ -2691,
-	/*log(0.51855)*/ -2689,
-	/*log(0.51880)*/ -2687,
-	/*log(0.51904)*/ -2686,
-	/*log(0.51929)*/ -2684,
-	/*log(0.51953)*/ -2682,
-	/*log(0.51978)*/ -2680,
-	/*log(0.52002)*/ -2678,
-	/*log(0.52026)*/ -2676,
-	/*log(0.52051)*/ -2674,
-	/*log(0.52075)*/ -2672,
-	/*log(0.52100)*/ -2670,
-	/*log(0.52124)*/ -2668,
-	/*log(0.52148)*/ -2666,
-	/*log(0.52173)*/ -2664,
-	/*log(0.52197)*/ -2662,
-	/*log(0.52222)*/ -2661,
-	/*log(0.52246)*/ -2659,
-	/*log(0.52271)*/ -2657,
-	/*log(0.52295)*/ -2655,
-	/*log(0.52319)*/ -2653,
-	/*log(0.52344)*/ -2651,
-	/*log(0.52368)*/ -2649,
-	/*log(0.52393)*/ -2647,
-	/*log(0.52417)*/ -2645,
-	/*log(0.52441)*/ -2643,
-	/*log(0.52466)*/ -2641,
-	/*log(0.52490)*/ -2640,
-	/*log(0.52515)*/ -2638,
-	/*log(0.52539)*/ -2636,
-	/*log(0.52563)*/ -2634,
-	/*log(0.52588)*/ -2632,
-	/*log(0.52612)*/ -2630,
-	/*log(0.52637)*/ -2628,
-	/*log(0.52661)*/ -2626,
-	/*log(0.52686)*/ -2624,
-	/*log(0.52710)*/ -2622,
-	/*log(0.52734)*/ -2621,
-	/*log(0.52759)*/ -2619,
-	/*log(0.52783)*/ -2617,
-	/*log(0.52808)*/ -2615,
-	/*log(0.52832)*/ -2613,
-	/*log(0.52856)*/ -2611,
-	/*log(0.52881)*/ -2609,
-	/*log(0.52905)*/ -2607,
-	/*log(0.52930)*/ -2605,
-	/*log(0.52954)*/ -2604,
-	/*log(0.52979)*/ -2602,
-	/*log(0.53003)*/ -2600,
-	/*log(0.53027)*/ -2598,
-	/*log(0.53052)*/ -2596,
-	/*log(0.53076)*/ -2594,
-	/*log(0.53101)*/ -2592,
-	/*log(0.53125)*/ -2590,
-	/*log(0.53149)*/ -2588,
-	/*log(0.53174)*/ -2587,
-	/*log(0.53198)*/ -2585,
-	/*log(0.53223)*/ -2583,
-	/*log(0.53247)*/ -2581,
-	/*log(0.53271)*/ -2579,
-	/*log(0.53296)*/ -2577,
-	/*log(0.53320)*/ -2575,
-	/*log(0.53345)*/ -2573,
-	/*log(0.53369)*/ -2572,
-	/*log(0.53394)*/ -2570,
-	/*log(0.53418)*/ -2568,
-	/*log(0.53442)*/ -2566,
-	/*log(0.53467)*/ -2564,
-	/*log(0.53491)*/ -2562,
-	/*log(0.53516)*/ -2560,
-	/*log(0.53540)*/ -2558,
-	/*log(0.53564)*/ -2557,
-	/*log(0.53589)*/ -2555,
-	/*log(0.53613)*/ -2553,
-	/*log(0.53638)*/ -2551,
-	/*log(0.53662)*/ -2549,
-	/*log(0.53687)*/ -2547,
-	/*log(0.53711)*/ -2545,
-	/*log(0.53735)*/ -2544,
-	/*log(0.53760)*/ -2542,
-	/*log(0.53784)*/ -2540,
-	/*log(0.53809)*/ -2538,
-	/*log(0.53833)*/ -2536,
-	/*log(0.53857)*/ -2534,
-	/*log(0.53882)*/ -2532,
-	/*log(0.53906)*/ -2531,
-	/*log(0.53931)*/ -2529,
-	/*log(0.53955)*/ -2527,
-	/*log(0.53979)*/ -2525,
-	/*log(0.54004)*/ -2523,
-	/*log(0.54028)*/ -2521,
-	/*log(0.54053)*/ -2519,
-	/*log(0.54077)*/ -2518,
-	/*log(0.54102)*/ -2516,
-	/*log(0.54126)*/ -2514,
-	/*log(0.54150)*/ -2512,
-	/*log(0.54175)*/ -2510,
-	/*log(0.54199)*/ -2508,
-	/*log(0.54224)*/ -2506,
-	/*log(0.54248)*/ -2505,
-	/*log(0.54272)*/ -2503,
-	/*log(0.54297)*/ -2501,
-	/*log(0.54321)*/ -2499,
-	/*log(0.54346)*/ -2497,
-	/*log(0.54370)*/ -2495,
-	/*log(0.54395)*/ -2494,
-	/*log(0.54419)*/ -2492,
-	/*log(0.54443)*/ -2490,
-	/*log(0.54468)*/ -2488,
-	/*log(0.54492)*/ -2486,
-	/*log(0.54517)*/ -2484,
-	/*log(0.54541)*/ -2483,
-	/*log(0.54565)*/ -2481,
-	/*log(0.54590)*/ -2479,
-	/*log(0.54614)*/ -2477,
-	/*log(0.54639)*/ -2475,
-	/*log(0.54663)*/ -2473,
-	/*log(0.54688)*/ -2472,
-	/*log(0.54712)*/ -2470,
-	/*log(0.54736)*/ -2468,
-	/*log(0.54761)*/ -2466,
-	/*log(0.54785)*/ -2464,
-	/*log(0.54810)*/ -2462,
-	/*log(0.54834)*/ -2461,
-	/*log(0.54858)*/ -2459,
-	/*log(0.54883)*/ -2457,
-	/*log(0.54907)*/ -2455,
-	/*log(0.54932)*/ -2453,
-	/*log(0.54956)*/ -2452,
-	/*log(0.54980)*/ -2450,
-	/*log(0.55005)*/ -2448,
-	/*log(0.55029)*/ -2446,
-	/*log(0.55054)*/ -2444,
-	/*log(0.55078)*/ -2442,
-	/*log(0.55103)*/ -2441,
-	/*log(0.55127)*/ -2439,
-	/*log(0.55151)*/ -2437,
-	/*log(0.55176)*/ -2435,
-	/*log(0.55200)*/ -2433,
-	/*log(0.55225)*/ -2432,
-	/*log(0.55249)*/ -2430,
-	/*log(0.55273)*/ -2428,
-	/*log(0.55298)*/ -2426,
-	/*log(0.55322)*/ -2424,
-	/*log(0.55347)*/ -2423,
-	/*log(0.55371)*/ -2421,
-	/*log(0.55396)*/ -2419,
-	/*log(0.55420)*/ -2417,
-	/*log(0.55444)*/ -2415,
-	/*log(0.55469)*/ -2413,
-	/*log(0.55493)*/ -2412,
-	/*log(0.55518)*/ -2410,
-	/*log(0.55542)*/ -2408,
-	/*log(0.55566)*/ -2406,
-	/*log(0.55591)*/ -2404,
-	/*log(0.55615)*/ -2403,
-	/*log(0.55640)*/ -2401,
-	/*log(0.55664)*/ -2399,
-	/*log(0.55688)*/ -2397,
-	/*log(0.55713)*/ -2395,
-	/*log(0.55737)*/ -2394,
-	/*log(0.55762)*/ -2392,
-	/*log(0.55786)*/ -2390,
-	/*log(0.55811)*/ -2388,
-	/*log(0.55835)*/ -2387,
-	/*log(0.55859)*/ -2385,
-	/*log(0.55884)*/ -2383,
-	/*log(0.55908)*/ -2381,
-	/*log(0.55933)*/ -2379,
-	/*log(0.55957)*/ -2378,
-	/*log(0.55981)*/ -2376,
-	/*log(0.56006)*/ -2374,
-	/*log(0.56030)*/ -2372,
-	/*log(0.56055)*/ -2370,
-	/*log(0.56079)*/ -2369,
-	/*log(0.56104)*/ -2367,
-	/*log(0.56128)*/ -2365,
-	/*log(0.56152)*/ -2363,
-	/*log(0.56177)*/ -2362,
-	/*log(0.56201)*/ -2360,
-	/*log(0.56226)*/ -2358,
-	/*log(0.56250)*/ -2356,
-	/*log(0.56274)*/ -2354,
-	/*log(0.56299)*/ -2353,
-	/*log(0.56323)*/ -2351,
-	/*log(0.56348)*/ -2349,
-	/*log(0.56372)*/ -2347,
-	/*log(0.56396)*/ -2346,
-	/*log(0.56421)*/ -2344,
-	/*log(0.56445)*/ -2342,
-	/*log(0.56470)*/ -2340,
-	/*log(0.56494)*/ -2338,
-	/*log(0.56519)*/ -2337,
-	/*log(0.56543)*/ -2335,
-	/*log(0.56567)*/ -2333,
-	/*log(0.56592)*/ -2331,
-	/*log(0.56616)*/ -2330,
-	/*log(0.56641)*/ -2328,
-	/*log(0.56665)*/ -2326,
-	/*log(0.56689)*/ -2324,
-	/*log(0.56714)*/ -2323,
-	/*log(0.56738)*/ -2321,
-	/*log(0.56763)*/ -2319,
-	/*log(0.56787)*/ -2317,
-	/*log(0.56812)*/ -2316,
-	/*log(0.56836)*/ -2314,
-	/*log(0.56860)*/ -2312,
-	/*log(0.56885)*/ -2310,
-	/*log(0.56909)*/ -2308,
-	/*log(0.56934)*/ -2307,
-	/*log(0.56958)*/ -2305,
-	/*log(0.56982)*/ -2303,
-	/*log(0.57007)*/ -2301,
-	/*log(0.57031)*/ -2300,
-	/*log(0.57056)*/ -2298,
-	/*log(0.57080)*/ -2296,
-	/*log(0.57104)*/ -2294,
-	/*log(0.57129)*/ -2293,
-	/*log(0.57153)*/ -2291,
-	/*log(0.57178)*/ -2289,
-	/*log(0.57202)*/ -2287,
-	/*log(0.57227)*/ -2286,
-	/*log(0.57251)*/ -2284,
-	/*log(0.57275)*/ -2282,
-	/*log(0.57300)*/ -2280,
-	/*log(0.57324)*/ -2279,
-	/*log(0.57349)*/ -2277,
-	/*log(0.57373)*/ -2275,
-	/*log(0.57397)*/ -2273,
-	/*log(0.57422)*/ -2272,
-	/*log(0.57446)*/ -2270,
-	/*log(0.57471)*/ -2268,
-	/*log(0.57495)*/ -2267,
-	/*log(0.57520)*/ -2265,
-	/*log(0.57544)*/ -2263,
-	/*log(0.57568)*/ -2261,
-	/*log(0.57593)*/ -2260,
-	/*log(0.57617)*/ -2258,
-	/*log(0.57642)*/ -2256,
-	/*log(0.57666)*/ -2254,
-	/*log(0.57690)*/ -2253,
-	/*log(0.57715)*/ -2251,
-	/*log(0.57739)*/ -2249,
-	/*log(0.57764)*/ -2247,
-	/*log(0.57788)*/ -2246,
-	/*log(0.57812)*/ -2244,
-	/*log(0.57837)*/ -2242,
-	/*log(0.57861)*/ -2241,
-	/*log(0.57886)*/ -2239,
-	/*log(0.57910)*/ -2237,
-	/*log(0.57935)*/ -2235,
-	/*log(0.57959)*/ -2234,
-	/*log(0.57983)*/ -2232,
-	/*log(0.58008)*/ -2230,
-	/*log(0.58032)*/ -2228,
-	/*log(0.58057)*/ -2227,
-	/*log(0.58081)*/ -2225,
-	/*log(0.58105)*/ -2223,
-	/*log(0.58130)*/ -2222,
-	/*log(0.58154)*/ -2220,
-	/*log(0.58179)*/ -2218,
-	/*log(0.58203)*/ -2216,
-	/*log(0.58228)*/ -2215,
-	/*log(0.58252)*/ -2213,
-	/*log(0.58276)*/ -2211,
-	/*log(0.58301)*/ -2210,
-	/*log(0.58325)*/ -2208,
-	/*log(0.58350)*/ -2206,
-	/*log(0.58374)*/ -2204,
-	/*log(0.58398)*/ -2203,
-	/*log(0.58423)*/ -2201,
-	/*log(0.58447)*/ -2199,
-	/*log(0.58472)*/ -2198,
-	/*log(0.58496)*/ -2196,
-	/*log(0.58521)*/ -2194,
-	/*log(0.58545)*/ -2192,
-	/*log(0.58569)*/ -2191,
-	/*log(0.58594)*/ -2189,
-	/*log(0.58618)*/ -2187,
-	/*log(0.58643)*/ -2186,
-	/*log(0.58667)*/ -2184,
-	/*log(0.58691)*/ -2182,
-	/*log(0.58716)*/ -2180,
-	/*log(0.58740)*/ -2179,
-	/*log(0.58765)*/ -2177,
-	/*log(0.58789)*/ -2175,
-	/*log(0.58813)*/ -2174,
-	/*log(0.58838)*/ -2172,
-	/*log(0.58862)*/ -2170,
-	/*log(0.58887)*/ -2169,
-	/*log(0.58911)*/ -2167,
-	/*log(0.58936)*/ -2165,
-	/*log(0.58960)*/ -2163,
-	/*log(0.58984)*/ -2162,
-	/*log(0.59009)*/ -2160,
-	/*log(0.59033)*/ -2158,
-	/*log(0.59058)*/ -2157,
-	/*log(0.59082)*/ -2155,
-	/*log(0.59106)*/ -2153,
-	/*log(0.59131)*/ -2152,
-	/*log(0.59155)*/ -2150,
-	/*log(0.59180)*/ -2148,
-	/*log(0.59204)*/ -2147,
-	/*log(0.59229)*/ -2145,
-	/*log(0.59253)*/ -2143,
-	/*log(0.59277)*/ -2141,
-	/*log(0.59302)*/ -2140,
-	/*log(0.59326)*/ -2138,
-	/*log(0.59351)*/ -2136,
-	/*log(0.59375)*/ -2135,
-	/*log(0.59399)*/ -2133,
-	/*log(0.59424)*/ -2131,
-	/*log(0.59448)*/ -2130,
-	/*log(0.59473)*/ -2128,
-	/*log(0.59497)*/ -2126,
-	/*log(0.59521)*/ -2125,
-	/*log(0.59546)*/ -2123,
-	/*log(0.59570)*/ -2121,
-	/*log(0.59595)*/ -2120,
-	/*log(0.59619)*/ -2118,
-	/*log(0.59644)*/ -2116,
-	/*log(0.59668)*/ -2115,
-	/*log(0.59692)*/ -2113,
-	/*log(0.59717)*/ -2111,
-	/*log(0.59741)*/ -2110,
-	/*log(0.59766)*/ -2108,
-	/*log(0.59790)*/ -2106,
-	/*log(0.59814)*/ -2105,
-	/*log(0.59839)*/ -2103,
-	/*log(0.59863)*/ -2101,
-	/*log(0.59888)*/ -2100,
-	/*log(0.59912)*/ -2098,
-	/*log(0.59937)*/ -2096,
-	/*log(0.59961)*/ -2095,
-	/*log(0.59985)*/ -2093,
-	/*log(0.60010)*/ -2091,
-	/*log(0.60034)*/ -2090,
-	/*log(0.60059)*/ -2088,
-	/*log(0.60083)*/ -2086,
-	/*log(0.60107)*/ -2085,
-	/*log(0.60132)*/ -2083,
-	/*log(0.60156)*/ -2081,
-	/*log(0.60181)*/ -2080,
-	/*log(0.60205)*/ -2078,
-	/*log(0.60229)*/ -2076,
-	/*log(0.60254)*/ -2075,
-	/*log(0.60278)*/ -2073,
-	/*log(0.60303)*/ -2071,
-	/*log(0.60327)*/ -2070,
-	/*log(0.60352)*/ -2068,
-	/*log(0.60376)*/ -2066,
-	/*log(0.60400)*/ -2065,
-	/*log(0.60425)*/ -2063,
-	/*log(0.60449)*/ -2061,
-	/*log(0.60474)*/ -2060,
-	/*log(0.60498)*/ -2058,
-	/*log(0.60522)*/ -2056,
-	/*log(0.60547)*/ -2055,
-	/*log(0.60571)*/ -2053,
-	/*log(0.60596)*/ -2051,
-	/*log(0.60620)*/ -2050,
-	/*log(0.60645)*/ -2048,
-	/*log(0.60669)*/ -2046,
-	/*log(0.60693)*/ -2045,
-	/*log(0.60718)*/ -2043,
-	/*log(0.60742)*/ -2041,
-	/*log(0.60767)*/ -2040,
-	/*log(0.60791)*/ -2038,
-	/*log(0.60815)*/ -2037,
-	/*log(0.60840)*/ -2035,
-	/*log(0.60864)*/ -2033,
-	/*log(0.60889)*/ -2032,
-	/*log(0.60913)*/ -2030,
-	/*log(0.60938)*/ -2028,
-	/*log(0.60962)*/ -2027,
-	/*log(0.60986)*/ -2025,
-	/*log(0.61011)*/ -2023,
-	/*log(0.61035)*/ -2022,
-	/*log(0.61060)*/ -2020,
-	/*log(0.61084)*/ -2019,
-	/*log(0.61108)*/ -2017,
-	/*log(0.61133)*/ -2015,
-	/*log(0.61157)*/ -2014,
-	/*log(0.61182)*/ -2012,
-	/*log(0.61206)*/ -2010,
-	/*log(0.61230)*/ -2009,
-	/*log(0.61255)*/ -2007,
-	/*log(0.61279)*/ -2005,
-	/*log(0.61304)*/ -2004,
-	/*log(0.61328)*/ -2002,
-	/*log(0.61353)*/ -2001,
-	/*log(0.61377)*/ -1999,
-	/*log(0.61401)*/ -1997,
-	/*log(0.61426)*/ -1996,
-	/*log(0.61450)*/ -1994,
-	/*log(0.61475)*/ -1992,
-	/*log(0.61499)*/ -1991,
-	/*log(0.61523)*/ -1989,
-	/*log(0.61548)*/ -1988,
-	/*log(0.61572)*/ -1986,
-	/*log(0.61597)*/ -1984,
-	/*log(0.61621)*/ -1983,
-	/*log(0.61646)*/ -1981,
-	/*log(0.61670)*/ -1979,
-	/*log(0.61694)*/ -1978,
-	/*log(0.61719)*/ -1976,
-	/*log(0.61743)*/ -1975,
-	/*log(0.61768)*/ -1973,
-	/*log(0.61792)*/ -1971,
-	/*log(0.61816)*/ -1970,
-	/*log(0.61841)*/ -1968,
-	/*log(0.61865)*/ -1966,
-	/*log(0.61890)*/ -1965,
-	/*log(0.61914)*/ -1963,
-	/*log(0.61938)*/ -1962,
-	/*log(0.61963)*/ -1960,
-	/*log(0.61987)*/ -1958,
-	/*log(0.62012)*/ -1957,
-	/*log(0.62036)*/ -1955,
-	/*log(0.62061)*/ -1954,
-	/*log(0.62085)*/ -1952,
-	/*log(0.62109)*/ -1950,
-	/*log(0.62134)*/ -1949,
-	/*log(0.62158)*/ -1947,
-	/*log(0.62183)*/ -1945,
-	/*log(0.62207)*/ -1944,
-	/*log(0.62231)*/ -1942,
-	/*log(0.62256)*/ -1941,
-	/*log(0.62280)*/ -1939,
-	/*log(0.62305)*/ -1937,
-	/*log(0.62329)*/ -1936,
-	/*log(0.62354)*/ -1934,
-	/*log(0.62378)*/ -1933,
-	/*log(0.62402)*/ -1931,
-	/*log(0.62427)*/ -1929,
-	/*log(0.62451)*/ -1928,
-	/*log(0.62476)*/ -1926,
-	/*log(0.62500)*/ -1925,
-	/*log(0.62524)*/ -1923,
-	/*log(0.62549)*/ -1921,
-	/*log(0.62573)*/ -1920,
-	/*log(0.62598)*/ -1918,
-	/*log(0.62622)*/ -1917,
-	/*log(0.62646)*/ -1915,
-	/*log(0.62671)*/ -1913,
-	/*log(0.62695)*/ -1912,
-	/*log(0.62720)*/ -1910,
-	/*log(0.62744)*/ -1909,
-	/*log(0.62769)*/ -1907,
-	/*log(0.62793)*/ -1905,
-	/*log(0.62817)*/ -1904,
-	/*log(0.62842)*/ -1902,
-	/*log(0.62866)*/ -1901,
-	/*log(0.62891)*/ -1899,
-	/*log(0.62915)*/ -1898,
-	/*log(0.62939)*/ -1896,
-	/*log(0.62964)*/ -1894,
-	/*log(0.62988)*/ -1893,
-	/*log(0.63013)*/ -1891,
-	/*log(0.63037)*/ -1890,
-	/*log(0.63062)*/ -1888,
-	/*log(0.63086)*/ -1886,
-	/*log(0.63110)*/ -1885,
-	/*log(0.63135)*/ -1883,
-	/*log(0.63159)*/ -1882,
-	/*log(0.63184)*/ -1880,
-	/*log(0.63208)*/ -1878,
-	/*log(0.63232)*/ -1877,
-	/*log(0.63257)*/ -1875,
-	/*log(0.63281)*/ -1874,
-	/*log(0.63306)*/ -1872,
-	/*log(0.63330)*/ -1871,
-	/*log(0.63354)*/ -1869,
-	/*log(0.63379)*/ -1867,
-	/*log(0.63403)*/ -1866,
-	/*log(0.63428)*/ -1864,
-	/*log(0.63452)*/ -1863,
-	/*log(0.63477)*/ -1861,
-	/*log(0.63501)*/ -1860,
-	/*log(0.63525)*/ -1858,
-	/*log(0.63550)*/ -1856,
-	/*log(0.63574)*/ -1855,
-	/*log(0.63599)*/ -1853,
-	/*log(0.63623)*/ -1852,
-	/*log(0.63647)*/ -1850,
-	/*log(0.63672)*/ -1849,
-	/*log(0.63696)*/ -1847,
-	/*log(0.63721)*/ -1845,
-	/*log(0.63745)*/ -1844,
-	/*log(0.63770)*/ -1842,
-	/*log(0.63794)*/ -1841,
-	/*log(0.63818)*/ -1839,
-	/*log(0.63843)*/ -1838,
-	/*log(0.63867)*/ -1836,
-	/*log(0.63892)*/ -1834,
-	/*log(0.63916)*/ -1833,
-	/*log(0.63940)*/ -1831,
-	/*log(0.63965)*/ -1830,
-	/*log(0.63989)*/ -1828,
-	/*log(0.64014)*/ -1827,
-	/*log(0.64038)*/ -1825,
-	/*log(0.64062)*/ -1823,
-	/*log(0.64087)*/ -1822,
-	/*log(0.64111)*/ -1820,
-	/*log(0.64136)*/ -1819,
-	/*log(0.64160)*/ -1817,
-	/*log(0.64185)*/ -1816,
-	/*log(0.64209)*/ -1814,
-	/*log(0.64233)*/ -1813,
-	/*log(0.64258)*/ -1811,
-	/*log(0.64282)*/ -1809,
-	/*log(0.64307)*/ -1808,
-	/*log(0.64331)*/ -1806,
-	/*log(0.64355)*/ -1805,
-	/*log(0.64380)*/ -1803,
-	/*log(0.64404)*/ -1802,
-	/*log(0.64429)*/ -1800,
-	/*log(0.64453)*/ -1799,
-	/*log(0.64478)*/ -1797,
-	/*log(0.64502)*/ -1795,
-	/*log(0.64526)*/ -1794,
-	/*log(0.64551)*/ -1792,
-	/*log(0.64575)*/ -1791,
-	/*log(0.64600)*/ -1789,
-	/*log(0.64624)*/ -1788,
-	/*log(0.64648)*/ -1786,
-	/*log(0.64673)*/ -1785,
-	/*log(0.64697)*/ -1783,
-	/*log(0.64722)*/ -1782,
-	/*log(0.64746)*/ -1780,
-	/*log(0.64771)*/ -1778,
-	/*log(0.64795)*/ -1777,
-	/*log(0.64819)*/ -1775,
-	/*log(0.64844)*/ -1774,
-	/*log(0.64868)*/ -1772,
-	/*log(0.64893)*/ -1771,
-	/*log(0.64917)*/ -1769,
-	/*log(0.64941)*/ -1768,
-	/*log(0.64966)*/ -1766,
-	/*log(0.64990)*/ -1765,
-	/*log(0.65015)*/ -1763,
-	/*log(0.65039)*/ -1762,
-	/*log(0.65063)*/ -1760,
-	/*log(0.65088)*/ -1758,
-	/*log(0.65112)*/ -1757,
-	/*log(0.65137)*/ -1755,
-	/*log(0.65161)*/ -1754,
-	/*log(0.65186)*/ -1752,
-	/*log(0.65210)*/ -1751,
-	/*log(0.65234)*/ -1749,
-	/*log(0.65259)*/ -1748,
-	/*log(0.65283)*/ -1746,
-	/*log(0.65308)*/ -1745,
-	/*log(0.65332)*/ -1743,
-	/*log(0.65356)*/ -1742,
-	/*log(0.65381)*/ -1740,
-	/*log(0.65405)*/ -1739,
-	/*log(0.65430)*/ -1737,
-	/*log(0.65454)*/ -1735,
-	/*log(0.65479)*/ -1734,
-	/*log(0.65503)*/ -1732,
-	/*log(0.65527)*/ -1731,
-	/*log(0.65552)*/ -1729,
-	/*log(0.65576)*/ -1728,
-	/*log(0.65601)*/ -1726,
-	/*log(0.65625)*/ -1725,
-	/*log(0.65649)*/ -1723,
-	/*log(0.65674)*/ -1722,
-	/*log(0.65698)*/ -1720,
-	/*log(0.65723)*/ -1719,
-	/*log(0.65747)*/ -1717,
-	/*log(0.65771)*/ -1716,
-	/*log(0.65796)*/ -1714,
-	/*log(0.65820)*/ -1713,
-	/*log(0.65845)*/ -1711,
-	/*log(0.65869)*/ -1710,
-	/*log(0.65894)*/ -1708,
-	/*log(0.65918)*/ -1707,
-	/*log(0.65942)*/ -1705,
-	/*log(0.65967)*/ -1704,
-	/*log(0.65991)*/ -1702,
-	/*log(0.66016)*/ -1700,
-	/*log(0.66040)*/ -1699,
-	/*log(0.66064)*/ -1697,
-	/*log(0.66089)*/ -1696,
-	/*log(0.66113)*/ -1694,
-	/*log(0.66138)*/ -1693,
-	/*log(0.66162)*/ -1691,
-	/*log(0.66187)*/ -1690,
-	/*log(0.66211)*/ -1688,
-	/*log(0.66235)*/ -1687,
-	/*log(0.66260)*/ -1685,
-	/*log(0.66284)*/ -1684,
-	/*log(0.66309)*/ -1682,
-	/*log(0.66333)*/ -1681,
-	/*log(0.66357)*/ -1679,
-	/*log(0.66382)*/ -1678,
-	/*log(0.66406)*/ -1676,
-	/*log(0.66431)*/ -1675,
-	/*log(0.66455)*/ -1673,
-	/*log(0.66479)*/ -1672,
-	/*log(0.66504)*/ -1670,
-	/*log(0.66528)*/ -1669,
-	/*log(0.66553)*/ -1667,
-	/*log(0.66577)*/ -1666,
-	/*log(0.66602)*/ -1664,
-	/*log(0.66626)*/ -1663,
-	/*log(0.66650)*/ -1661,
-	/*log(0.66675)*/ -1660,
-	/*log(0.66699)*/ -1658,
-	/*log(0.66724)*/ -1657,
-	/*log(0.66748)*/ -1655,
-	/*log(0.66772)*/ -1654,
-	/*log(0.66797)*/ -1652,
-	/*log(0.66821)*/ -1651,
-	/*log(0.66846)*/ -1649,
-	/*log(0.66870)*/ -1648,
-	/*log(0.66895)*/ -1646,
-	/*log(0.66919)*/ -1645,
-	/*log(0.66943)*/ -1643,
-	/*log(0.66968)*/ -1642,
-	/*log(0.66992)*/ -1640,
-	/*log(0.67017)*/ -1639,
-	/*log(0.67041)*/ -1637,
-	/*log(0.67065)*/ -1636,
-	/*log(0.67090)*/ -1634,
-	/*log(0.67114)*/ -1633,
-	/*log(0.67139)*/ -1631,
-	/*log(0.67163)*/ -1630,
-	/*log(0.67188)*/ -1628,
-	/*log(0.67212)*/ -1627,
-	/*log(0.67236)*/ -1625,
-	/*log(0.67261)*/ -1624,
-	/*log(0.67285)*/ -1622,
-	/*log(0.67310)*/ -1621,
-	/*log(0.67334)*/ -1619,
-	/*log(0.67358)*/ -1618,
-	/*log(0.67383)*/ -1617,
-	/*log(0.67407)*/ -1615,
-	/*log(0.67432)*/ -1614,
-	/*log(0.67456)*/ -1612,
-	/*log(0.67480)*/ -1611,
-	/*log(0.67505)*/ -1609,
-	/*log(0.67529)*/ -1608,
-	/*log(0.67554)*/ -1606,
-	/*log(0.67578)*/ -1605,
-	/*log(0.67603)*/ -1603,
-	/*log(0.67627)*/ -1602,
-	/*log(0.67651)*/ -1600,
-	/*log(0.67676)*/ -1599,
-	/*log(0.67700)*/ -1597,
-	/*log(0.67725)*/ -1596,
-	/*log(0.67749)*/ -1594,
-	/*log(0.67773)*/ -1593,
-	/*log(0.67798)*/ -1591,
-	/*log(0.67822)*/ -1590,
-	/*log(0.67847)*/ -1588,
-	/*log(0.67871)*/ -1587,
-	/*log(0.67896)*/ -1585,
-	/*log(0.67920)*/ -1584,
-	/*log(0.67944)*/ -1583,
-	/*log(0.67969)*/ -1581,
-	/*log(0.67993)*/ -1580,
-	/*log(0.68018)*/ -1578,
-	/*log(0.68042)*/ -1577,
-	/*log(0.68066)*/ -1575,
-	/*log(0.68091)*/ -1574,
-	/*log(0.68115)*/ -1572,
-	/*log(0.68140)*/ -1571,
-	/*log(0.68164)*/ -1569,
-	/*log(0.68188)*/ -1568,
-	/*log(0.68213)*/ -1566,
-	/*log(0.68237)*/ -1565,
-	/*log(0.68262)*/ -1563,
-	/*log(0.68286)*/ -1562,
-	/*log(0.68311)*/ -1561,
-	/*log(0.68335)*/ -1559,
-	/*log(0.68359)*/ -1558,
-	/*log(0.68384)*/ -1556,
-	/*log(0.68408)*/ -1555,
-	/*log(0.68433)*/ -1553,
-	/*log(0.68457)*/ -1552,
-	/*log(0.68481)*/ -1550,
-	/*log(0.68506)*/ -1549,
-	/*log(0.68530)*/ -1547,
-	/*log(0.68555)*/ -1546,
-	/*log(0.68579)*/ -1544,
-	/*log(0.68604)*/ -1543,
-	/*log(0.68628)*/ -1542,
-	/*log(0.68652)*/ -1540,
-	/*log(0.68677)*/ -1539,
-	/*log(0.68701)*/ -1537,
-	/*log(0.68726)*/ -1536,
-	/*log(0.68750)*/ -1534,
-	/*log(0.68774)*/ -1533,
-	/*log(0.68799)*/ -1531,
-	/*log(0.68823)*/ -1530,
-	/*log(0.68848)*/ -1528,
-	/*log(0.68872)*/ -1527,
-	/*log(0.68896)*/ -1526,
-	/*log(0.68921)*/ -1524,
-	/*log(0.68945)*/ -1523,
-	/*log(0.68970)*/ -1521,
-	/*log(0.68994)*/ -1520,
-	/*log(0.69019)*/ -1518,
-	/*log(0.69043)*/ -1517,
-	/*log(0.69067)*/ -1515,
-	/*log(0.69092)*/ -1514,
-	/*log(0.69116)*/ -1512,
-	/*log(0.69141)*/ -1511,
-	/*log(0.69165)*/ -1510,
-	/*log(0.69189)*/ -1508,
-	/*log(0.69214)*/ -1507,
-	/*log(0.69238)*/ -1505,
-	/*log(0.69263)*/ -1504,
-	/*log(0.69287)*/ -1502,
-	/*log(0.69312)*/ -1501,
-	/*log(0.69336)*/ -1499,
-	/*log(0.69360)*/ -1498,
-	/*log(0.69385)*/ -1497,
-	/*log(0.69409)*/ -1495,
-	/*log(0.69434)*/ -1494,
-	/*log(0.69458)*/ -1492,
-	/*log(0.69482)*/ -1491,
-	/*log(0.69507)*/ -1489,
-	/*log(0.69531)*/ -1488,
-	/*log(0.69556)*/ -1487,
-	/*log(0.69580)*/ -1485,
-	/*log(0.69604)*/ -1484,
-	/*log(0.69629)*/ -1482,
-	/*log(0.69653)*/ -1481,
-	/*log(0.69678)*/ -1479,
-	/*log(0.69702)*/ -1478,
-	/*log(0.69727)*/ -1476,
-	/*log(0.69751)*/ -1475,
-	/*log(0.69775)*/ -1474,
-	/*log(0.69800)*/ -1472,
-	/*log(0.69824)*/ -1471,
-	/*log(0.69849)*/ -1469,
-	/*log(0.69873)*/ -1468,
-	/*log(0.69897)*/ -1466,
-	/*log(0.69922)*/ -1465,
-	/*log(0.69946)*/ -1464,
-	/*log(0.69971)*/ -1462,
-	/*log(0.69995)*/ -1461,
-	/*log(0.70020)*/ -1459,
-	/*log(0.70044)*/ -1458,
-	/*log(0.70068)*/ -1456,
-	/*log(0.70093)*/ -1455,
-	/*log(0.70117)*/ -1454,
-	/*log(0.70142)*/ -1452,
-	/*log(0.70166)*/ -1451,
-	/*log(0.70190)*/ -1449,
-	/*log(0.70215)*/ -1448,
-	/*log(0.70239)*/ -1446,
-	/*log(0.70264)*/ -1445,
-	/*log(0.70288)*/ -1444,
-	/*log(0.70312)*/ -1442,
-	/*log(0.70337)*/ -1441,
-	/*log(0.70361)*/ -1439,
-	/*log(0.70386)*/ -1438,
-	/*log(0.70410)*/ -1437,
-	/*log(0.70435)*/ -1435,
-	/*log(0.70459)*/ -1434,
-	/*log(0.70483)*/ -1432,
-	/*log(0.70508)*/ -1431,
-	/*log(0.70532)*/ -1429,
-	/*log(0.70557)*/ -1428,
-	/*log(0.70581)*/ -1427,
-	/*log(0.70605)*/ -1425,
-	/*log(0.70630)*/ -1424,
-	/*log(0.70654)*/ -1422,
-	/*log(0.70679)*/ -1421,
-	/*log(0.70703)*/ -1420,
-	/*log(0.70728)*/ -1418,
-	/*log(0.70752)*/ -1417,
-	/*log(0.70776)*/ -1415,
-	/*log(0.70801)*/ -1414,
-	/*log(0.70825)*/ -1412,
-	/*log(0.70850)*/ -1411,
-	/*log(0.70874)*/ -1410,
-	/*log(0.70898)*/ -1408,
-	/*log(0.70923)*/ -1407,
-	/*log(0.70947)*/ -1405,
-	/*log(0.70972)*/ -1404,
-	/*log(0.70996)*/ -1403,
-	/*log(0.71021)*/ -1401,
-	/*log(0.71045)*/ -1400,
-	/*log(0.71069)*/ -1398,
-	/*log(0.71094)*/ -1397,
-	/*log(0.71118)*/ -1396,
-	/*log(0.71143)*/ -1394,
-	/*log(0.71167)*/ -1393,
-	/*log(0.71191)*/ -1391,
-	/*log(0.71216)*/ -1390,
-	/*log(0.71240)*/ -1389,
-	/*log(0.71265)*/ -1387,
-	/*log(0.71289)*/ -1386,
-	/*log(0.71313)*/ -1384,
-	/*log(0.71338)*/ -1383,
-	/*log(0.71362)*/ -1381,
-	/*log(0.71387)*/ -1380,
-	/*log(0.71411)*/ -1379,
-	/*log(0.71436)*/ -1377,
-	/*log(0.71460)*/ -1376,
-	/*log(0.71484)*/ -1374,
-	/*log(0.71509)*/ -1373,
-	/*log(0.71533)*/ -1372,
-	/*log(0.71558)*/ -1370,
-	/*log(0.71582)*/ -1369,
-	/*log(0.71606)*/ -1368,
-	/*log(0.71631)*/ -1366,
-	/*log(0.71655)*/ -1365,
-	/*log(0.71680)*/ -1363,
-	/*log(0.71704)*/ -1362,
-	/*log(0.71729)*/ -1361,
-	/*log(0.71753)*/ -1359,
-	/*log(0.71777)*/ -1358,
-	/*log(0.71802)*/ -1356,
-	/*log(0.71826)*/ -1355,
-	/*log(0.71851)*/ -1354,
-	/*log(0.71875)*/ -1352,
-	/*log(0.71899)*/ -1351,
-	/*log(0.71924)*/ -1349,
-	/*log(0.71948)*/ -1348,
-	/*log(0.71973)*/ -1347,
-	/*log(0.71997)*/ -1345,
-	/*log(0.72021)*/ -1344,
-	/*log(0.72046)*/ -1342,
-	/*log(0.72070)*/ -1341,
-	/*log(0.72095)*/ -1340,
-	/*log(0.72119)*/ -1338,
-	/*log(0.72144)*/ -1337,
-	/*log(0.72168)*/ -1336,
-	/*log(0.72192)*/ -1334,
-	/*log(0.72217)*/ -1333,
-	/*log(0.72241)*/ -1331,
-	/*log(0.72266)*/ -1330,
-	/*log(0.72290)*/ -1329,
-	/*log(0.72314)*/ -1327,
-	/*log(0.72339)*/ -1326,
-	/*log(0.72363)*/ -1324,
-	/*log(0.72388)*/ -1323,
-	/*log(0.72412)*/ -1322,
-	/*log(0.72437)*/ -1320,
-	/*log(0.72461)*/ -1319,
-	/*log(0.72485)*/ -1318,
-	/*log(0.72510)*/ -1316,
-	/*log(0.72534)*/ -1315,
-	/*log(0.72559)*/ -1313,
-	/*log(0.72583)*/ -1312,
-	/*log(0.72607)*/ -1311,
-	/*log(0.72632)*/ -1309,
-	/*log(0.72656)*/ -1308,
-	/*log(0.72681)*/ -1307,
-	/*log(0.72705)*/ -1305,
-	/*log(0.72729)*/ -1304,
-	/*log(0.72754)*/ -1302,
-	/*log(0.72778)*/ -1301,
-	/*log(0.72803)*/ -1300,
-	/*log(0.72827)*/ -1298,
-	/*log(0.72852)*/ -1297,
-	/*log(0.72876)*/ -1296,
-	/*log(0.72900)*/ -1294,
-	/*log(0.72925)*/ -1293,
-	/*log(0.72949)*/ -1291,
-	/*log(0.72974)*/ -1290,
-	/*log(0.72998)*/ -1289,
-	/*log(0.73022)*/ -1287,
-	/*log(0.73047)*/ -1286,
-	/*log(0.73071)*/ -1285,
-	/*log(0.73096)*/ -1283,
-	/*log(0.73120)*/ -1282,
-	/*log(0.73145)*/ -1280,
-	/*log(0.73169)*/ -1279,
-	/*log(0.73193)*/ -1278,
-	/*log(0.73218)*/ -1276,
-	/*log(0.73242)*/ -1275,
-	/*log(0.73267)*/ -1274,
-	/*log(0.73291)*/ -1272,
-	/*log(0.73315)*/ -1271,
-	/*log(0.73340)*/ -1270,
-	/*log(0.73364)*/ -1268,
-	/*log(0.73389)*/ -1267,
-	/*log(0.73413)*/ -1265,
-	/*log(0.73438)*/ -1264,
-	/*log(0.73462)*/ -1263,
-	/*log(0.73486)*/ -1261,
-	/*log(0.73511)*/ -1260,
-	/*log(0.73535)*/ -1259,
-	/*log(0.73560)*/ -1257,
-	/*log(0.73584)*/ -1256,
-	/*log(0.73608)*/ -1255,
-	/*log(0.73633)*/ -1253,
-	/*log(0.73657)*/ -1252,
-	/*log(0.73682)*/ -1250,
-	/*log(0.73706)*/ -1249,
-	/*log(0.73730)*/ -1248,
-	/*log(0.73755)*/ -1246,
-	/*log(0.73779)*/ -1245,
-	/*log(0.73804)*/ -1244,
-	/*log(0.73828)*/ -1242,
-	/*log(0.73853)*/ -1241,
-	/*log(0.73877)*/ -1240,
-	/*log(0.73901)*/ -1238,
-	/*log(0.73926)*/ -1237,
-	/*log(0.73950)*/ -1236,
-	/*log(0.73975)*/ -1234,
-	/*log(0.73999)*/ -1233,
-	/*log(0.74023)*/ -1232,
-	/*log(0.74048)*/ -1230,
-	/*log(0.74072)*/ -1229,
-	/*log(0.74097)*/ -1227,
-	/*log(0.74121)*/ -1226,
-	/*log(0.74146)*/ -1225,
-	/*log(0.74170)*/ -1223,
-	/*log(0.74194)*/ -1222,
-	/*log(0.74219)*/ -1221,
-	/*log(0.74243)*/ -1219,
-	/*log(0.74268)*/ -1218,
-	/*log(0.74292)*/ -1217,
-	/*log(0.74316)*/ -1215,
-	/*log(0.74341)*/ -1214,
-	/*log(0.74365)*/ -1213,
-	/*log(0.74390)*/ -1211,
-	/*log(0.74414)*/ -1210,
-	/*log(0.74438)*/ -1209,
-	/*log(0.74463)*/ -1207,
-	/*log(0.74487)*/ -1206,
-	/*log(0.74512)*/ -1205,
-	/*log(0.74536)*/ -1203,
-	/*log(0.74561)*/ -1202,
-	/*log(0.74585)*/ -1201,
-	/*log(0.74609)*/ -1199,
-	/*log(0.74634)*/ -1198,
-	/*log(0.74658)*/ -1197,
-	/*log(0.74683)*/ -1195,
-	/*log(0.74707)*/ -1194,
-	/*log(0.74731)*/ -1193,
-	/*log(0.74756)*/ -1191,
-	/*log(0.74780)*/ -1190,
-	/*log(0.74805)*/ -1189,
-	/*log(0.74829)*/ -1187,
-	/*log(0.74854)*/ -1186,
-	/*log(0.74878)*/ -1185,
-	/*log(0.74902)*/ -1183,
-	/*log(0.74927)*/ -1182,
-	/*log(0.74951)*/ -1181,
-	/*log(0.74976)*/ -1179,
-	/*log(0.75000)*/ -1178,
-	/*log(0.75024)*/ -1177,
-	/*log(0.75049)*/ -1175,
-	/*log(0.75073)*/ -1174,
-	/*log(0.75098)*/ -1173,
-	/*log(0.75122)*/ -1171,
-	/*log(0.75146)*/ -1170,
-	/*log(0.75171)*/ -1169,
-	/*log(0.75195)*/ -1167,
-	/*log(0.75220)*/ -1166,
-	/*log(0.75244)*/ -1165,
-	/*log(0.75269)*/ -1163,
-	/*log(0.75293)*/ -1162,
-	/*log(0.75317)*/ -1161,
-	/*log(0.75342)*/ -1159,
-	/*log(0.75366)*/ -1158,
-	/*log(0.75391)*/ -1157,
-	/*log(0.75415)*/ -1155,
-	/*log(0.75439)*/ -1154,
-	/*log(0.75464)*/ -1153,
-	/*log(0.75488)*/ -1151,
-	/*log(0.75513)*/ -1150,
-	/*log(0.75537)*/ -1149,
-	/*log(0.75562)*/ -1147,
-	/*log(0.75586)*/ -1146,
-	/*log(0.75610)*/ -1145,
-	/*log(0.75635)*/ -1143,
-	/*log(0.75659)*/ -1142,
-	/*log(0.75684)*/ -1141,
-	/*log(0.75708)*/ -1139,
-	/*log(0.75732)*/ -1138,
-	/*log(0.75757)*/ -1137,
-	/*log(0.75781)*/ -1135,
-	/*log(0.75806)*/ -1134,
-	/*log(0.75830)*/ -1133,
-	/*log(0.75854)*/ -1131,
-	/*log(0.75879)*/ -1130,
-	/*log(0.75903)*/ -1129,
-	/*log(0.75928)*/ -1127,
-	/*log(0.75952)*/ -1126,
-	/*log(0.75977)*/ -1125,
-	/*log(0.76001)*/ -1124,
-	/*log(0.76025)*/ -1122,
-	/*log(0.76050)*/ -1121,
-	/*log(0.76074)*/ -1120,
-	/*log(0.76099)*/ -1118,
-	/*log(0.76123)*/ -1117,
-	/*log(0.76147)*/ -1116,
-	/*log(0.76172)*/ -1114,
-	/*log(0.76196)*/ -1113,
-	/*log(0.76221)*/ -1112,
-	/*log(0.76245)*/ -1110,
-	/*log(0.76270)*/ -1109,
-	/*log(0.76294)*/ -1108,
-	/*log(0.76318)*/ -1106,
-	/*log(0.76343)*/ -1105,
-	/*log(0.76367)*/ -1104,
-	/*log(0.76392)*/ -1103,
-	/*log(0.76416)*/ -1101,
-	/*log(0.76440)*/ -1100,
-	/*log(0.76465)*/ -1099,
-	/*log(0.76489)*/ -1097,
-	/*log(0.76514)*/ -1096,
-	/*log(0.76538)*/ -1095,
-	/*log(0.76562)*/ -1093,
-	/*log(0.76587)*/ -1092,
-	/*log(0.76611)*/ -1091,
-	/*log(0.76636)*/ -1089,
-	/*log(0.76660)*/ -1088,
-	/*log(0.76685)*/ -1087,
-	/*log(0.76709)*/ -1086,
-	/*log(0.76733)*/ -1084,
-	/*log(0.76758)*/ -1083,
-	/*log(0.76782)*/ -1082,
-	/*log(0.76807)*/ -1080,
-	/*log(0.76831)*/ -1079,
-	/*log(0.76855)*/ -1078,
-	/*log(0.76880)*/ -1076,
-	/*log(0.76904)*/ -1075,
-	/*log(0.76929)*/ -1074,
-	/*log(0.76953)*/ -1073,
-	/*log(0.76978)*/ -1071,
-	/*log(0.77002)*/ -1070,
-	/*log(0.77026)*/ -1069,
-	/*log(0.77051)*/ -1067,
-	/*log(0.77075)*/ -1066,
-	/*log(0.77100)*/ -1065,
-	/*log(0.77124)*/ -1063,
-	/*log(0.77148)*/ -1062,
-	/*log(0.77173)*/ -1061,
-	/*log(0.77197)*/ -1060,
-	/*log(0.77222)*/ -1058,
-	/*log(0.77246)*/ -1057,
-	/*log(0.77271)*/ -1056,
-	/*log(0.77295)*/ -1054,
-	/*log(0.77319)*/ -1053,
-	/*log(0.77344)*/ -1052,
-	/*log(0.77368)*/ -1051,
-	/*log(0.77393)*/ -1049,
-	/*log(0.77417)*/ -1048,
-	/*log(0.77441)*/ -1047,
-	/*log(0.77466)*/ -1045,
-	/*log(0.77490)*/ -1044,
-	/*log(0.77515)*/ -1043,
-	/*log(0.77539)*/ -1041,
-	/*log(0.77563)*/ -1040,
-	/*log(0.77588)*/ -1039,
-	/*log(0.77612)*/ -1038,
-	/*log(0.77637)*/ -1036,
-	/*log(0.77661)*/ -1035,
-	/*log(0.77686)*/ -1034,
-	/*log(0.77710)*/ -1032,
-	/*log(0.77734)*/ -1031,
-	/*log(0.77759)*/ -1030,
-	/*log(0.77783)*/ -1029,
-	/*log(0.77808)*/ -1027,
-	/*log(0.77832)*/ -1026,
-	/*log(0.77856)*/ -1025,
-	/*log(0.77881)*/ -1023,
-	/*log(0.77905)*/ -1022,
-	/*log(0.77930)*/ -1021,
-	/*log(0.77954)*/ -1020,
-	/*log(0.77979)*/ -1018,
-	/*log(0.78003)*/ -1017,
-	/*log(0.78027)*/ -1016,
-	/*log(0.78052)*/ -1014,
-	/*log(0.78076)*/ -1013,
-	/*log(0.78101)*/ -1012,
-	/*log(0.78125)*/ -1011,
-	/*log(0.78149)*/ -1009,
-	/*log(0.78174)*/ -1008,
-	/*log(0.78198)*/ -1007,
-	/*log(0.78223)*/ -1006,
-	/*log(0.78247)*/ -1004,
-	/*log(0.78271)*/ -1003,
-	/*log(0.78296)*/ -1002,
-	/*log(0.78320)*/ -1000,
-	/*log(0.78345)*/ -999,
-	/*log(0.78369)*/ -998,
-	/*log(0.78394)*/ -997,
-	/*log(0.78418)*/ -995,
-	/*log(0.78442)*/ -994,
-	/*log(0.78467)*/ -993,
-	/*log(0.78491)*/ -991,
-	/*log(0.78516)*/ -990,
-	/*log(0.78540)*/ -989,
-	/*log(0.78564)*/ -988,
-	/*log(0.78589)*/ -986,
-	/*log(0.78613)*/ -985,
-	/*log(0.78638)*/ -984,
-	/*log(0.78662)*/ -983,
-	/*log(0.78687)*/ -981,
-	/*log(0.78711)*/ -980,
-	/*log(0.78735)*/ -979,
-	/*log(0.78760)*/ -977,
-	/*log(0.78784)*/ -976,
-	/*log(0.78809)*/ -975,
-	/*log(0.78833)*/ -974,
-	/*log(0.78857)*/ -972,
-	/*log(0.78882)*/ -971,
-	/*log(0.78906)*/ -970,
-	/*log(0.78931)*/ -969,
-	/*log(0.78955)*/ -967,
-	/*log(0.78979)*/ -966,
-	/*log(0.79004)*/ -965,
-	/*log(0.79028)*/ -964,
-	/*log(0.79053)*/ -962,
-	/*log(0.79077)*/ -961,
-	/*log(0.79102)*/ -960,
-	/*log(0.79126)*/ -958,
-	/*log(0.79150)*/ -957,
-	/*log(0.79175)*/ -956,
-	/*log(0.79199)*/ -955,
-	/*log(0.79224)*/ -953,
-	/*log(0.79248)*/ -952,
-	/*log(0.79272)*/ -951,
-	/*log(0.79297)*/ -950,
-	/*log(0.79321)*/ -948,
-	/*log(0.79346)*/ -947,
-	/*log(0.79370)*/ -946,
-	/*log(0.79395)*/ -945,
-	/*log(0.79419)*/ -943,
-	/*log(0.79443)*/ -942,
-	/*log(0.79468)*/ -941,
-	/*log(0.79492)*/ -940,
-	/*log(0.79517)*/ -938,
-	/*log(0.79541)*/ -937,
-	/*log(0.79565)*/ -936,
-	/*log(0.79590)*/ -935,
-	/*log(0.79614)*/ -933,
-	/*log(0.79639)*/ -932,
-	/*log(0.79663)*/ -931,
-	/*log(0.79688)*/ -930,
-	/*log(0.79712)*/ -928,
-	/*log(0.79736)*/ -927,
-	/*log(0.79761)*/ -926,
-	/*log(0.79785)*/ -925,
-	/*log(0.79810)*/ -923,
-	/*log(0.79834)*/ -922,
-	/*log(0.79858)*/ -921,
-	/*log(0.79883)*/ -920,
-	/*log(0.79907)*/ -918,
-	/*log(0.79932)*/ -917,
-	/*log(0.79956)*/ -916,
-	/*log(0.79980)*/ -914,
-	/*log(0.80005)*/ -913,
-	/*log(0.80029)*/ -912,
-	/*log(0.80054)*/ -911,
-	/*log(0.80078)*/ -909,
-	/*log(0.80103)*/ -908,
-	/*log(0.80127)*/ -907,
-	/*log(0.80151)*/ -906,
-	/*log(0.80176)*/ -905,
-	/*log(0.80200)*/ -903,
-	/*log(0.80225)*/ -902,
-	/*log(0.80249)*/ -901,
-	/*log(0.80273)*/ -900,
-	/*log(0.80298)*/ -898,
-	/*log(0.80322)*/ -897,
-	/*log(0.80347)*/ -896,
-	/*log(0.80371)*/ -895,
-	/*log(0.80396)*/ -893,
-	/*log(0.80420)*/ -892,
-	/*log(0.80444)*/ -891,
-	/*log(0.80469)*/ -890,
-	/*log(0.80493)*/ -888,
-	/*log(0.80518)*/ -887,
-	/*log(0.80542)*/ -886,
-	/*log(0.80566)*/ -885,
-	/*log(0.80591)*/ -883,
-	/*log(0.80615)*/ -882,
-	/*log(0.80640)*/ -881,
-	/*log(0.80664)*/ -880,
-	/*log(0.80688)*/ -878,
-	/*log(0.80713)*/ -877,
-	/*log(0.80737)*/ -876,
-	/*log(0.80762)*/ -875,
-	/*log(0.80786)*/ -873,
-	/*log(0.80811)*/ -872,
-	/*log(0.80835)*/ -871,
-	/*log(0.80859)*/ -870,
-	/*log(0.80884)*/ -868,
-	/*log(0.80908)*/ -867,
-	/*log(0.80933)*/ -866,
-	/*log(0.80957)*/ -865,
-	/*log(0.80981)*/ -864,
-	/*log(0.81006)*/ -862,
-	/*log(0.81030)*/ -861,
-	/*log(0.81055)*/ -860,
-	/*log(0.81079)*/ -859,
-	/*log(0.81104)*/ -857,
-	/*log(0.81128)*/ -856,
-	/*log(0.81152)*/ -855,
-	/*log(0.81177)*/ -854,
-	/*log(0.81201)*/ -852,
-	/*log(0.81226)*/ -851,
-	/*log(0.81250)*/ -850,
-	/*log(0.81274)*/ -849,
-	/*log(0.81299)*/ -848,
-	/*log(0.81323)*/ -846,
-	/*log(0.81348)*/ -845,
-	/*log(0.81372)*/ -844,
-	/*log(0.81396)*/ -843,
-	/*log(0.81421)*/ -841,
-	/*log(0.81445)*/ -840,
-	/*log(0.81470)*/ -839,
-	/*log(0.81494)*/ -838,
-	/*log(0.81519)*/ -836,
-	/*log(0.81543)*/ -835,
-	/*log(0.81567)*/ -834,
-	/*log(0.81592)*/ -833,
-	/*log(0.81616)*/ -832,
-	/*log(0.81641)*/ -830,
-	/*log(0.81665)*/ -829,
-	/*log(0.81689)*/ -828,
-	/*log(0.81714)*/ -827,
-	/*log(0.81738)*/ -825,
-	/*log(0.81763)*/ -824,
-	/*log(0.81787)*/ -823,
-	/*log(0.81812)*/ -822,
-	/*log(0.81836)*/ -821,
-	/*log(0.81860)*/ -819,
-	/*log(0.81885)*/ -818,
-	/*log(0.81909)*/ -817,
-	/*log(0.81934)*/ -816,
-	/*log(0.81958)*/ -814,
-	/*log(0.81982)*/ -813,
-	/*log(0.82007)*/ -812,
-	/*log(0.82031)*/ -811,
-	/*log(0.82056)*/ -810,
-	/*log(0.82080)*/ -808,
-	/*log(0.82104)*/ -807,
-	/*log(0.82129)*/ -806,
-	/*log(0.82153)*/ -805,
-	/*log(0.82178)*/ -803,
-	/*log(0.82202)*/ -802,
-	/*log(0.82227)*/ -801,
-	/*log(0.82251)*/ -800,
-	/*log(0.82275)*/ -799,
-	/*log(0.82300)*/ -797,
-	/*log(0.82324)*/ -796,
-	/*log(0.82349)*/ -795,
-	/*log(0.82373)*/ -794,
-	/*log(0.82397)*/ -793,
-	/*log(0.82422)*/ -791,
-	/*log(0.82446)*/ -790,
-	/*log(0.82471)*/ -789,
-	/*log(0.82495)*/ -788,
-	/*log(0.82520)*/ -786,
-	/*log(0.82544)*/ -785,
-	/*log(0.82568)*/ -784,
-	/*log(0.82593)*/ -783,
-	/*log(0.82617)*/ -782,
-	/*log(0.82642)*/ -780,
-	/*log(0.82666)*/ -779,
-	/*log(0.82690)*/ -778,
-	/*log(0.82715)*/ -777,
-	/*log(0.82739)*/ -776,
-	/*log(0.82764)*/ -774,
-	/*log(0.82788)*/ -773,
-	/*log(0.82812)*/ -772,
-	/*log(0.82837)*/ -771,
-	/*log(0.82861)*/ -770,
-	/*log(0.82886)*/ -768,
-	/*log(0.82910)*/ -767,
-	/*log(0.82935)*/ -766,
-	/*log(0.82959)*/ -765,
-	/*log(0.82983)*/ -764,
-	/*log(0.83008)*/ -762,
-	/*log(0.83032)*/ -761,
-	/*log(0.83057)*/ -760,
-	/*log(0.83081)*/ -759,
-	/*log(0.83105)*/ -758,
-	/*log(0.83130)*/ -756,
-	/*log(0.83154)*/ -755,
-	/*log(0.83179)*/ -754,
-	/*log(0.83203)*/ -753,
-	/*log(0.83228)*/ -751,
-	/*log(0.83252)*/ -750,
-	/*log(0.83276)*/ -749,
-	/*log(0.83301)*/ -748,
-	/*log(0.83325)*/ -747,
-	/*log(0.83350)*/ -745,
-	/*log(0.83374)*/ -744,
-	/*log(0.83398)*/ -743,
-	/*log(0.83423)*/ -742,
-	/*log(0.83447)*/ -741,
-	/*log(0.83472)*/ -739,
-	/*log(0.83496)*/ -738,
-	/*log(0.83521)*/ -737,
-	/*log(0.83545)*/ -736,
-	/*log(0.83569)*/ -735,
-	/*log(0.83594)*/ -734,
-	/*log(0.83618)*/ -732,
-	/*log(0.83643)*/ -731,
-	/*log(0.83667)*/ -730,
-	/*log(0.83691)*/ -729,
-	/*log(0.83716)*/ -728,
-	/*log(0.83740)*/ -726,
-	/*log(0.83765)*/ -725,
-	/*log(0.83789)*/ -724,
-	/*log(0.83813)*/ -723,
-	/*log(0.83838)*/ -722,
-	/*log(0.83862)*/ -720,
-	/*log(0.83887)*/ -719,
-	/*log(0.83911)*/ -718,
-	/*log(0.83936)*/ -717,
-	/*log(0.83960)*/ -716,
-	/*log(0.83984)*/ -714,
-	/*log(0.84009)*/ -713,
-	/*log(0.84033)*/ -712,
-	/*log(0.84058)*/ -711,
-	/*log(0.84082)*/ -710,
-	/*log(0.84106)*/ -708,
-	/*log(0.84131)*/ -707,
-	/*log(0.84155)*/ -706,
-	/*log(0.84180)*/ -705,
-	/*log(0.84204)*/ -704,
-	/*log(0.84229)*/ -703,
-	/*log(0.84253)*/ -701,
-	/*log(0.84277)*/ -700,
-	/*log(0.84302)*/ -699,
-	/*log(0.84326)*/ -698,
-	/*log(0.84351)*/ -697,
-	/*log(0.84375)*/ -695,
-	/*log(0.84399)*/ -694,
-	/*log(0.84424)*/ -693,
-	/*log(0.84448)*/ -692,
-	/*log(0.84473)*/ -691,
-	/*log(0.84497)*/ -689,
-	/*log(0.84521)*/ -688,
-	/*log(0.84546)*/ -687,
-	/*log(0.84570)*/ -686,
-	/*log(0.84595)*/ -685,
-	/*log(0.84619)*/ -684,
-	/*log(0.84644)*/ -682,
-	/*log(0.84668)*/ -681,
-	/*log(0.84692)*/ -680,
-	/*log(0.84717)*/ -679,
-	/*log(0.84741)*/ -678,
-	/*log(0.84766)*/ -676,
-	/*log(0.84790)*/ -675,
-	/*log(0.84814)*/ -674,
-	/*log(0.84839)*/ -673,
-	/*log(0.84863)*/ -672,
-	/*log(0.84888)*/ -671,
-	/*log(0.84912)*/ -669,
-	/*log(0.84937)*/ -668,
-	/*log(0.84961)*/ -667,
-	/*log(0.84985)*/ -666,
-	/*log(0.85010)*/ -665,
-	/*log(0.85034)*/ -664,
-	/*log(0.85059)*/ -662,
-	/*log(0.85083)*/ -661,
-	/*log(0.85107)*/ -660,
-	/*log(0.85132)*/ -659,
-	/*log(0.85156)*/ -658,
-	/*log(0.85181)*/ -656,
-	/*log(0.85205)*/ -655,
-	/*log(0.85229)*/ -654,
-	/*log(0.85254)*/ -653,
-	/*log(0.85278)*/ -652,
-	/*log(0.85303)*/ -651,
-	/*log(0.85327)*/ -649,
-	/*log(0.85352)*/ -648,
-	/*log(0.85376)*/ -647,
-	/*log(0.85400)*/ -646,
-	/*log(0.85425)*/ -645,
-	/*log(0.85449)*/ -644,
-	/*log(0.85474)*/ -642,
-	/*log(0.85498)*/ -641,
-	/*log(0.85522)*/ -640,
-	/*log(0.85547)*/ -639,
-	/*log(0.85571)*/ -638,
-	/*log(0.85596)*/ -637,
-	/*log(0.85620)*/ -635,
-	/*log(0.85645)*/ -634,
-	/*log(0.85669)*/ -633,
-	/*log(0.85693)*/ -632,
-	/*log(0.85718)*/ -631,
-	/*log(0.85742)*/ -630,
-	/*log(0.85767)*/ -628,
-	/*log(0.85791)*/ -627,
-	/*log(0.85815)*/ -626,
-	/*log(0.85840)*/ -625,
-	/*log(0.85864)*/ -624,
-	/*log(0.85889)*/ -623,
-	/*log(0.85913)*/ -621,
-	/*log(0.85938)*/ -620,
-	/*log(0.85962)*/ -619,
-	/*log(0.85986)*/ -618,
-	/*log(0.86011)*/ -617,
-	/*log(0.86035)*/ -616,
-	/*log(0.86060)*/ -614,
-	/*log(0.86084)*/ -613,
-	/*log(0.86108)*/ -612,
-	/*log(0.86133)*/ -611,
-	/*log(0.86157)*/ -610,
-	/*log(0.86182)*/ -609,
-	/*log(0.86206)*/ -607,
-	/*log(0.86230)*/ -606,
-	/*log(0.86255)*/ -605,
-	/*log(0.86279)*/ -604,
-	/*log(0.86304)*/ -603,
-	/*log(0.86328)*/ -602,
-	/*log(0.86353)*/ -601,
-	/*log(0.86377)*/ -599,
-	/*log(0.86401)*/ -598,
-	/*log(0.86426)*/ -597,
-	/*log(0.86450)*/ -596,
-	/*log(0.86475)*/ -595,
-	/*log(0.86499)*/ -594,
-	/*log(0.86523)*/ -592,
-	/*log(0.86548)*/ -591,
-	/*log(0.86572)*/ -590,
-	/*log(0.86597)*/ -589,
-	/*log(0.86621)*/ -588,
-	/*log(0.86646)*/ -587,
-	/*log(0.86670)*/ -585,
-	/*log(0.86694)*/ -584,
-	/*log(0.86719)*/ -583,
-	/*log(0.86743)*/ -582,
-	/*log(0.86768)*/ -581,
-	/*log(0.86792)*/ -580,
-	/*log(0.86816)*/ -579,
-	/*log(0.86841)*/ -577,
-	/*log(0.86865)*/ -576,
-	/*log(0.86890)*/ -575,
-	/*log(0.86914)*/ -574,
-	/*log(0.86938)*/ -573,
-	/*log(0.86963)*/ -572,
-	/*log(0.86987)*/ -571,
-	/*log(0.87012)*/ -569,
-	/*log(0.87036)*/ -568,
-	/*log(0.87061)*/ -567,
-	/*log(0.87085)*/ -566,
-	/*log(0.87109)*/ -565,
-	/*log(0.87134)*/ -564,
-	/*log(0.87158)*/ -562,
-	/*log(0.87183)*/ -561,
-	/*log(0.87207)*/ -560,
-	/*log(0.87231)*/ -559,
-	/*log(0.87256)*/ -558,
-	/*log(0.87280)*/ -557,
-	/*log(0.87305)*/ -556,
-	/*log(0.87329)*/ -554,
-	/*log(0.87354)*/ -553,
-	/*log(0.87378)*/ -552,
-	/*log(0.87402)*/ -551,
-	/*log(0.87427)*/ -550,
-	/*log(0.87451)*/ -549,
-	/*log(0.87476)*/ -548,
-	/*log(0.87500)*/ -546,
-	/*log(0.87524)*/ -545,
-	/*log(0.87549)*/ -544,
-	/*log(0.87573)*/ -543,
-	/*log(0.87598)*/ -542,
-	/*log(0.87622)*/ -541,
-	/*log(0.87646)*/ -540,
-	/*log(0.87671)*/ -538,
-	/*log(0.87695)*/ -537,
-	/*log(0.87720)*/ -536,
-	/*log(0.87744)*/ -535,
-	/*log(0.87769)*/ -534,
-	/*log(0.87793)*/ -533,
-	/*log(0.87817)*/ -532,
-	/*log(0.87842)*/ -530,
-	/*log(0.87866)*/ -529,
-	/*log(0.87891)*/ -528,
-	/*log(0.87915)*/ -527,
-	/*log(0.87939)*/ -526,
-	/*log(0.87964)*/ -525,
-	/*log(0.87988)*/ -524,
-	/*log(0.88013)*/ -523,
-	/*log(0.88037)*/ -521,
-	/*log(0.88062)*/ -520,
-	/*log(0.88086)*/ -519,
-	/*log(0.88110)*/ -518,
-	/*log(0.88135)*/ -517,
-	/*log(0.88159)*/ -516,
-	/*log(0.88184)*/ -515,
-	/*log(0.88208)*/ -513,
-	/*log(0.88232)*/ -512,
-	/*log(0.88257)*/ -511,
-	/*log(0.88281)*/ -510,
-	/*log(0.88306)*/ -509,
-	/*log(0.88330)*/ -508,
-	/*log(0.88354)*/ -507,
-	/*log(0.88379)*/ -506,
-	/*log(0.88403)*/ -504,
-	/*log(0.88428)*/ -503,
-	/*log(0.88452)*/ -502,
-	/*log(0.88477)*/ -501,
-	/*log(0.88501)*/ -500,
-	/*log(0.88525)*/ -499,
-	/*log(0.88550)*/ -498,
-	/*log(0.88574)*/ -496,
-	/*log(0.88599)*/ -495,
-	/*log(0.88623)*/ -494,
-	/*log(0.88647)*/ -493,
-	/*log(0.88672)*/ -492,
-	/*log(0.88696)*/ -491,
-	/*log(0.88721)*/ -490,
-	/*log(0.88745)*/ -489,
-	/*log(0.88770)*/ -487,
-	/*log(0.88794)*/ -486,
-	/*log(0.88818)*/ -485,
-	/*log(0.88843)*/ -484,
-	/*log(0.88867)*/ -483,
-	/*log(0.88892)*/ -482,
-	/*log(0.88916)*/ -481,
-	/*log(0.88940)*/ -480,
-	/*log(0.88965)*/ -478,
-	/*log(0.88989)*/ -477,
-	/*log(0.89014)*/ -476,
-	/*log(0.89038)*/ -475,
-	/*log(0.89062)*/ -474,
-	/*log(0.89087)*/ -473,
-	/*log(0.89111)*/ -472,
-	/*log(0.89136)*/ -471,
-	/*log(0.89160)*/ -469,
-	/*log(0.89185)*/ -468,
-	/*log(0.89209)*/ -467,
-	/*log(0.89233)*/ -466,
-	/*log(0.89258)*/ -465,
-	/*log(0.89282)*/ -464,
-	/*log(0.89307)*/ -463,
-	/*log(0.89331)*/ -462,
-	/*log(0.89355)*/ -460,
-	/*log(0.89380)*/ -459,
-	/*log(0.89404)*/ -458,
-	/*log(0.89429)*/ -457,
-	/*log(0.89453)*/ -456,
-	/*log(0.89478)*/ -455,
-	/*log(0.89502)*/ -454,
-	/*log(0.89526)*/ -453,
-	/*log(0.89551)*/ -452,
-	/*log(0.89575)*/ -450,
-	/*log(0.89600)*/ -449,
-	/*log(0.89624)*/ -448,
-	/*log(0.89648)*/ -447,
-	/*log(0.89673)*/ -446,
-	/*log(0.89697)*/ -445,
-	/*log(0.89722)*/ -444,
-	/*log(0.89746)*/ -443,
-	/*log(0.89771)*/ -442,
-	/*log(0.89795)*/ -440,
-	/*log(0.89819)*/ -439,
-	/*log(0.89844)*/ -438,
-	/*log(0.89868)*/ -437,
-	/*log(0.89893)*/ -436,
-	/*log(0.89917)*/ -435,
-	/*log(0.89941)*/ -434,
-	/*log(0.89966)*/ -433,
-	/*log(0.89990)*/ -432,
-	/*log(0.90015)*/ -430,
-	/*log(0.90039)*/ -429,
-	/*log(0.90063)*/ -428,
-	/*log(0.90088)*/ -427,
-	/*log(0.90112)*/ -426,
-	/*log(0.90137)*/ -425,
-	/*log(0.90161)*/ -424,
-	/*log(0.90186)*/ -423,
-	/*log(0.90210)*/ -422,
-	/*log(0.90234)*/ -420,
-	/*log(0.90259)*/ -419,
-	/*log(0.90283)*/ -418,
-	/*log(0.90308)*/ -417,
-	/*log(0.90332)*/ -416,
-	/*log(0.90356)*/ -415,
-	/*log(0.90381)*/ -414,
-	/*log(0.90405)*/ -413,
-	/*log(0.90430)*/ -412,
-	/*log(0.90454)*/ -410,
-	/*log(0.90479)*/ -409,
-	/*log(0.90503)*/ -408,
-	/*log(0.90527)*/ -407,
-	/*log(0.90552)*/ -406,
-	/*log(0.90576)*/ -405,
-	/*log(0.90601)*/ -404,
-	/*log(0.90625)*/ -403,
-	/*log(0.90649)*/ -402,
-	/*log(0.90674)*/ -401,
-	/*log(0.90698)*/ -399,
-	/*log(0.90723)*/ -398,
-	/*log(0.90747)*/ -397,
-	/*log(0.90771)*/ -396,
-	/*log(0.90796)*/ -395,
-	/*log(0.90820)*/ -394,
-	/*log(0.90845)*/ -393,
-	/*log(0.90869)*/ -392,
-	/*log(0.90894)*/ -391,
-	/*log(0.90918)*/ -389,
-	/*log(0.90942)*/ -388,
-	/*log(0.90967)*/ -387,
-	/*log(0.90991)*/ -386,
-	/*log(0.91016)*/ -385,
-	/*log(0.91040)*/ -384,
-	/*log(0.91064)*/ -383,
-	/*log(0.91089)*/ -382,
-	/*log(0.91113)*/ -381,
-	/*log(0.91138)*/ -380,
-	/*log(0.91162)*/ -379,
-	/*log(0.91187)*/ -377,
-	/*log(0.91211)*/ -376,
-	/*log(0.91235)*/ -375,
-	/*log(0.91260)*/ -374,
-	/*log(0.91284)*/ -373,
-	/*log(0.91309)*/ -372,
-	/*log(0.91333)*/ -371,
-	/*log(0.91357)*/ -370,
-	/*log(0.91382)*/ -369,
-	/*log(0.91406)*/ -368,
-	/*log(0.91431)*/ -366,
-	/*log(0.91455)*/ -365,
-	/*log(0.91479)*/ -364,
-	/*log(0.91504)*/ -363,
-	/*log(0.91528)*/ -362,
-	/*log(0.91553)*/ -361,
-	/*log(0.91577)*/ -360,
-	/*log(0.91602)*/ -359,
-	/*log(0.91626)*/ -358,
-	/*log(0.91650)*/ -357,
-	/*log(0.91675)*/ -356,
-	/*log(0.91699)*/ -354,
-	/*log(0.91724)*/ -353,
-	/*log(0.91748)*/ -352,
-	/*log(0.91772)*/ -351,
-	/*log(0.91797)*/ -350,
-	/*log(0.91821)*/ -349,
-	/*log(0.91846)*/ -348,
-	/*log(0.91870)*/ -347,
-	/*log(0.91895)*/ -346,
-	/*log(0.91919)*/ -345,
-	/*log(0.91943)*/ -344,
-	/*log(0.91968)*/ -342,
-	/*log(0.91992)*/ -341,
-	/*log(0.92017)*/ -340,
-	/*log(0.92041)*/ -339,
-	/*log(0.92065)*/ -338,
-	/*log(0.92090)*/ -337,
-	/*log(0.92114)*/ -336,
-	/*log(0.92139)*/ -335,
-	/*log(0.92163)*/ -334,
-	/*log(0.92188)*/ -333,
-	/*log(0.92212)*/ -332,
-	/*log(0.92236)*/ -331,
-	/*log(0.92261)*/ -329,
-	/*log(0.92285)*/ -328,
-	/*log(0.92310)*/ -327,
-	/*log(0.92334)*/ -326,
-	/*log(0.92358)*/ -325,
-	/*log(0.92383)*/ -324,
-	/*log(0.92407)*/ -323,
-	/*log(0.92432)*/ -322,
-	/*log(0.92456)*/ -321,
-	/*log(0.92480)*/ -320,
-	/*log(0.92505)*/ -319,
-	/*log(0.92529)*/ -318,
-	/*log(0.92554)*/ -316,
-	/*log(0.92578)*/ -315,
-	/*log(0.92603)*/ -314,
-	/*log(0.92627)*/ -313,
-	/*log(0.92651)*/ -312,
-	/*log(0.92676)*/ -311,
-	/*log(0.92700)*/ -310,
-	/*log(0.92725)*/ -309,
-	/*log(0.92749)*/ -308,
-	/*log(0.92773)*/ -307,
-	/*log(0.92798)*/ -306,
-	/*log(0.92822)*/ -305,
-	/*log(0.92847)*/ -304,
-	/*log(0.92871)*/ -302,
-	/*log(0.92896)*/ -301,
-	/*log(0.92920)*/ -300,
-	/*log(0.92944)*/ -299,
-	/*log(0.92969)*/ -298,
-	/*log(0.92993)*/ -297,
-	/*log(0.93018)*/ -296,
-	/*log(0.93042)*/ -295,
-	/*log(0.93066)*/ -294,
-	/*log(0.93091)*/ -293,
-	/*log(0.93115)*/ -292,
-	/*log(0.93140)*/ -291,
-	/*log(0.93164)*/ -290,
-	/*log(0.93188)*/ -288,
-	/*log(0.93213)*/ -287,
-	/*log(0.93237)*/ -286,
-	/*log(0.93262)*/ -285,
-	/*log(0.93286)*/ -284,
-	/*log(0.93311)*/ -283,
-	/*log(0.93335)*/ -282,
-	/*log(0.93359)*/ -281,
-	/*log(0.93384)*/ -280,
-	/*log(0.93408)*/ -279,
-	/*log(0.93433)*/ -278,
-	/*log(0.93457)*/ -277,
-	/*log(0.93481)*/ -276,
-	/*log(0.93506)*/ -275,
-	/*log(0.93530)*/ -273,
-	/*log(0.93555)*/ -272,
-	/*log(0.93579)*/ -271,
-	/*log(0.93604)*/ -270,
-	/*log(0.93628)*/ -269,
-	/*log(0.93652)*/ -268,
-	/*log(0.93677)*/ -267,
-	/*log(0.93701)*/ -266,
-	/*log(0.93726)*/ -265,
-	/*log(0.93750)*/ -264,
-	/*log(0.93774)*/ -263,
-	/*log(0.93799)*/ -262,
-	/*log(0.93823)*/ -261,
-	/*log(0.93848)*/ -260,
-	/*log(0.93872)*/ -259,
-	/*log(0.93896)*/ -257,
-	/*log(0.93921)*/ -256,
-	/*log(0.93945)*/ -255,
-	/*log(0.93970)*/ -254,
-	/*log(0.93994)*/ -253,
-	/*log(0.94019)*/ -252,
-	/*log(0.94043)*/ -251,
-	/*log(0.94067)*/ -250,
-	/*log(0.94092)*/ -249,
-	/*log(0.94116)*/ -248,
-	/*log(0.94141)*/ -247,
-	/*log(0.94165)*/ -246,
-	/*log(0.94189)*/ -245,
-	/*log(0.94214)*/ -244,
-	/*log(0.94238)*/ -243,
-	/*log(0.94263)*/ -242,
-	/*log(0.94287)*/ -240,
-	/*log(0.94312)*/ -239,
-	/*log(0.94336)*/ -238,
-	/*log(0.94360)*/ -237,
-	/*log(0.94385)*/ -236,
-	/*log(0.94409)*/ -235,
-	/*log(0.94434)*/ -234,
-	/*log(0.94458)*/ -233,
-	/*log(0.94482)*/ -232,
-	/*log(0.94507)*/ -231,
-	/*log(0.94531)*/ -230,
-	/*log(0.94556)*/ -229,
-	/*log(0.94580)*/ -228,
-	/*log(0.94604)*/ -227,
-	/*log(0.94629)*/ -226,
-	/*log(0.94653)*/ -225,
-	/*log(0.94678)*/ -224,
-	/*log(0.94702)*/ -222,
-	/*log(0.94727)*/ -221,
-	/*log(0.94751)*/ -220,
-	/*log(0.94775)*/ -219,
-	/*log(0.94800)*/ -218,
-	/*log(0.94824)*/ -217,
-	/*log(0.94849)*/ -216,
-	/*log(0.94873)*/ -215,
-	/*log(0.94897)*/ -214,
-	/*log(0.94922)*/ -213,
-	/*log(0.94946)*/ -212,
-	/*log(0.94971)*/ -211,
-	/*log(0.94995)*/ -210,
-	/*log(0.95020)*/ -209,
-	/*log(0.95044)*/ -208,
-	/*log(0.95068)*/ -207,
-	/*log(0.95093)*/ -206,
-	/*log(0.95117)*/ -205,
-	/*log(0.95142)*/ -203,
-	/*log(0.95166)*/ -202,
-	/*log(0.95190)*/ -201,
-	/*log(0.95215)*/ -200,
-	/*log(0.95239)*/ -199,
-	/*log(0.95264)*/ -198,
-	/*log(0.95288)*/ -197,
-	/*log(0.95312)*/ -196,
-	/*log(0.95337)*/ -195,
-	/*log(0.95361)*/ -194,
-	/*log(0.95386)*/ -193,
-	/*log(0.95410)*/ -192,
-	/*log(0.95435)*/ -191,
-	/*log(0.95459)*/ -190,
-	/*log(0.95483)*/ -189,
-	/*log(0.95508)*/ -188,
-	/*log(0.95532)*/ -187,
-	/*log(0.95557)*/ -186,
-	/*log(0.95581)*/ -185,
-	/*log(0.95605)*/ -184,
-	/*log(0.95630)*/ -183,
-	/*log(0.95654)*/ -181,
-	/*log(0.95679)*/ -180,
-	/*log(0.95703)*/ -179,
-	/*log(0.95728)*/ -178,
-	/*log(0.95752)*/ -177,
-	/*log(0.95776)*/ -176,
-	/*log(0.95801)*/ -175,
-	/*log(0.95825)*/ -174,
-	/*log(0.95850)*/ -173,
-	/*log(0.95874)*/ -172,
-	/*log(0.95898)*/ -171,
-	/*log(0.95923)*/ -170,
-	/*log(0.95947)*/ -169,
-	/*log(0.95972)*/ -168,
-	/*log(0.95996)*/ -167,
-	/*log(0.96021)*/ -166,
-	/*log(0.96045)*/ -165,
-	/*log(0.96069)*/ -164,
-	/*log(0.96094)*/ -163,
-	/*log(0.96118)*/ -162,
-	/*log(0.96143)*/ -161,
-	/*log(0.96167)*/ -160,
-	/*log(0.96191)*/ -159,
-	/*log(0.96216)*/ -158,
-	/*log(0.96240)*/ -156,
-	/*log(0.96265)*/ -155,
-	/*log(0.96289)*/ -154,
-	/*log(0.96313)*/ -153,
-	/*log(0.96338)*/ -152,
-	/*log(0.96362)*/ -151,
-	/*log(0.96387)*/ -150,
-	/*log(0.96411)*/ -149,
-	/*log(0.96436)*/ -148,
-	/*log(0.96460)*/ -147,
-	/*log(0.96484)*/ -146,
-	/*log(0.96509)*/ -145,
-	/*log(0.96533)*/ -144,
-	/*log(0.96558)*/ -143,
-	/*log(0.96582)*/ -142,
-	/*log(0.96606)*/ -141,
-	/*log(0.96631)*/ -140,
-	/*log(0.96655)*/ -139,
-	/*log(0.96680)*/ -138,
-	/*log(0.96704)*/ -137,
-	/*log(0.96729)*/ -136,
-	/*log(0.96753)*/ -135,
-	/*log(0.96777)*/ -134,
-	/*log(0.96802)*/ -133,
-	/*log(0.96826)*/ -132,
-	/*log(0.96851)*/ -131,
-	/*log(0.96875)*/ -130,
-	/*log(0.96899)*/ -129,
-	/*log(0.96924)*/ -127,
-	/*log(0.96948)*/ -126,
-	/*log(0.96973)*/ -125,
-	/*log(0.96997)*/ -124,
-	/*log(0.97021)*/ -123,
-	/*log(0.97046)*/ -122,
-	/*log(0.97070)*/ -121,
-	/*log(0.97095)*/ -120,
-	/*log(0.97119)*/ -119,
-	/*log(0.97144)*/ -118,
-	/*log(0.97168)*/ -117,
-	/*log(0.97192)*/ -116,
-	/*log(0.97217)*/ -115,
-	/*log(0.97241)*/ -114,
-	/*log(0.97266)*/ -113,
-	/*log(0.97290)*/ -112,
-	/*log(0.97314)*/ -111,
-	/*log(0.97339)*/ -110,
-	/*log(0.97363)*/ -109,
-	/*log(0.97388)*/ -108,
-	/*log(0.97412)*/ -107,
-	/*log(0.97437)*/ -106,
-	/*log(0.97461)*/ -105,
-	/*log(0.97485)*/ -104,
-	/*log(0.97510)*/ -103,
-	/*log(0.97534)*/ -102,
-	/*log(0.97559)*/ -101,
-	/*log(0.97583)*/ -100,
-	/*log(0.97607)*/ -99,
-	/*log(0.97632)*/ -98,
-	/*log(0.97656)*/ -97,
-	/*log(0.97681)*/ -96,
-	/*log(0.97705)*/ -95,
-	/*log(0.97729)*/ -94,
-	/*log(0.97754)*/ -93,
-	/*log(0.97778)*/ -92,
-	/*log(0.97803)*/ -91,
-	/*log(0.97827)*/ -89,
-	/*log(0.97852)*/ -88,
-	/*log(0.97876)*/ -87,
-	/*log(0.97900)*/ -86,
-	/*log(0.97925)*/ -85,
-	/*log(0.97949)*/ -84,
-	/*log(0.97974)*/ -83,
-	/*log(0.97998)*/ -82,
-	/*log(0.98022)*/ -81,
-	/*log(0.98047)*/ -80,
-	/*log(0.98071)*/ -79,
-	/*log(0.98096)*/ -78,
-	/*log(0.98120)*/ -77,
-	/*log(0.98145)*/ -76,
-	/*log(0.98169)*/ -75,
-	/*log(0.98193)*/ -74,
-	/*log(0.98218)*/ -73,
-	/*log(0.98242)*/ -72,
-	/*log(0.98267)*/ -71,
-	/*log(0.98291)*/ -70,
-	/*log(0.98315)*/ -69,
-	/*log(0.98340)*/ -68,
-	/*log(0.98364)*/ -67,
-	/*log(0.98389)*/ -66,
-	/*log(0.98413)*/ -65,
-	/*log(0.98438)*/ -64,
-	/*log(0.98462)*/ -63,
-	/*log(0.98486)*/ -62,
-	/*log(0.98511)*/ -61,
-	/*log(0.98535)*/ -60,
-	/*log(0.98560)*/ -59,
-	/*log(0.98584)*/ -58,
-	/*log(0.98608)*/ -57,
-	/*log(0.98633)*/ -56,
-	/*log(0.98657)*/ -55,
-	/*log(0.98682)*/ -54,
-	/*log(0.98706)*/ -53,
-	/*log(0.98730)*/ -52,
-	/*log(0.98755)*/ -51,
-	/*log(0.98779)*/ -50,
-	/*log(0.98804)*/ -49,
-	/*log(0.98828)*/ -48,
-	/*log(0.98853)*/ -47,
-	/*log(0.98877)*/ -46,
-	/*log(0.98901)*/ -45,
-	/*log(0.98926)*/ -44,
-	/*log(0.98950)*/ -43,
-	/*log(0.98975)*/ -42,
-	/*log(0.98999)*/ -41,
-	/*log(0.99023)*/ -40,
-	/*log(0.99048)*/ -39,
-	/*log(0.99072)*/ -38,
-	/*log(0.99097)*/ -37,
-	/*log(0.99121)*/ -36,
-	/*log(0.99146)*/ -35,
-	/*log(0.99170)*/ -34,
-	/*log(0.99194)*/ -33,
-	/*log(0.99219)*/ -32,
-	/*log(0.99243)*/ -31,
-	/*log(0.99268)*/ -30,
-	/*log(0.99292)*/ -29,
-	/*log(0.99316)*/ -28,
-	/*log(0.99341)*/ -27,
-	/*log(0.99365)*/ -26,
-	/*log(0.99390)*/ -25,
-	/*log(0.99414)*/ -24,
-	/*log(0.99438)*/ -23,
-	/*log(0.99463)*/ -22,
-	/*log(0.99487)*/ -21,
-	/*log(0.99512)*/ -20,
-	/*log(0.99536)*/ -19,
-	/*log(0.99561)*/ -18,
-	/*log(0.99585)*/ -17,
-	/*log(0.99609)*/ -16,
-	/*log(0.99634)*/ -15,
-	/*log(0.99658)*/ -14,
-	/*log(0.99683)*/ -13,
-	/*log(0.99707)*/ -12,
-	/*log(0.99731)*/ -11,
-	/*log(0.99756)*/ -10,
-	/*log(0.99780)*/ -9,
-	/*log(0.99805)*/ -8,
-	/*log(0.99829)*/ -7,
-	/*log(0.99854)*/ -6,
-	/*log(0.99878)*/ -5,
-	/*log(0.99902)*/ -4,
-	/*log(0.99927)*/ -3,
-	/*log(0.99951)*/ -2,
-	/*log(0.99976)*/ -1,
 	/*log(1.00000)*/ 0,
+	/*log(1.00024)*/ 0,
+	/*log(1.00049)*/ 1,
+	/*log(1.00073)*/ 2,
+	/*log(1.00098)*/ 3,
+	/*log(1.00122)*/ 4,
+	/*log(1.00146)*/ 5,
+	/*log(1.00171)*/ 6,
+	/*log(1.00195)*/ 7,
+	/*log(1.00220)*/ 8,
+	/*log(1.00244)*/ 9,
+	/*log(1.00269)*/ 10,
+	/*log(1.00293)*/ 11,
+	/*log(1.00317)*/ 12,
+	/*log(1.00342)*/ 13,
+	/*log(1.00366)*/ 14,
+	/*log(1.00391)*/ 15,
+	/*log(1.00415)*/ 16,
+	/*log(1.00439)*/ 17,
+	/*log(1.00464)*/ 18,
+	/*log(1.00488)*/ 19,
+	/*log(1.00513)*/ 20,
+	/*log(1.00537)*/ 21,
+	/*log(1.00562)*/ 22,
+	/*log(1.00586)*/ 23,
+	/*log(1.00610)*/ 24,
+	/*log(1.00635)*/ 25,
+	/*log(1.00659)*/ 26,
+	/*log(1.00684)*/ 27,
+	/*log(1.00708)*/ 28,
+	/*log(1.00732)*/ 29,
+	/*log(1.00757)*/ 30,
+	/*log(1.00781)*/ 31,
+	/*log(1.00806)*/ 32,
+	/*log(1.00830)*/ 33,
+	/*log(1.00854)*/ 34,
+	/*log(1.00879)*/ 35,
+	/*log(1.00903)*/ 36,
+	/*log(1.00928)*/ 37,
+	/*log(1.00952)*/ 38,
+	/*log(1.00977)*/ 39,
+	/*log(1.01001)*/ 40,
+	/*log(1.01025)*/ 41,
+	/*log(1.01050)*/ 42,
+	/*log(1.01074)*/ 43,
+	/*log(1.01099)*/ 44,
+	/*log(1.01123)*/ 45,
+	/*log(1.01147)*/ 46,
+	/*log(1.01172)*/ 47,
+	/*log(1.01196)*/ 48,
+	/*log(1.01221)*/ 49,
+	/*log(1.01245)*/ 50,
+	/*log(1.01270)*/ 51,
+	/*log(1.01294)*/ 52,
+	/*log(1.01318)*/ 53,
+	/*log(1.01343)*/ 54,
+	/*log(1.01367)*/ 55,
+	/*log(1.01392)*/ 56,
+	/*log(1.01416)*/ 57,
+	/*log(1.01440)*/ 58,
+	/*log(1.01465)*/ 59,
+	/*log(1.01489)*/ 60,
+	/*log(1.01514)*/ 61,
+	/*log(1.01538)*/ 62,
+	/*log(1.01562)*/ 63,
+	/*log(1.01587)*/ 64,
+	/*log(1.01611)*/ 65,
+	/*log(1.01636)*/ 66,
+	/*log(1.01660)*/ 67,
+	/*log(1.01685)*/ 68,
+	/*log(1.01709)*/ 69,
+	/*log(1.01733)*/ 70,
+	/*log(1.01758)*/ 71,
+	/*log(1.01782)*/ 72,
+	/*log(1.01807)*/ 73,
+	/*log(1.01831)*/ 74,
+	/*log(1.01855)*/ 75,
+	/*log(1.01880)*/ 76,
+	/*log(1.01904)*/ 77,
+	/*log(1.01929)*/ 78,
+	/*log(1.01953)*/ 79,
+	/*log(1.01978)*/ 80,
+	/*log(1.02002)*/ 81,
+	/*log(1.02026)*/ 82,
+	/*log(1.02051)*/ 83,
+	/*log(1.02075)*/ 84,
+	/*log(1.02100)*/ 85,
+	/*log(1.02124)*/ 86,
+	/*log(1.02148)*/ 87,
+	/*log(1.02173)*/ 88,
+	/*log(1.02197)*/ 89,
+	/*log(1.02222)*/ 90,
+	/*log(1.02246)*/ 90,
+	/*log(1.02271)*/ 91,
+	/*log(1.02295)*/ 92,
+	/*log(1.02319)*/ 93,
+	/*log(1.02344)*/ 94,
+	/*log(1.02368)*/ 95,
+	/*log(1.02393)*/ 96,
+	/*log(1.02417)*/ 97,
+	/*log(1.02441)*/ 98,
+	/*log(1.02466)*/ 99,
+	/*log(1.02490)*/ 100,
+	/*log(1.02515)*/ 101,
+	/*log(1.02539)*/ 102,
+	/*log(1.02563)*/ 103,
+	/*log(1.02588)*/ 104,
+	/*log(1.02612)*/ 105,
+	/*log(1.02637)*/ 106,
+	/*log(1.02661)*/ 107,
+	/*log(1.02686)*/ 108,
+	/*log(1.02710)*/ 109,
+	/*log(1.02734)*/ 110,
+	/*log(1.02759)*/ 111,
+	/*log(1.02783)*/ 112,
+	/*log(1.02808)*/ 113,
+	/*log(1.02832)*/ 114,
+	/*log(1.02856)*/ 115,
+	/*log(1.02881)*/ 116,
+	/*log(1.02905)*/ 117,
+	/*log(1.02930)*/ 118,
+	/*log(1.02954)*/ 119,
+	/*log(1.02979)*/ 120,
+	/*log(1.03003)*/ 121,
+	/*log(1.03027)*/ 122,
+	/*log(1.03052)*/ 123,
+	/*log(1.03076)*/ 124,
+	/*log(1.03101)*/ 125,
+	/*log(1.03125)*/ 126,
+	/*log(1.03149)*/ 127,
+	/*log(1.03174)*/ 127,
+	/*log(1.03198)*/ 128,
+	/*log(1.03223)*/ 129,
+	/*log(1.03247)*/ 130,
+	/*log(1.03271)*/ 131,
+	/*log(1.03296)*/ 132,
+	/*log(1.03320)*/ 133,
+	/*log(1.03345)*/ 134,
+	/*log(1.03369)*/ 135,
+	/*log(1.03394)*/ 136,
+	/*log(1.03418)*/ 137,
+	/*log(1.03442)*/ 138,
+	/*log(1.03467)*/ 139,
+	/*log(1.03491)*/ 140,
+	/*log(1.03516)*/ 141,
+	/*log(1.03540)*/ 142,
+	/*log(1.03564)*/ 143,
+	/*log(1.03589)*/ 144,
+	/*log(1.03613)*/ 145,
+	/*log(1.03638)*/ 146,
+	/*log(1.03662)*/ 147,
+	/*log(1.03687)*/ 148,
+	/*log(1.03711)*/ 149,
+	/*log(1.03735)*/ 150,
+	/*log(1.03760)*/ 151,
+	/*log(1.03784)*/ 152,
+	/*log(1.03809)*/ 153,
+	/*log(1.03833)*/ 154,
+	/*log(1.03857)*/ 155,
+	/*log(1.03882)*/ 155,
+	/*log(1.03906)*/ 156,
+	/*log(1.03931)*/ 157,
+	/*log(1.03955)*/ 158,
+	/*log(1.03979)*/ 159,
+	/*log(1.04004)*/ 160,
+	/*log(1.04028)*/ 161,
+	/*log(1.04053)*/ 162,
+	/*log(1.04077)*/ 163,
+	/*log(1.04102)*/ 164,
+	/*log(1.04126)*/ 165,
+	/*log(1.04150)*/ 166,
+	/*log(1.04175)*/ 167,
+	/*log(1.04199)*/ 168,
+	/*log(1.04224)*/ 169,
+	/*log(1.04248)*/ 170,
+	/*log(1.04272)*/ 171,
+	/*log(1.04297)*/ 172,
+	/*log(1.04321)*/ 173,
+	/*log(1.04346)*/ 174,
+	/*log(1.04370)*/ 175,
+	/*log(1.04395)*/ 176,
+	/*log(1.04419)*/ 177,
+	/*log(1.04443)*/ 178,
+	/*log(1.04468)*/ 179,
+	/*log(1.04492)*/ 179,
+	/*log(1.04517)*/ 180,
+	/*log(1.04541)*/ 181,
+	/*log(1.04565)*/ 182,
+	/*log(1.04590)*/ 183,
+	/*log(1.04614)*/ 184,
+	/*log(1.04639)*/ 185,
+	/*log(1.04663)*/ 186,
+	/*log(1.04688)*/ 187,
+	/*log(1.04712)*/ 188,
+	/*log(1.04736)*/ 189,
+	/*log(1.04761)*/ 190,
+	/*log(1.04785)*/ 191,
+	/*log(1.04810)*/ 192,
+	/*log(1.04834)*/ 193,
+	/*log(1.04858)*/ 194,
+	/*log(1.04883)*/ 195,
+	/*log(1.04907)*/ 196,
+	/*log(1.04932)*/ 197,
+	/*log(1.04956)*/ 198,
+	/*log(1.04980)*/ 199,
+	/*log(1.05005)*/ 200,
+	/*log(1.05029)*/ 200,
+	/*log(1.05054)*/ 201,
+	/*log(1.05078)*/ 202,
+	/*log(1.05103)*/ 203,
+	/*log(1.05127)*/ 204,
+	/*log(1.05151)*/ 205,
+	/*log(1.05176)*/ 206,
+	/*log(1.05200)*/ 207,
+	/*log(1.05225)*/ 208,
+	/*log(1.05249)*/ 209,
+	/*log(1.05273)*/ 210,
+	/*log(1.05298)*/ 211,
+	/*log(1.05322)*/ 212,
+	/*log(1.05347)*/ 213,
+	/*log(1.05371)*/ 214,
+	/*log(1.05396)*/ 215,
+	/*log(1.05420)*/ 216,
+	/*log(1.05444)*/ 217,
+	/*log(1.05469)*/ 218,
+	/*log(1.05493)*/ 219,
+	/*log(1.05518)*/ 219,
+	/*log(1.05542)*/ 220,
+	/*log(1.05566)*/ 221,
+	/*log(1.05591)*/ 222,
+	/*log(1.05615)*/ 223,
+	/*log(1.05640)*/ 224,
+	/*log(1.05664)*/ 225,
+	/*log(1.05688)*/ 226,
+	/*log(1.05713)*/ 227,
+	/*log(1.05737)*/ 228,
+	/*log(1.05762)*/ 229,
+	/*log(1.05786)*/ 230,
+	/*log(1.05811)*/ 231,
+	/*log(1.05835)*/ 232,
+	/*log(1.05859)*/ 233,
+	/*log(1.05884)*/ 234,
+	/*log(1.05908)*/ 235,
+	/*log(1.05933)*/ 236,
+	/*log(1.05957)*/ 237,
+	/*log(1.05981)*/ 237,
+	/*log(1.06006)*/ 238,
+	/*log(1.06030)*/ 239,
+	/*log(1.06055)*/ 240,
+	/*log(1.06079)*/ 241,
+	/*log(1.06104)*/ 242,
+	/*log(1.06128)*/ 243,
+	/*log(1.06152)*/ 244,
+	/*log(1.06177)*/ 245,
+	/*log(1.06201)*/ 246,
+	/*log(1.06226)*/ 247,
+	/*log(1.06250)*/ 248,
+	/*log(1.06274)*/ 249,
+	/*log(1.06299)*/ 250,
+	/*log(1.06323)*/ 251,
+	/*log(1.06348)*/ 252,
+	/*log(1.06372)*/ 253,
+	/*log(1.06396)*/ 253,
+	/*log(1.06421)*/ 254,
+	/*log(1.06445)*/ 255,
+	/*log(1.06470)*/ 256,
+	/*log(1.06494)*/ 257,
+	/*log(1.06519)*/ 258,
+	/*log(1.06543)*/ 259,
+	/*log(1.06567)*/ 260,
+	/*log(1.06592)*/ 261,
+	/*log(1.06616)*/ 262,
+	/*log(1.06641)*/ 263,
+	/*log(1.06665)*/ 264,
+	/*log(1.06689)*/ 265,
+	/*log(1.06714)*/ 266,
+	/*log(1.06738)*/ 267,
+	/*log(1.06763)*/ 268,
+	/*log(1.06787)*/ 268,
+	/*log(1.06812)*/ 269,
+	/*log(1.06836)*/ 270,
+	/*log(1.06860)*/ 271,
+	/*log(1.06885)*/ 272,
+	/*log(1.06909)*/ 273,
+	/*log(1.06934)*/ 274,
+	/*log(1.06958)*/ 275,
+	/*log(1.06982)*/ 276,
+	/*log(1.07007)*/ 277,
+	/*log(1.07031)*/ 278,
+	/*log(1.07056)*/ 279,
+	/*log(1.07080)*/ 280,
+	/*log(1.07104)*/ 281,
+	/*log(1.07129)*/ 282,
+	/*log(1.07153)*/ 282,
+	/*log(1.07178)*/ 283,
+	/*log(1.07202)*/ 284,
+	/*log(1.07227)*/ 285,
+	/*log(1.07251)*/ 286,
+	/*log(1.07275)*/ 287,
+	/*log(1.07300)*/ 288,
+	/*log(1.07324)*/ 289,
+	/*log(1.07349)*/ 290,
+	/*log(1.07373)*/ 291,
+	/*log(1.07397)*/ 292,
+	/*log(1.07422)*/ 293,
+	/*log(1.07446)*/ 294,
+	/*log(1.07471)*/ 295,
+	/*log(1.07495)*/ 296,
+	/*log(1.07520)*/ 296,
+	/*log(1.07544)*/ 297,
+	/*log(1.07568)*/ 298,
+	/*log(1.07593)*/ 299,
+	/*log(1.07617)*/ 300,
+	/*log(1.07642)*/ 301,
+	/*log(1.07666)*/ 302,
+	/*log(1.07690)*/ 303,
+	/*log(1.07715)*/ 304,
+	/*log(1.07739)*/ 305,
+	/*log(1.07764)*/ 306,
+	/*log(1.07788)*/ 307,
+	/*log(1.07812)*/ 308,
+	/*log(1.07837)*/ 309,
+	/*log(1.07861)*/ 309,
+	/*log(1.07886)*/ 310,
+	/*log(1.07910)*/ 311,
+	/*log(1.07935)*/ 312,
+	/*log(1.07959)*/ 313,
+	/*log(1.07983)*/ 314,
+	/*log(1.08008)*/ 315,
+	/*log(1.08032)*/ 316,
+	/*log(1.08057)*/ 317,
+	/*log(1.08081)*/ 318,
+	/*log(1.08105)*/ 319,
+	/*log(1.08130)*/ 320,
+	/*log(1.08154)*/ 321,
+	/*log(1.08179)*/ 322,
+	/*log(1.08203)*/ 322,
+	/*log(1.08228)*/ 323,
+	/*log(1.08252)*/ 324,
+	/*log(1.08276)*/ 325,
+	/*log(1.08301)*/ 326,
+	/*log(1.08325)*/ 327,
+	/*log(1.08350)*/ 328,
+	/*log(1.08374)*/ 329,
+	/*log(1.08398)*/ 330,
+	/*log(1.08423)*/ 331,
+	/*log(1.08447)*/ 332,
+	/*log(1.08472)*/ 333,
+	/*log(1.08496)*/ 334,
+	/*log(1.08521)*/ 334,
+	/*log(1.08545)*/ 335,
+	/*log(1.08569)*/ 336,
+	/*log(1.08594)*/ 337,
+	/*log(1.08618)*/ 338,
+	/*log(1.08643)*/ 339,
+	/*log(1.08667)*/ 340,
+	/*log(1.08691)*/ 341,
+	/*log(1.08716)*/ 342,
+	/*log(1.08740)*/ 343,
+	/*log(1.08765)*/ 344,
+	/*log(1.08789)*/ 345,
+	/*log(1.08813)*/ 345,
+	/*log(1.08838)*/ 346,
+	/*log(1.08862)*/ 347,
+	/*log(1.08887)*/ 348,
+	/*log(1.08911)*/ 349,
+	/*log(1.08936)*/ 350,
+	/*log(1.08960)*/ 351,
+	/*log(1.08984)*/ 352,
+	/*log(1.09009)*/ 353,
+	/*log(1.09033)*/ 354,
+	/*log(1.09058)*/ 355,
+	/*log(1.09082)*/ 356,
+	/*log(1.09106)*/ 356,
+	/*log(1.09131)*/ 357,
+	/*log(1.09155)*/ 358,
+	/*log(1.09180)*/ 359,
+	/*log(1.09204)*/ 360,
+	/*log(1.09229)*/ 361,
+	/*log(1.09253)*/ 362,
+	/*log(1.09277)*/ 363,
+	/*log(1.09302)*/ 364,
+	/*log(1.09326)*/ 365,
+	/*log(1.09351)*/ 366,
+	/*log(1.09375)*/ 367,
+	/*log(1.09399)*/ 367,
+	/*log(1.09424)*/ 368,
+	/*log(1.09448)*/ 369,
+	/*log(1.09473)*/ 370,
+	/*log(1.09497)*/ 371,
+	/*log(1.09521)*/ 372,
+	/*log(1.09546)*/ 373,
+	/*log(1.09570)*/ 374,
+	/*log(1.09595)*/ 375,
+	/*log(1.09619)*/ 376,
+	/*log(1.09644)*/ 377,
+	/*log(1.09668)*/ 378,
+	/*log(1.09692)*/ 378,
+	/*log(1.09717)*/ 379,
+	/*log(1.09741)*/ 380,
+	/*log(1.09766)*/ 381,
+	/*log(1.09790)*/ 382,
+	/*log(1.09814)*/ 383,
+	/*log(1.09839)*/ 384,
+	/*log(1.09863)*/ 385,
+	/*log(1.09888)*/ 386,
+	/*log(1.09912)*/ 387,
+	/*log(1.09937)*/ 388,
+	/*log(1.09961)*/ 388,
+	/*log(1.09985)*/ 389,
+	/*log(1.10010)*/ 390,
+	/*log(1.10034)*/ 391,
+	/*log(1.10059)*/ 392,
+	/*log(1.10083)*/ 393,
+	/*log(1.10107)*/ 394,
+	/*log(1.10132)*/ 395,
+	/*log(1.10156)*/ 396,
+	/*log(1.10181)*/ 397,
+	/*log(1.10205)*/ 398,
+	/*log(1.10229)*/ 398,
+	/*log(1.10254)*/ 399,
+	/*log(1.10278)*/ 400,
+	/*log(1.10303)*/ 401,
+	/*log(1.10327)*/ 402,
+	/*log(1.10352)*/ 403,
+	/*log(1.10376)*/ 404,
+	/*log(1.10400)*/ 405,
+	/*log(1.10425)*/ 406,
+	/*log(1.10449)*/ 407,
+	/*log(1.10474)*/ 407,
+	/*log(1.10498)*/ 408,
+	/*log(1.10522)*/ 409,
+	/*log(1.10547)*/ 410,
+	/*log(1.10571)*/ 411,
+	/*log(1.10596)*/ 412,
+	/*log(1.10620)*/ 413,
+	/*log(1.10645)*/ 414,
+	/*log(1.10669)*/ 415,
+	/*log(1.10693)*/ 416,
+	/*log(1.10718)*/ 417,
+	/*log(1.10742)*/ 417,
+	/*log(1.10767)*/ 418,
+	/*log(1.10791)*/ 419,
+	/*log(1.10815)*/ 420,
+	/*log(1.10840)*/ 421,
+	/*log(1.10864)*/ 422,
+	/*log(1.10889)*/ 423,
+	/*log(1.10913)*/ 424,
+	/*log(1.10938)*/ 425,
+	/*log(1.10962)*/ 426,
+	/*log(1.10986)*/ 426,
+	/*log(1.11011)*/ 427,
+	/*log(1.11035)*/ 428,
+	/*log(1.11060)*/ 429,
+	/*log(1.11084)*/ 430,
+	/*log(1.11108)*/ 431,
+	/*log(1.11133)*/ 432,
+	/*log(1.11157)*/ 433,
+	/*log(1.11182)*/ 434,
+	/*log(1.11206)*/ 435,
+	/*log(1.11230)*/ 435,
+	/*log(1.11255)*/ 436,
+	/*log(1.11279)*/ 437,
+	/*log(1.11304)*/ 438,
+	/*log(1.11328)*/ 439,
+	/*log(1.11353)*/ 440,
+	/*log(1.11377)*/ 441,
+	/*log(1.11401)*/ 442,
+	/*log(1.11426)*/ 443,
+	/*log(1.11450)*/ 444,
+	/*log(1.11475)*/ 444,
+	/*log(1.11499)*/ 445,
+	/*log(1.11523)*/ 446,
+	/*log(1.11548)*/ 447,
+	/*log(1.11572)*/ 448,
+	/*log(1.11597)*/ 449,
+	/*log(1.11621)*/ 450,
+	/*log(1.11646)*/ 451,
+	/*log(1.11670)*/ 452,
+	/*log(1.11694)*/ 453,
+	/*log(1.11719)*/ 453,
+	/*log(1.11743)*/ 454,
+	/*log(1.11768)*/ 455,
+	/*log(1.11792)*/ 456,
+	/*log(1.11816)*/ 457,
+	/*log(1.11841)*/ 458,
+	/*log(1.11865)*/ 459,
+	/*log(1.11890)*/ 460,
+	/*log(1.11914)*/ 461,
+	/*log(1.11938)*/ 461,
+	/*log(1.11963)*/ 462,
+	/*log(1.11987)*/ 463,
+	/*log(1.12012)*/ 464,
+	/*log(1.12036)*/ 465,
+	/*log(1.12061)*/ 466,
+	/*log(1.12085)*/ 467,
+	/*log(1.12109)*/ 468,
+	/*log(1.12134)*/ 469,
+	/*log(1.12158)*/ 469,
+	/*log(1.12183)*/ 470,
+	/*log(1.12207)*/ 471,
+	/*log(1.12231)*/ 472,
+	/*log(1.12256)*/ 473,
+	/*log(1.12280)*/ 474,
+	/*log(1.12305)*/ 475,
+	/*log(1.12329)*/ 476,
+	/*log(1.12354)*/ 477,
+	/*log(1.12378)*/ 477,
+	/*log(1.12402)*/ 478,
+	/*log(1.12427)*/ 479,
+	/*log(1.12451)*/ 480,
+	/*log(1.12476)*/ 481,
+	/*log(1.12500)*/ 482,
+	/*log(1.12524)*/ 483,
+	/*log(1.12549)*/ 484,
+	/*log(1.12573)*/ 485,
+	/*log(1.12598)*/ 485,
+	/*log(1.12622)*/ 486,
+	/*log(1.12646)*/ 487,
+	/*log(1.12671)*/ 488,
+	/*log(1.12695)*/ 489,
+	/*log(1.12720)*/ 490,
+	/*log(1.12744)*/ 491,
+	/*log(1.12769)*/ 492,
+	/*log(1.12793)*/ 493,
+	/*log(1.12817)*/ 493,
+	/*log(1.12842)*/ 494,
+	/*log(1.12866)*/ 495,
+	/*log(1.12891)*/ 496,
+	/*log(1.12915)*/ 497,
+	/*log(1.12939)*/ 498,
+	/*log(1.12964)*/ 499,
+	/*log(1.12988)*/ 500,
+	/*log(1.13013)*/ 501,
+	/*log(1.13037)*/ 501,
+	/*log(1.13062)*/ 502,
+	/*log(1.13086)*/ 503,
+	/*log(1.13110)*/ 504,
+	/*log(1.13135)*/ 505,
+	/*log(1.13159)*/ 506,
+	/*log(1.13184)*/ 507,
+	/*log(1.13208)*/ 508,
+	/*log(1.13232)*/ 509,
+	/*log(1.13257)*/ 509,
+	/*log(1.13281)*/ 510,
+	/*log(1.13306)*/ 511,
+	/*log(1.13330)*/ 512,
+	/*log(1.13354)*/ 513,
+	/*log(1.13379)*/ 514,
+	/*log(1.13403)*/ 515,
+	/*log(1.13428)*/ 516,
+	/*log(1.13452)*/ 516,
+	/*log(1.13477)*/ 517,
+	/*log(1.13501)*/ 518,
+	/*log(1.13525)*/ 519,
+	/*log(1.13550)*/ 520,
+	/*log(1.13574)*/ 521,
+	/*log(1.13599)*/ 522,
+	/*log(1.13623)*/ 523,
+	/*log(1.13647)*/ 524,
+	/*log(1.13672)*/ 524,
+	/*log(1.13696)*/ 525,
+	/*log(1.13721)*/ 526,
+	/*log(1.13745)*/ 527,
+	/*log(1.13770)*/ 528,
+	/*log(1.13794)*/ 529,
+	/*log(1.13818)*/ 530,
+	/*log(1.13843)*/ 531,
+	/*log(1.13867)*/ 531,
+	/*log(1.13892)*/ 532,
+	/*log(1.13916)*/ 533,
+	/*log(1.13940)*/ 534,
+	/*log(1.13965)*/ 535,
+	/*log(1.13989)*/ 536,
+	/*log(1.14014)*/ 537,
+	/*log(1.14038)*/ 538,
+	/*log(1.14062)*/ 538,
+	/*log(1.14087)*/ 539,
+	/*log(1.14111)*/ 540,
+	/*log(1.14136)*/ 541,
+	/*log(1.14160)*/ 542,
+	/*log(1.14185)*/ 543,
+	/*log(1.14209)*/ 544,
+	/*log(1.14233)*/ 545,
+	/*log(1.14258)*/ 545,
+	/*log(1.14282)*/ 546,
+	/*log(1.14307)*/ 547,
+	/*log(1.14331)*/ 548,
+	/*log(1.14355)*/ 549,
+	/*log(1.14380)*/ 550,
+	/*log(1.14404)*/ 551,
+	/*log(1.14429)*/ 552,
+	/*log(1.14453)*/ 552,
+	/*log(1.14478)*/ 553,
+	/*log(1.14502)*/ 554,
+	/*log(1.14526)*/ 555,
+	/*log(1.14551)*/ 556,
+	/*log(1.14575)*/ 557,
+	/*log(1.14600)*/ 558,
+	/*log(1.14624)*/ 559,
+	/*log(1.14648)*/ 559,
+	/*log(1.14673)*/ 560,
+	/*log(1.14697)*/ 561,
+	/*log(1.14722)*/ 562,
+	/*log(1.14746)*/ 563,
+	/*log(1.14771)*/ 564,
+	/*log(1.14795)*/ 565,
+	/*log(1.14819)*/ 566,
+	/*log(1.14844)*/ 566,
+	/*log(1.14868)*/ 567,
+	/*log(1.14893)*/ 568,
+	/*log(1.14917)*/ 569,
+	/*log(1.14941)*/ 570,
+	/*log(1.14966)*/ 571,
+	/*log(1.14990)*/ 572,
+	/*log(1.15015)*/ 572,
+	/*log(1.15039)*/ 573,
+	/*log(1.15063)*/ 574,
+	/*log(1.15088)*/ 575,
+	/*log(1.15112)*/ 576,
+	/*log(1.15137)*/ 577,
+	/*log(1.15161)*/ 578,
+	/*log(1.15186)*/ 579,
+	/*log(1.15210)*/ 579,
+	/*log(1.15234)*/ 580,
+	/*log(1.15259)*/ 581,
+	/*log(1.15283)*/ 582,
+	/*log(1.15308)*/ 583,
+	/*log(1.15332)*/ 584,
+	/*log(1.15356)*/ 585,
+	/*log(1.15381)*/ 586,
+	/*log(1.15405)*/ 586,
+	/*log(1.15430)*/ 587,
+	/*log(1.15454)*/ 588,
+	/*log(1.15479)*/ 589,
+	/*log(1.15503)*/ 590,
+	/*log(1.15527)*/ 591,
+	/*log(1.15552)*/ 592,
+	/*log(1.15576)*/ 592,
+	/*log(1.15601)*/ 593,
+	/*log(1.15625)*/ 594,
+	/*log(1.15649)*/ 595,
+	/*log(1.15674)*/ 596,
+	/*log(1.15698)*/ 597,
+	/*log(1.15723)*/ 598,
+	/*log(1.15747)*/ 598,
+	/*log(1.15771)*/ 599,
+	/*log(1.15796)*/ 600,
+	/*log(1.15820)*/ 601,
+	/*log(1.15845)*/ 602,
+	/*log(1.15869)*/ 603,
+	/*log(1.15894)*/ 604,
+	/*log(1.15918)*/ 605,
+	/*log(1.15942)*/ 605,
+	/*log(1.15967)*/ 606,
+	/*log(1.15991)*/ 607,
+	/*log(1.16016)*/ 608,
+	/*log(1.16040)*/ 609,
+	/*log(1.16064)*/ 610,
+	/*log(1.16089)*/ 611,
+	/*log(1.16113)*/ 611,
+	/*log(1.16138)*/ 612,
+	/*log(1.16162)*/ 613,
+	/*log(1.16187)*/ 614,
+	/*log(1.16211)*/ 615,
+	/*log(1.16235)*/ 616,
+	/*log(1.16260)*/ 617,
+	/*log(1.16284)*/ 617,
+	/*log(1.16309)*/ 618,
+	/*log(1.16333)*/ 619,
+	/*log(1.16357)*/ 620,
+	/*log(1.16382)*/ 621,
+	/*log(1.16406)*/ 622,
+	/*log(1.16431)*/ 623,
+	/*log(1.16455)*/ 623,
+	/*log(1.16479)*/ 624,
+	/*log(1.16504)*/ 625,
+	/*log(1.16528)*/ 626,
+	/*log(1.16553)*/ 627,
+	/*log(1.16577)*/ 628,
+	/*log(1.16602)*/ 629,
+	/*log(1.16626)*/ 629,
+	/*log(1.16650)*/ 630,
+	/*log(1.16675)*/ 631,
+	/*log(1.16699)*/ 632,
+	/*log(1.16724)*/ 633,
+	/*log(1.16748)*/ 634,
+	/*log(1.16772)*/ 635,
+	/*log(1.16797)*/ 635,
+	/*log(1.16821)*/ 636,
+	/*log(1.16846)*/ 637,
+	/*log(1.16870)*/ 638,
+	/*log(1.16895)*/ 639,
+	/*log(1.16919)*/ 640,
+	/*log(1.16943)*/ 641,
+	/*log(1.16968)*/ 641,
+	/*log(1.16992)*/ 642,
+	/*log(1.17017)*/ 643,
+	/*log(1.17041)*/ 644,
+	/*log(1.17065)*/ 645,
+	/*log(1.17090)*/ 646,
+	/*log(1.17114)*/ 647,
+	/*log(1.17139)*/ 647,
+	/*log(1.17163)*/ 648,
+	/*log(1.17188)*/ 649,
+	/*log(1.17212)*/ 650,
+	/*log(1.17236)*/ 651,
+	/*log(1.17261)*/ 652,
+	/*log(1.17285)*/ 653,
+	/*log(1.17310)*/ 653,
+	/*log(1.17334)*/ 654,
+	/*log(1.17358)*/ 655,
+	/*log(1.17383)*/ 656,
+	/*log(1.17407)*/ 657,
+	/*log(1.17432)*/ 658,
+	/*log(1.17456)*/ 659,
+	/*log(1.17480)*/ 659,
+	/*log(1.17505)*/ 660,
+	/*log(1.17529)*/ 661,
+	/*log(1.17554)*/ 662,
+	/*log(1.17578)*/ 663,
+	/*log(1.17603)*/ 664,
+	/*log(1.17627)*/ 664,
+	/*log(1.17651)*/ 665,
+	/*log(1.17676)*/ 666,
+	/*log(1.17700)*/ 667,
+	/*log(1.17725)*/ 668,
+	/*log(1.17749)*/ 669,
+	/*log(1.17773)*/ 670,
+	/*log(1.17798)*/ 670,
+	/*log(1.17822)*/ 671,
+	/*log(1.17847)*/ 672,
+	/*log(1.17871)*/ 673,
+	/*log(1.17896)*/ 674,
+	/*log(1.17920)*/ 675,
+	/*log(1.17944)*/ 676,
+	/*log(1.17969)*/ 676,
+	/*log(1.17993)*/ 677,
+	/*log(1.18018)*/ 678,
+	/*log(1.18042)*/ 679,
+	/*log(1.18066)*/ 680,
+	/*log(1.18091)*/ 681,
+	/*log(1.18115)*/ 681,
+	/*log(1.18140)*/ 682,
+	/*log(1.18164)*/ 683,
+	/*log(1.18188)*/ 684,
+	/*log(1.18213)*/ 685,
+	/*log(1.18237)*/ 686,
+	/*log(1.18262)*/ 687,
+	/*log(1.18286)*/ 687,
+	/*log(1.18311)*/ 688,
+	/*log(1.18335)*/ 689,
+	/*log(1.18359)*/ 690,
+	/*log(1.18384)*/ 691,
+	/*log(1.18408)*/ 692,
+	/*log(1.18433)*/ 692,
+	/*log(1.18457)*/ 693,
+	/*log(1.18481)*/ 694,
+	/*log(1.18506)*/ 695,
+	/*log(1.18530)*/ 696,
+	/*log(1.18555)*/ 697,
+	/*log(1.18579)*/ 697,
+	/*log(1.18604)*/ 698,
+	/*log(1.18628)*/ 699,
+	/*log(1.18652)*/ 700,
+	/*log(1.18677)*/ 701,
+	/*log(1.18701)*/ 702,
+	/*log(1.18726)*/ 703,
+	/*log(1.18750)*/ 703,
+	/*log(1.18774)*/ 704,
+	/*log(1.18799)*/ 705,
+	/*log(1.18823)*/ 706,
+	/*log(1.18848)*/ 707,
+	/*log(1.18872)*/ 708,
+	/*log(1.18896)*/ 708,
+	/*log(1.18921)*/ 709,
+	/*log(1.18945)*/ 710,
+	/*log(1.18970)*/ 711,
+	/*log(1.18994)*/ 712,
+	/*log(1.19019)*/ 713,
+	/*log(1.19043)*/ 713,
+	/*log(1.19067)*/ 714,
+	/*log(1.19092)*/ 715,
+	/*log(1.19116)*/ 716,
+	/*log(1.19141)*/ 717,
+	/*log(1.19165)*/ 718,
+	/*log(1.19189)*/ 719,
+	/*log(1.19214)*/ 719,
+	/*log(1.19238)*/ 720,
+	/*log(1.19263)*/ 721,
+	/*log(1.19287)*/ 722,
+	/*log(1.19312)*/ 723,
+	/*log(1.19336)*/ 724,
+	/*log(1.19360)*/ 724,
+	/*log(1.19385)*/ 725,
+	/*log(1.19409)*/ 726,
+	/*log(1.19434)*/ 727,
+	/*log(1.19458)*/ 728,
+	/*log(1.19482)*/ 729,
+	/*log(1.19507)*/ 729,
+	/*log(1.19531)*/ 730,
+	/*log(1.19556)*/ 731,
+	/*log(1.19580)*/ 732,
+	/*log(1.19604)*/ 733,
+	/*log(1.19629)*/ 734,
+	/*log(1.19653)*/ 734,
+	/*log(1.19678)*/ 735,
+	/*log(1.19702)*/ 736,
+	/*log(1.19727)*/ 737,
+	/*log(1.19751)*/ 738,
+	/*log(1.19775)*/ 739,
+	/*log(1.19800)*/ 739,
+	/*log(1.19824)*/ 740,
+	/*log(1.19849)*/ 741,
+	/*log(1.19873)*/ 742,
+	/*log(1.19897)*/ 743,
+	/*log(1.19922)*/ 744,
+	/*log(1.19946)*/ 744,
+	/*log(1.19971)*/ 745,
+	/*log(1.19995)*/ 746,
+	/*log(1.20020)*/ 747,
+	/*log(1.20044)*/ 748,
+	/*log(1.20068)*/ 749,
+	/*log(1.20093)*/ 749,
+	/*log(1.20117)*/ 750,
+	/*log(1.20142)*/ 751,
+	/*log(1.20166)*/ 752,
+	/*log(1.20190)*/ 753,
+	/*log(1.20215)*/ 754,
+	/*log(1.20239)*/ 754,
+	/*log(1.20264)*/ 755,
+	/*log(1.20288)*/ 756,
+	/*log(1.20312)*/ 757,
+	/*log(1.20337)*/ 758,
+	/*log(1.20361)*/ 759,
+	/*log(1.20386)*/ 759,
+	/*log(1.20410)*/ 760,
+	/*log(1.20435)*/ 761,
+	/*log(1.20459)*/ 762,
+	/*log(1.20483)*/ 763,
+	/*log(1.20508)*/ 764,
+	/*log(1.20532)*/ 764,
+	/*log(1.20557)*/ 765,
+	/*log(1.20581)*/ 766,
+	/*log(1.20605)*/ 767,
+	/*log(1.20630)*/ 768,
+	/*log(1.20654)*/ 769,
+	/*log(1.20679)*/ 769,
+	/*log(1.20703)*/ 770,
+	/*log(1.20728)*/ 771,
+	/*log(1.20752)*/ 772,
+	/*log(1.20776)*/ 773,
+	/*log(1.20801)*/ 774,
+	/*log(1.20825)*/ 774,
+	/*log(1.20850)*/ 775,
+	/*log(1.20874)*/ 776,
+	/*log(1.20898)*/ 777,
+	/*log(1.20923)*/ 778,
+	/*log(1.20947)*/ 778,
+	/*log(1.20972)*/ 779,
+	/*log(1.20996)*/ 780,
+	/*log(1.21021)*/ 781,
+	/*log(1.21045)*/ 782,
+	/*log(1.21069)*/ 783,
+	/*log(1.21094)*/ 783,
+	/*log(1.21118)*/ 784,
+	/*log(1.21143)*/ 785,
+	/*log(1.21167)*/ 786,
+	/*log(1.21191)*/ 787,
+	/*log(1.21216)*/ 788,
+	/*log(1.21240)*/ 788,
+	/*log(1.21265)*/ 789,
+	/*log(1.21289)*/ 790,
+	/*log(1.21313)*/ 791,
+	/*log(1.21338)*/ 792,
+	/*log(1.21362)*/ 793,
+	/*log(1.21387)*/ 793,
+	/*log(1.21411)*/ 794,
+	/*log(1.21436)*/ 795,
+	/*log(1.21460)*/ 796,
+	/*log(1.21484)*/ 797,
+	/*log(1.21509)*/ 797,
+	/*log(1.21533)*/ 798,
+	/*log(1.21558)*/ 799,
+	/*log(1.21582)*/ 800,
+	/*log(1.21606)*/ 801,
+	/*log(1.21631)*/ 802,
+	/*log(1.21655)*/ 802,
+	/*log(1.21680)*/ 803,
+	/*log(1.21704)*/ 804,
+	/*log(1.21729)*/ 805,
+	/*log(1.21753)*/ 806,
+	/*log(1.21777)*/ 807,
+	/*log(1.21802)*/ 807,
+	/*log(1.21826)*/ 808,
+	/*log(1.21851)*/ 809,
+	/*log(1.21875)*/ 810,
+	/*log(1.21899)*/ 811,
+	/*log(1.21924)*/ 811,
+	/*log(1.21948)*/ 812,
+	/*log(1.21973)*/ 813,
+	/*log(1.21997)*/ 814,
+	/*log(1.22021)*/ 815,
+	/*log(1.22046)*/ 816,
+	/*log(1.22070)*/ 816,
+	/*log(1.22095)*/ 817,
+	/*log(1.22119)*/ 818,
+	/*log(1.22144)*/ 819,
+	/*log(1.22168)*/ 820,
+	/*log(1.22192)*/ 820,
+	/*log(1.22217)*/ 821,
+	/*log(1.22241)*/ 822,
+	/*log(1.22266)*/ 823,
+	/*log(1.22290)*/ 824,
+	/*log(1.22314)*/ 825,
+	/*log(1.22339)*/ 825,
+	/*log(1.22363)*/ 826,
+	/*log(1.22388)*/ 827,
+	/*log(1.22412)*/ 828,
+	/*log(1.22437)*/ 829,
+	/*log(1.22461)*/ 829,
+	/*log(1.22485)*/ 830,
+	/*log(1.22510)*/ 831,
+	/*log(1.22534)*/ 832,
+	/*log(1.22559)*/ 833,
+	/*log(1.22583)*/ 834,
+	/*log(1.22607)*/ 834,
+	/*log(1.22632)*/ 835,
+	/*log(1.22656)*/ 836,
+	/*log(1.22681)*/ 837,
+	/*log(1.22705)*/ 838,
+	/*log(1.22729)*/ 838,
+	/*log(1.22754)*/ 839,
+	/*log(1.22778)*/ 840,
+	/*log(1.22803)*/ 841,
+	/*log(1.22827)*/ 842,
+	/*log(1.22852)*/ 842,
+	/*log(1.22876)*/ 843,
+	/*log(1.22900)*/ 844,
+	/*log(1.22925)*/ 845,
+	/*log(1.22949)*/ 846,
+	/*log(1.22974)*/ 847,
+	/*log(1.22998)*/ 847,
+	/*log(1.23022)*/ 848,
+	/*log(1.23047)*/ 849,
+	/*log(1.23071)*/ 850,
+	/*log(1.23096)*/ 851,
+	/*log(1.23120)*/ 851,
+	/*log(1.23145)*/ 852,
+	/*log(1.23169)*/ 853,
+	/*log(1.23193)*/ 854,
+	/*log(1.23218)*/ 855,
+	/*log(1.23242)*/ 855,
+	/*log(1.23267)*/ 856,
+	/*log(1.23291)*/ 857,
+	/*log(1.23315)*/ 858,
+	/*log(1.23340)*/ 859,
+	/*log(1.23364)*/ 860,
+	/*log(1.23389)*/ 860,
+	/*log(1.23413)*/ 861,
+	/*log(1.23438)*/ 862,
+	/*log(1.23462)*/ 863,
+	/*log(1.23486)*/ 864,
+	/*log(1.23511)*/ 864,
+	/*log(1.23535)*/ 865,
+	/*log(1.23560)*/ 866,
+	/*log(1.23584)*/ 867,
+	/*log(1.23608)*/ 868,
+	/*log(1.23633)*/ 868,
+	/*log(1.23657)*/ 869,
+	/*log(1.23682)*/ 870,
+	/*log(1.23706)*/ 871,
+	/*log(1.23730)*/ 872,
+	/*log(1.23755)*/ 872,
+	/*log(1.23779)*/ 873,
+	/*log(1.23804)*/ 874,
+	/*log(1.23828)*/ 875,
+	/*log(1.23853)*/ 876,
+	/*log(1.23877)*/ 877,
+	/*log(1.23901)*/ 877,
+	/*log(1.23926)*/ 878,
+	/*log(1.23950)*/ 879,
+	/*log(1.23975)*/ 880,
+	/*log(1.23999)*/ 881,
+	/*log(1.24023)*/ 881,
+	/*log(1.24048)*/ 882,
+	/*log(1.24072)*/ 883,
+	/*log(1.24097)*/ 884,
+	/*log(1.24121)*/ 885,
+	/*log(1.24146)*/ 885,
+	/*log(1.24170)*/ 886,
+	/*log(1.24194)*/ 887,
+	/*log(1.24219)*/ 888,
+	/*log(1.24243)*/ 889,
+	/*log(1.24268)*/ 889,
+	/*log(1.24292)*/ 890,
+	/*log(1.24316)*/ 891,
+	/*log(1.24341)*/ 892,
+	/*log(1.24365)*/ 893,
+	/*log(1.24390)*/ 893,
+	/*log(1.24414)*/ 894,
+	/*log(1.24438)*/ 895,
+	/*log(1.24463)*/ 896,
+	/*log(1.24487)*/ 897,
+	/*log(1.24512)*/ 897,
+	/*log(1.24536)*/ 898,
+	/*log(1.24561)*/ 899,
+	/*log(1.24585)*/ 900,
+	/*log(1.24609)*/ 901,
+	/*log(1.24634)*/ 901,
+	/*log(1.24658)*/ 902,
+	/*log(1.24683)*/ 903,
+	/*log(1.24707)*/ 904,
+	/*log(1.24731)*/ 905,
+	/*log(1.24756)*/ 905,
+	/*log(1.24780)*/ 906,
+	/*log(1.24805)*/ 907,
+	/*log(1.24829)*/ 908,
+	/*log(1.24854)*/ 909,
+	/*log(1.24878)*/ 909,
+	/*log(1.24902)*/ 910,
+	/*log(1.24927)*/ 911,
+	/*log(1.24951)*/ 912,
+	/*log(1.24976)*/ 913,
+	/*log(1.25000)*/ 913,
+	/*log(1.25024)*/ 914,
+	/*log(1.25049)*/ 915,
+	/*log(1.25073)*/ 916,
+	/*log(1.25098)*/ 917,
+	/*log(1.25122)*/ 917,
+	/*log(1.25146)*/ 918,
+	/*log(1.25171)*/ 919,
+	/*log(1.25195)*/ 920,
+	/*log(1.25220)*/ 921,
+	/*log(1.25244)*/ 921,
+	/*log(1.25269)*/ 922,
+	/*log(1.25293)*/ 923,
+	/*log(1.25317)*/ 924,
+	/*log(1.25342)*/ 925,
+	/*log(1.25366)*/ 925,
+	/*log(1.25391)*/ 926,
+	/*log(1.25415)*/ 927,
+	/*log(1.25439)*/ 928,
+	/*log(1.25464)*/ 929,
+	/*log(1.25488)*/ 929,
+	/*log(1.25513)*/ 930,
+	/*log(1.25537)*/ 931,
+	/*log(1.25562)*/ 932,
+	/*log(1.25586)*/ 933,
+	/*log(1.25610)*/ 933,
+	/*log(1.25635)*/ 934,
+	/*log(1.25659)*/ 935,
+	/*log(1.25684)*/ 936,
+	/*log(1.25708)*/ 937,
+	/*log(1.25732)*/ 937,
+	/*log(1.25757)*/ 938,
+	/*log(1.25781)*/ 939,
+	/*log(1.25806)*/ 940,
+	/*log(1.25830)*/ 941,
+	/*log(1.25854)*/ 941,
+	/*log(1.25879)*/ 942,
+	/*log(1.25903)*/ 943,
+	/*log(1.25928)*/ 944,
+	/*log(1.25952)*/ 945,
+	/*log(1.25977)*/ 945,
+	/*log(1.26001)*/ 946,
+	/*log(1.26025)*/ 947,
+	/*log(1.26050)*/ 948,
+	/*log(1.26074)*/ 949,
+	/*log(1.26099)*/ 949,
+	/*log(1.26123)*/ 950,
+	/*log(1.26147)*/ 951,
+	/*log(1.26172)*/ 952,
+	/*log(1.26196)*/ 953,
+	/*log(1.26221)*/ 953,
+	/*log(1.26245)*/ 954,
+	/*log(1.26270)*/ 955,
+	/*log(1.26294)*/ 956,
+	/*log(1.26318)*/ 956,
+	/*log(1.26343)*/ 957,
+	/*log(1.26367)*/ 958,
+	/*log(1.26392)*/ 959,
+	/*log(1.26416)*/ 960,
+	/*log(1.26440)*/ 960,
+	/*log(1.26465)*/ 961,
+	/*log(1.26489)*/ 962,
+	/*log(1.26514)*/ 963,
+	/*log(1.26538)*/ 964,
+	/*log(1.26562)*/ 964,
+	/*log(1.26587)*/ 965,
+	/*log(1.26611)*/ 966,
+	/*log(1.26636)*/ 967,
+	/*log(1.26660)*/ 968,
+	/*log(1.26685)*/ 968,
+	/*log(1.26709)*/ 969,
+	/*log(1.26733)*/ 970,
+	/*log(1.26758)*/ 971,
+	/*log(1.26782)*/ 971,
+	/*log(1.26807)*/ 972,
+	/*log(1.26831)*/ 973,
+	/*log(1.26855)*/ 974,
+	/*log(1.26880)*/ 975,
+	/*log(1.26904)*/ 975,
+	/*log(1.26929)*/ 976,
+	/*log(1.26953)*/ 977,
+	/*log(1.26978)*/ 978,
+	/*log(1.27002)*/ 979,
+	/*log(1.27026)*/ 979,
+	/*log(1.27051)*/ 980,
+	/*log(1.27075)*/ 981,
+	/*log(1.27100)*/ 982,
+	/*log(1.27124)*/ 983,
+	/*log(1.27148)*/ 983,
+	/*log(1.27173)*/ 984,
+	/*log(1.27197)*/ 985,
+	/*log(1.27222)*/ 986,
+	/*log(1.27246)*/ 986,
+	/*log(1.27271)*/ 987,
+	/*log(1.27295)*/ 988,
+	/*log(1.27319)*/ 989,
+	/*log(1.27344)*/ 990,
+	/*log(1.27368)*/ 990,
+	/*log(1.27393)*/ 991,
+	/*log(1.27417)*/ 992,
+	/*log(1.27441)*/ 993,
+	/*log(1.27466)*/ 994,
+	/*log(1.27490)*/ 994,
+	/*log(1.27515)*/ 995,
+	/*log(1.27539)*/ 996,
+	/*log(1.27563)*/ 997,
+	/*log(1.27588)*/ 997,
+	/*log(1.27612)*/ 998,
+	/*log(1.27637)*/ 999,
+	/*log(1.27661)*/ 1000,
+	/*log(1.27686)*/ 1001,
+	/*log(1.27710)*/ 1001,
+	/*log(1.27734)*/ 1002,
+	/*log(1.27759)*/ 1003,
+	/*log(1.27783)*/ 1004,
+	/*log(1.27808)*/ 1004,
+	/*log(1.27832)*/ 1005,
+	/*log(1.27856)*/ 1006,
+	/*log(1.27881)*/ 1007,
+	/*log(1.27905)*/ 1008,
+	/*log(1.27930)*/ 1008,
+	/*log(1.27954)*/ 1009,
+	/*log(1.27979)*/ 1010,
+	/*log(1.28003)*/ 1011,
+	/*log(1.28027)*/ 1012,
+	/*log(1.28052)*/ 1012,
+	/*log(1.28076)*/ 1013,
+	/*log(1.28101)*/ 1014,
+	/*log(1.28125)*/ 1015,
+	/*log(1.28149)*/ 1015,
+	/*log(1.28174)*/ 1016,
+	/*log(1.28198)*/ 1017,
+	/*log(1.28223)*/ 1018,
+	/*log(1.28247)*/ 1019,
+	/*log(1.28271)*/ 1019,
+	/*log(1.28296)*/ 1020,
+	/*log(1.28320)*/ 1021,
+	/*log(1.28345)*/ 1022,
+	/*log(1.28369)*/ 1022,
+	/*log(1.28394)*/ 1023,
+	/*log(1.28418)*/ 1024,
+	/*log(1.28442)*/ 1025,
+	/*log(1.28467)*/ 1026,
+	/*log(1.28491)*/ 1026,
+	/*log(1.28516)*/ 1027,
+	/*log(1.28540)*/ 1028,
+	/*log(1.28564)*/ 1029,
+	/*log(1.28589)*/ 1029,
+	/*log(1.28613)*/ 1030,
+	/*log(1.28638)*/ 1031,
+	/*log(1.28662)*/ 1032,
+	/*log(1.28687)*/ 1033,
+	/*log(1.28711)*/ 1033,
+	/*log(1.28735)*/ 1034,
+	/*log(1.28760)*/ 1035,
+	/*log(1.28784)*/ 1036,
+	/*log(1.28809)*/ 1036,
+	/*log(1.28833)*/ 1037,
+	/*log(1.28857)*/ 1038,
+	/*log(1.28882)*/ 1039,
+	/*log(1.28906)*/ 1040,
+	/*log(1.28931)*/ 1040,
+	/*log(1.28955)*/ 1041,
+	/*log(1.28979)*/ 1042,
+	/*log(1.29004)*/ 1043,
+	/*log(1.29028)*/ 1043,
+	/*log(1.29053)*/ 1044,
+	/*log(1.29077)*/ 1045,
+	/*log(1.29102)*/ 1046,
+	/*log(1.29126)*/ 1047,
+	/*log(1.29150)*/ 1047,
+	/*log(1.29175)*/ 1048,
+	/*log(1.29199)*/ 1049,
+	/*log(1.29224)*/ 1050,
+	/*log(1.29248)*/ 1050,
+	/*log(1.29272)*/ 1051,
+	/*log(1.29297)*/ 1052,
+	/*log(1.29321)*/ 1053,
+	/*log(1.29346)*/ 1053,
+	/*log(1.29370)*/ 1054,
+	/*log(1.29395)*/ 1055,
+	/*log(1.29419)*/ 1056,
+	/*log(1.29443)*/ 1057,
+	/*log(1.29468)*/ 1057,
+	/*log(1.29492)*/ 1058,
+	/*log(1.29517)*/ 1059,
+	/*log(1.29541)*/ 1060,
+	/*log(1.29565)*/ 1060,
+	/*log(1.29590)*/ 1061,
+	/*log(1.29614)*/ 1062,
+	/*log(1.29639)*/ 1063,
+	/*log(1.29663)*/ 1064,
+	/*log(1.29688)*/ 1064,
+	/*log(1.29712)*/ 1065,
+	/*log(1.29736)*/ 1066,
+	/*log(1.29761)*/ 1067,
+	/*log(1.29785)*/ 1067,
+	/*log(1.29810)*/ 1068,
+	/*log(1.29834)*/ 1069,
+	/*log(1.29858)*/ 1070,
+	/*log(1.29883)*/ 1070,
+	/*log(1.29907)*/ 1071,
+	/*log(1.29932)*/ 1072,
+	/*log(1.29956)*/ 1073,
+	/*log(1.29980)*/ 1074,
+	/*log(1.30005)*/ 1074,
+	/*log(1.30029)*/ 1075,
+	/*log(1.30054)*/ 1076,
+	/*log(1.30078)*/ 1077,
+	/*log(1.30103)*/ 1077,
+	/*log(1.30127)*/ 1078,
+	/*log(1.30151)*/ 1079,
+	/*log(1.30176)*/ 1080,
+	/*log(1.30200)*/ 1080,
+	/*log(1.30225)*/ 1081,
+	/*log(1.30249)*/ 1082,
+	/*log(1.30273)*/ 1083,
+	/*log(1.30298)*/ 1084,
+	/*log(1.30322)*/ 1084,
+	/*log(1.30347)*/ 1085,
+	/*log(1.30371)*/ 1086,
+	/*log(1.30396)*/ 1087,
+	/*log(1.30420)*/ 1087,
+	/*log(1.30444)*/ 1088,
+	/*log(1.30469)*/ 1089,
+	/*log(1.30493)*/ 1090,
+	/*log(1.30518)*/ 1090,
+	/*log(1.30542)*/ 1091,
+	/*log(1.30566)*/ 1092,
+	/*log(1.30591)*/ 1093,
+	/*log(1.30615)*/ 1093,
+	/*log(1.30640)*/ 1094,
+	/*log(1.30664)*/ 1095,
+	/*log(1.30688)*/ 1096,
+	/*log(1.30713)*/ 1097,
+	/*log(1.30737)*/ 1097,
+	/*log(1.30762)*/ 1098,
+	/*log(1.30786)*/ 1099,
+	/*log(1.30811)*/ 1100,
+	/*log(1.30835)*/ 1100,
+	/*log(1.30859)*/ 1101,
+	/*log(1.30884)*/ 1102,
+	/*log(1.30908)*/ 1103,
+	/*log(1.30933)*/ 1103,
+	/*log(1.30957)*/ 1104,
+	/*log(1.30981)*/ 1105,
+	/*log(1.31006)*/ 1106,
+	/*log(1.31030)*/ 1106,
+	/*log(1.31055)*/ 1107,
+	/*log(1.31079)*/ 1108,
+	/*log(1.31104)*/ 1109,
+	/*log(1.31128)*/ 1110,
+	/*log(1.31152)*/ 1110,
+	/*log(1.31177)*/ 1111,
+	/*log(1.31201)*/ 1112,
+	/*log(1.31226)*/ 1113,
+	/*log(1.31250)*/ 1113,
+	/*log(1.31274)*/ 1114,
+	/*log(1.31299)*/ 1115,
+	/*log(1.31323)*/ 1116,
+	/*log(1.31348)*/ 1116,
+	/*log(1.31372)*/ 1117,
+	/*log(1.31396)*/ 1118,
+	/*log(1.31421)*/ 1119,
+	/*log(1.31445)*/ 1119,
+	/*log(1.31470)*/ 1120,
+	/*log(1.31494)*/ 1121,
+	/*log(1.31519)*/ 1122,
+	/*log(1.31543)*/ 1122,
+	/*log(1.31567)*/ 1123,
+	/*log(1.31592)*/ 1124,
+	/*log(1.31616)*/ 1125,
+	/*log(1.31641)*/ 1126,
+	/*log(1.31665)*/ 1126,
+	/*log(1.31689)*/ 1127,
+	/*log(1.31714)*/ 1128,
+	/*log(1.31738)*/ 1129,
+	/*log(1.31763)*/ 1129,
+	/*log(1.31787)*/ 1130,
+	/*log(1.31812)*/ 1131,
+	/*log(1.31836)*/ 1132,
+	/*log(1.31860)*/ 1132,
+	/*log(1.31885)*/ 1133,
+	/*log(1.31909)*/ 1134,
+	/*log(1.31934)*/ 1135,
+	/*log(1.31958)*/ 1135,
+	/*log(1.31982)*/ 1136,
+	/*log(1.32007)*/ 1137,
+	/*log(1.32031)*/ 1138,
+	/*log(1.32056)*/ 1138,
+	/*log(1.32080)*/ 1139,
+	/*log(1.32104)*/ 1140,
+	/*log(1.32129)*/ 1141,
+	/*log(1.32153)*/ 1141,
+	/*log(1.32178)*/ 1142,
+	/*log(1.32202)*/ 1143,
+	/*log(1.32227)*/ 1144,
+	/*log(1.32251)*/ 1144,
+	/*log(1.32275)*/ 1145,
+	/*log(1.32300)*/ 1146,
+	/*log(1.32324)*/ 1147,
+	/*log(1.32349)*/ 1147,
+	/*log(1.32373)*/ 1148,
+	/*log(1.32397)*/ 1149,
+	/*log(1.32422)*/ 1150,
+	/*log(1.32446)*/ 1151,
+	/*log(1.32471)*/ 1151,
+	/*log(1.32495)*/ 1152,
+	/*log(1.32520)*/ 1153,
+	/*log(1.32544)*/ 1154,
+	/*log(1.32568)*/ 1154,
+	/*log(1.32593)*/ 1155,
+	/*log(1.32617)*/ 1156,
+	/*log(1.32642)*/ 1157,
+	/*log(1.32666)*/ 1157,
+	/*log(1.32690)*/ 1158,
+	/*log(1.32715)*/ 1159,
+	/*log(1.32739)*/ 1160,
+	/*log(1.32764)*/ 1160,
+	/*log(1.32788)*/ 1161,
+	/*log(1.32812)*/ 1162,
+	/*log(1.32837)*/ 1163,
+	/*log(1.32861)*/ 1163,
+	/*log(1.32886)*/ 1164,
+	/*log(1.32910)*/ 1165,
+	/*log(1.32935)*/ 1166,
+	/*log(1.32959)*/ 1166,
+	/*log(1.32983)*/ 1167,
+	/*log(1.33008)*/ 1168,
+	/*log(1.33032)*/ 1169,
+	/*log(1.33057)*/ 1169,
+	/*log(1.33081)*/ 1170,
+	/*log(1.33105)*/ 1171,
+	/*log(1.33130)*/ 1172,
+	/*log(1.33154)*/ 1172,
+	/*log(1.33179)*/ 1173,
+	/*log(1.33203)*/ 1174,
+	/*log(1.33228)*/ 1175,
+	/*log(1.33252)*/ 1175,
+	/*log(1.33276)*/ 1176,
+	/*log(1.33301)*/ 1177,
+	/*log(1.33325)*/ 1178,
+	/*log(1.33350)*/ 1178,
+	/*log(1.33374)*/ 1179,
+	/*log(1.33398)*/ 1180,
+	/*log(1.33423)*/ 1181,
+	/*log(1.33447)*/ 1181,
+	/*log(1.33472)*/ 1182,
+	/*log(1.33496)*/ 1183,
+	/*log(1.33521)*/ 1184,
+	/*log(1.33545)*/ 1184,
+	/*log(1.33569)*/ 1185,
+	/*log(1.33594)*/ 1186,
+	/*log(1.33618)*/ 1187,
+	/*log(1.33643)*/ 1187,
+	/*log(1.33667)*/ 1188,
+	/*log(1.33691)*/ 1189,
+	/*log(1.33716)*/ 1190,
+	/*log(1.33740)*/ 1190,
+	/*log(1.33765)*/ 1191,
+	/*log(1.33789)*/ 1192,
+	/*log(1.33813)*/ 1193,
+	/*log(1.33838)*/ 1193,
+	/*log(1.33862)*/ 1194,
+	/*log(1.33887)*/ 1195,
+	/*log(1.33911)*/ 1196,
+	/*log(1.33936)*/ 1196,
+	/*log(1.33960)*/ 1197,
+	/*log(1.33984)*/ 1198,
+	/*log(1.34009)*/ 1199,
+	/*log(1.34033)*/ 1199,
+	/*log(1.34058)*/ 1200,
+	/*log(1.34082)*/ 1201,
+	/*log(1.34106)*/ 1202,
+	/*log(1.34131)*/ 1202,
+	/*log(1.34155)*/ 1203,
+	/*log(1.34180)*/ 1204,
+	/*log(1.34204)*/ 1205,
+	/*log(1.34229)*/ 1205,
+	/*log(1.34253)*/ 1206,
+	/*log(1.34277)*/ 1207,
+	/*log(1.34302)*/ 1207,
+	/*log(1.34326)*/ 1208,
+	/*log(1.34351)*/ 1209,
+	/*log(1.34375)*/ 1210,
+	/*log(1.34399)*/ 1210,
+	/*log(1.34424)*/ 1211,
+	/*log(1.34448)*/ 1212,
+	/*log(1.34473)*/ 1213,
+	/*log(1.34497)*/ 1213,
+	/*log(1.34521)*/ 1214,
+	/*log(1.34546)*/ 1215,
+	/*log(1.34570)*/ 1216,
+	/*log(1.34595)*/ 1216,
+	/*log(1.34619)*/ 1217,
+	/*log(1.34644)*/ 1218,
+	/*log(1.34668)*/ 1219,
+	/*log(1.34692)*/ 1219,
+	/*log(1.34717)*/ 1220,
+	/*log(1.34741)*/ 1221,
+	/*log(1.34766)*/ 1222,
+	/*log(1.34790)*/ 1222,
+	/*log(1.34814)*/ 1223,
+	/*log(1.34839)*/ 1224,
+	/*log(1.34863)*/ 1225,
+	/*log(1.34888)*/ 1225,
+	/*log(1.34912)*/ 1226,
+	/*log(1.34937)*/ 1227,
+	/*log(1.34961)*/ 1228,
+	/*log(1.34985)*/ 1228,
+	/*log(1.35010)*/ 1229,
+	/*log(1.35034)*/ 1230,
+	/*log(1.35059)*/ 1231,
+	/*log(1.35083)*/ 1231,
+	/*log(1.35107)*/ 1232,
+	/*log(1.35132)*/ 1233,
+	/*log(1.35156)*/ 1233,
+	/*log(1.35181)*/ 1234,
+	/*log(1.35205)*/ 1235,
+	/*log(1.35229)*/ 1236,
+	/*log(1.35254)*/ 1236,
+	/*log(1.35278)*/ 1237,
+	/*log(1.35303)*/ 1238,
+	/*log(1.35327)*/ 1239,
+	/*log(1.35352)*/ 1239,
+	/*log(1.35376)*/ 1240,
+	/*log(1.35400)*/ 1241,
+	/*log(1.35425)*/ 1242,
+	/*log(1.35449)*/ 1242,
+	/*log(1.35474)*/ 1243,
+	/*log(1.35498)*/ 1244,
+	/*log(1.35522)*/ 1245,
+	/*log(1.35547)*/ 1245,
+	/*log(1.35571)*/ 1246,
+	/*log(1.35596)*/ 1247,
+	/*log(1.35620)*/ 1248,
+	/*log(1.35645)*/ 1248,
+	/*log(1.35669)*/ 1249,
+	/*log(1.35693)*/ 1250,
+	/*log(1.35718)*/ 1250,
+	/*log(1.35742)*/ 1251,
+	/*log(1.35767)*/ 1252,
+	/*log(1.35791)*/ 1253,
+	/*log(1.35815)*/ 1253,
+	/*log(1.35840)*/ 1254,
+	/*log(1.35864)*/ 1255,
+	/*log(1.35889)*/ 1256,
+	/*log(1.35913)*/ 1256,
+	/*log(1.35938)*/ 1257,
+	/*log(1.35962)*/ 1258,
+	/*log(1.35986)*/ 1259,
+	/*log(1.36011)*/ 1259,
+	/*log(1.36035)*/ 1260,
+	/*log(1.36060)*/ 1261,
+	/*log(1.36084)*/ 1261,
+	/*log(1.36108)*/ 1262,
+	/*log(1.36133)*/ 1263,
+	/*log(1.36157)*/ 1264,
+	/*log(1.36182)*/ 1264,
+	/*log(1.36206)*/ 1265,
+	/*log(1.36230)*/ 1266,
+	/*log(1.36255)*/ 1267,
+	/*log(1.36279)*/ 1267,
+	/*log(1.36304)*/ 1268,
+	/*log(1.36328)*/ 1269,
+	/*log(1.36353)*/ 1270,
+	/*log(1.36377)*/ 1270,
+	/*log(1.36401)*/ 1271,
+	/*log(1.36426)*/ 1272,
+	/*log(1.36450)*/ 1272,
+	/*log(1.36475)*/ 1273,
+	/*log(1.36499)*/ 1274,
+	/*log(1.36523)*/ 1275,
+	/*log(1.36548)*/ 1275,
+	/*log(1.36572)*/ 1276,
+	/*log(1.36597)*/ 1277,
+	/*log(1.36621)*/ 1278,
+	/*log(1.36646)*/ 1278,
+	/*log(1.36670)*/ 1279,
+	/*log(1.36694)*/ 1280,
+	/*log(1.36719)*/ 1281,
+	/*log(1.36743)*/ 1281,
+	/*log(1.36768)*/ 1282,
+	/*log(1.36792)*/ 1283,
+	/*log(1.36816)*/ 1283,
+	/*log(1.36841)*/ 1284,
+	/*log(1.36865)*/ 1285,
+	/*log(1.36890)*/ 1286,
+	/*log(1.36914)*/ 1286,
+	/*log(1.36938)*/ 1287,
+	/*log(1.36963)*/ 1288,
+	/*log(1.36987)*/ 1289,
+	/*log(1.37012)*/ 1289,
+	/*log(1.37036)*/ 1290,
+	/*log(1.37061)*/ 1291,
+	/*log(1.37085)*/ 1292,
+	/*log(1.37109)*/ 1292,
+	/*log(1.37134)*/ 1293,
+	/*log(1.37158)*/ 1294,
+	/*log(1.37183)*/ 1294,
+	/*log(1.37207)*/ 1295,
+	/*log(1.37231)*/ 1296,
+	/*log(1.37256)*/ 1297,
+	/*log(1.37280)*/ 1297,
+	/*log(1.37305)*/ 1298,
+	/*log(1.37329)*/ 1299,
+	/*log(1.37354)*/ 1300,
+	/*log(1.37378)*/ 1300,
+	/*log(1.37402)*/ 1301,
+	/*log(1.37427)*/ 1302,
+	/*log(1.37451)*/ 1302,
+	/*log(1.37476)*/ 1303,
+	/*log(1.37500)*/ 1304,
+	/*log(1.37524)*/ 1305,
+	/*log(1.37549)*/ 1305,
+	/*log(1.37573)*/ 1306,
+	/*log(1.37598)*/ 1307,
+	/*log(1.37622)*/ 1308,
+	/*log(1.37646)*/ 1308,
+	/*log(1.37671)*/ 1309,
+	/*log(1.37695)*/ 1310,
+	/*log(1.37720)*/ 1310,
+	/*log(1.37744)*/ 1311,
+	/*log(1.37769)*/ 1312,
+	/*log(1.37793)*/ 1313,
+	/*log(1.37817)*/ 1313,
+	/*log(1.37842)*/ 1314,
+	/*log(1.37866)*/ 1315,
+	/*log(1.37891)*/ 1316,
+	/*log(1.37915)*/ 1316,
+	/*log(1.37939)*/ 1317,
+	/*log(1.37964)*/ 1318,
+	/*log(1.37988)*/ 1318,
+	/*log(1.38013)*/ 1319,
+	/*log(1.38037)*/ 1320,
+	/*log(1.38062)*/ 1321,
+	/*log(1.38086)*/ 1321,
+	/*log(1.38110)*/ 1322,
+	/*log(1.38135)*/ 1323,
+	/*log(1.38159)*/ 1323,
+	/*log(1.38184)*/ 1324,
+	/*log(1.38208)*/ 1325,
+	/*log(1.38232)*/ 1326,
+	/*log(1.38257)*/ 1326,
+	/*log(1.38281)*/ 1327,
+	/*log(1.38306)*/ 1328,
+	/*log(1.38330)*/ 1329,
+	/*log(1.38354)*/ 1329,
+	/*log(1.38379)*/ 1330,
+	/*log(1.38403)*/ 1331,
+	/*log(1.38428)*/ 1331,
+	/*log(1.38452)*/ 1332,
+	/*log(1.38477)*/ 1333,
+	/*log(1.38501)*/ 1334,
+	/*log(1.38525)*/ 1334,
+	/*log(1.38550)*/ 1335,
+	/*log(1.38574)*/ 1336,
+	/*log(1.38599)*/ 1336,
+	/*log(1.38623)*/ 1337,
+	/*log(1.38647)*/ 1338,
+	/*log(1.38672)*/ 1339,
+	/*log(1.38696)*/ 1339,
+	/*log(1.38721)*/ 1340,
+	/*log(1.38745)*/ 1341,
+	/*log(1.38770)*/ 1342,
+	/*log(1.38794)*/ 1342,
+	/*log(1.38818)*/ 1343,
+	/*log(1.38843)*/ 1344,
+	/*log(1.38867)*/ 1344,
+	/*log(1.38892)*/ 1345,
+	/*log(1.38916)*/ 1346,
+	/*log(1.38940)*/ 1347,
+	/*log(1.38965)*/ 1347,
+	/*log(1.38989)*/ 1348,
+	/*log(1.39014)*/ 1349,
+	/*log(1.39038)*/ 1349,
+	/*log(1.39062)*/ 1350,
+	/*log(1.39087)*/ 1351,
+	/*log(1.39111)*/ 1352,
+	/*log(1.39136)*/ 1352,
+	/*log(1.39160)*/ 1353,
+	/*log(1.39185)*/ 1354,
+	/*log(1.39209)*/ 1354,
+	/*log(1.39233)*/ 1355,
+	/*log(1.39258)*/ 1356,
+	/*log(1.39282)*/ 1357,
+	/*log(1.39307)*/ 1357,
+	/*log(1.39331)*/ 1358,
+	/*log(1.39355)*/ 1359,
+	/*log(1.39380)*/ 1360,
+	/*log(1.39404)*/ 1360,
+	/*log(1.39429)*/ 1361,
+	/*log(1.39453)*/ 1362,
+	/*log(1.39478)*/ 1362,
+	/*log(1.39502)*/ 1363,
+	/*log(1.39526)*/ 1364,
+	/*log(1.39551)*/ 1365,
+	/*log(1.39575)*/ 1365,
+	/*log(1.39600)*/ 1366,
+	/*log(1.39624)*/ 1367,
+	/*log(1.39648)*/ 1367,
+	/*log(1.39673)*/ 1368,
+	/*log(1.39697)*/ 1369,
+	/*log(1.39722)*/ 1370,
+	/*log(1.39746)*/ 1370,
+	/*log(1.39771)*/ 1371,
+	/*log(1.39795)*/ 1372,
+	/*log(1.39819)*/ 1372,
+	/*log(1.39844)*/ 1373,
+	/*log(1.39868)*/ 1374,
+	/*log(1.39893)*/ 1375,
+	/*log(1.39917)*/ 1375,
+	/*log(1.39941)*/ 1376,
+	/*log(1.39966)*/ 1377,
+	/*log(1.39990)*/ 1377,
+	/*log(1.40015)*/ 1378,
+	/*log(1.40039)*/ 1379,
+	/*log(1.40063)*/ 1380,
+	/*log(1.40088)*/ 1380,
+	/*log(1.40112)*/ 1381,
+	/*log(1.40137)*/ 1382,
+	/*log(1.40161)*/ 1382,
+	/*log(1.40186)*/ 1383,
+	/*log(1.40210)*/ 1384,
+	/*log(1.40234)*/ 1385,
+	/*log(1.40259)*/ 1385,
+	/*log(1.40283)*/ 1386,
+	/*log(1.40308)*/ 1387,
+	/*log(1.40332)*/ 1387,
+	/*log(1.40356)*/ 1388,
+	/*log(1.40381)*/ 1389,
+	/*log(1.40405)*/ 1390,
+	/*log(1.40430)*/ 1390,
+	/*log(1.40454)*/ 1391,
+	/*log(1.40479)*/ 1392,
+	/*log(1.40503)*/ 1392,
+	/*log(1.40527)*/ 1393,
+	/*log(1.40552)*/ 1394,
+	/*log(1.40576)*/ 1395,
+	/*log(1.40601)*/ 1395,
+	/*log(1.40625)*/ 1396,
+	/*log(1.40649)*/ 1397,
+	/*log(1.40674)*/ 1397,
+	/*log(1.40698)*/ 1398,
+	/*log(1.40723)*/ 1399,
+	/*log(1.40747)*/ 1399,
+	/*log(1.40771)*/ 1400,
+	/*log(1.40796)*/ 1401,
+	/*log(1.40820)*/ 1402,
+	/*log(1.40845)*/ 1402,
+	/*log(1.40869)*/ 1403,
+	/*log(1.40894)*/ 1404,
+	/*log(1.40918)*/ 1404,
+	/*log(1.40942)*/ 1405,
+	/*log(1.40967)*/ 1406,
+	/*log(1.40991)*/ 1407,
+	/*log(1.41016)*/ 1407,
+	/*log(1.41040)*/ 1408,
+	/*log(1.41064)*/ 1409,
+	/*log(1.41089)*/ 1409,
+	/*log(1.41113)*/ 1410,
+	/*log(1.41138)*/ 1411,
+	/*log(1.41162)*/ 1412,
+	/*log(1.41187)*/ 1412,
+	/*log(1.41211)*/ 1413,
+	/*log(1.41235)*/ 1414,
+	/*log(1.41260)*/ 1414,
+	/*log(1.41284)*/ 1415,
+	/*log(1.41309)*/ 1416,
+	/*log(1.41333)*/ 1417,
+	/*log(1.41357)*/ 1417,
+	/*log(1.41382)*/ 1418,
+	/*log(1.41406)*/ 1419,
+	/*log(1.41431)*/ 1419,
+	/*log(1.41455)*/ 1420,
+	/*log(1.41479)*/ 1421,
+	/*log(1.41504)*/ 1421,
+	/*log(1.41528)*/ 1422,
+	/*log(1.41553)*/ 1423,
+	/*log(1.41577)*/ 1424,
+	/*log(1.41602)*/ 1424,
+	/*log(1.41626)*/ 1425,
+	/*log(1.41650)*/ 1426,
+	/*log(1.41675)*/ 1426,
+	/*log(1.41699)*/ 1427,
+	/*log(1.41724)*/ 1428,
+	/*log(1.41748)*/ 1429,
+	/*log(1.41772)*/ 1429,
+	/*log(1.41797)*/ 1430,
+	/*log(1.41821)*/ 1431,
+	/*log(1.41846)*/ 1431,
+	/*log(1.41870)*/ 1432,
+	/*log(1.41895)*/ 1433,
+	/*log(1.41919)*/ 1433,
+	/*log(1.41943)*/ 1434,
+	/*log(1.41968)*/ 1435,
+	/*log(1.41992)*/ 1436,
+	/*log(1.42017)*/ 1436,
+	/*log(1.42041)*/ 1437,
+	/*log(1.42065)*/ 1438,
+	/*log(1.42090)*/ 1438,
+	/*log(1.42114)*/ 1439,
+	/*log(1.42139)*/ 1440,
+	/*log(1.42163)*/ 1440,
+	/*log(1.42188)*/ 1441,
+	/*log(1.42212)*/ 1442,
+	/*log(1.42236)*/ 1443,
+	/*log(1.42261)*/ 1443,
+	/*log(1.42285)*/ 1444,
+	/*log(1.42310)*/ 1445,
+	/*log(1.42334)*/ 1445,
+	/*log(1.42358)*/ 1446,
+	/*log(1.42383)*/ 1447,
+	/*log(1.42407)*/ 1448,
+	/*log(1.42432)*/ 1448,
+	/*log(1.42456)*/ 1449,
+	/*log(1.42480)*/ 1450,
+	/*log(1.42505)*/ 1450,
+	/*log(1.42529)*/ 1451,
+	/*log(1.42554)*/ 1452,
+	/*log(1.42578)*/ 1452,
+	/*log(1.42603)*/ 1453,
+	/*log(1.42627)*/ 1454,
+	/*log(1.42651)*/ 1455,
+	/*log(1.42676)*/ 1455,
+	/*log(1.42700)*/ 1456,
+	/*log(1.42725)*/ 1457,
+	/*log(1.42749)*/ 1457,
+	/*log(1.42773)*/ 1458,
+	/*log(1.42798)*/ 1459,
+	/*log(1.42822)*/ 1459,
+	/*log(1.42847)*/ 1460,
+	/*log(1.42871)*/ 1461,
+	/*log(1.42896)*/ 1462,
+	/*log(1.42920)*/ 1462,
+	/*log(1.42944)*/ 1463,
+	/*log(1.42969)*/ 1464,
+	/*log(1.42993)*/ 1464,
+	/*log(1.43018)*/ 1465,
+	/*log(1.43042)*/ 1466,
+	/*log(1.43066)*/ 1466,
+	/*log(1.43091)*/ 1467,
+	/*log(1.43115)*/ 1468,
+	/*log(1.43140)*/ 1469,
+	/*log(1.43164)*/ 1469,
+	/*log(1.43188)*/ 1470,
+	/*log(1.43213)*/ 1471,
+	/*log(1.43237)*/ 1471,
+	/*log(1.43262)*/ 1472,
+	/*log(1.43286)*/ 1473,
+	/*log(1.43311)*/ 1473,
+	/*log(1.43335)*/ 1474,
+	/*log(1.43359)*/ 1475,
+	/*log(1.43384)*/ 1476,
+	/*log(1.43408)*/ 1476,
+	/*log(1.43433)*/ 1477,
+	/*log(1.43457)*/ 1478,
+	/*log(1.43481)*/ 1478,
+	/*log(1.43506)*/ 1479,
+	/*log(1.43530)*/ 1480,
+	/*log(1.43555)*/ 1480,
+	/*log(1.43579)*/ 1481,
+	/*log(1.43604)*/ 1482,
+	/*log(1.43628)*/ 1482,
+	/*log(1.43652)*/ 1483,
+	/*log(1.43677)*/ 1484,
+	/*log(1.43701)*/ 1485,
+	/*log(1.43726)*/ 1485,
+	/*log(1.43750)*/ 1486,
+	/*log(1.43774)*/ 1487,
+	/*log(1.43799)*/ 1487,
+	/*log(1.43823)*/ 1488,
+	/*log(1.43848)*/ 1489,
+	/*log(1.43872)*/ 1489,
+	/*log(1.43896)*/ 1490,
+	/*log(1.43921)*/ 1491,
+	/*log(1.43945)*/ 1492,
+	/*log(1.43970)*/ 1492,
+	/*log(1.43994)*/ 1493,
+	/*log(1.44019)*/ 1494,
+	/*log(1.44043)*/ 1494,
+	/*log(1.44067)*/ 1495,
+	/*log(1.44092)*/ 1496,
+	/*log(1.44116)*/ 1496,
+	/*log(1.44141)*/ 1497,
+	/*log(1.44165)*/ 1498,
+	/*log(1.44189)*/ 1498,
+	/*log(1.44214)*/ 1499,
+	/*log(1.44238)*/ 1500,
+	/*log(1.44263)*/ 1501,
+	/*log(1.44287)*/ 1501,
+	/*log(1.44312)*/ 1502,
+	/*log(1.44336)*/ 1503,
+	/*log(1.44360)*/ 1503,
+	/*log(1.44385)*/ 1504,
+	/*log(1.44409)*/ 1505,
+	/*log(1.44434)*/ 1505,
+	/*log(1.44458)*/ 1506,
+	/*log(1.44482)*/ 1507,
+	/*log(1.44507)*/ 1507,
+	/*log(1.44531)*/ 1508,
+	/*log(1.44556)*/ 1509,
+	/*log(1.44580)*/ 1510,
+	/*log(1.44604)*/ 1510,
+	/*log(1.44629)*/ 1511,
+	/*log(1.44653)*/ 1512,
+	/*log(1.44678)*/ 1512,
+	/*log(1.44702)*/ 1513,
+	/*log(1.44727)*/ 1514,
+	/*log(1.44751)*/ 1514,
+	/*log(1.44775)*/ 1515,
+	/*log(1.44800)*/ 1516,
+	/*log(1.44824)*/ 1516,
+	/*log(1.44849)*/ 1517,
+	/*log(1.44873)*/ 1518,
+	/*log(1.44897)*/ 1519,
+	/*log(1.44922)*/ 1519,
+	/*log(1.44946)*/ 1520,
+	/*log(1.44971)*/ 1521,
+	/*log(1.44995)*/ 1521,
+	/*log(1.45020)*/ 1522,
+	/*log(1.45044)*/ 1523,
+	/*log(1.45068)*/ 1523,
+	/*log(1.45093)*/ 1524,
+	/*log(1.45117)*/ 1525,
+	/*log(1.45142)*/ 1525,
+	/*log(1.45166)*/ 1526,
+	/*log(1.45190)*/ 1527,
+	/*log(1.45215)*/ 1527,
+	/*log(1.45239)*/ 1528,
+	/*log(1.45264)*/ 1529,
+	/*log(1.45288)*/ 1530,
+	/*log(1.45312)*/ 1530,
+	/*log(1.45337)*/ 1531,
+	/*log(1.45361)*/ 1532,
+	/*log(1.45386)*/ 1532,
+	/*log(1.45410)*/ 1533,
+	/*log(1.45435)*/ 1534,
+	/*log(1.45459)*/ 1534,
+	/*log(1.45483)*/ 1535,
+	/*log(1.45508)*/ 1536,
+	/*log(1.45532)*/ 1536,
+	/*log(1.45557)*/ 1537,
+	/*log(1.45581)*/ 1538,
+	/*log(1.45605)*/ 1538,
+	/*log(1.45630)*/ 1539,
+	/*log(1.45654)*/ 1540,
+	/*log(1.45679)*/ 1541,
+	/*log(1.45703)*/ 1541,
+	/*log(1.45728)*/ 1542,
+	/*log(1.45752)*/ 1543,
+	/*log(1.45776)*/ 1543,
+	/*log(1.45801)*/ 1544,
+	/*log(1.45825)*/ 1545,
+	/*log(1.45850)*/ 1545,
+	/*log(1.45874)*/ 1546,
+	/*log(1.45898)*/ 1547,
+	/*log(1.45923)*/ 1547,
+	/*log(1.45947)*/ 1548,
+	/*log(1.45972)*/ 1549,
+	/*log(1.45996)*/ 1549,
+	/*log(1.46021)*/ 1550,
+	/*log(1.46045)*/ 1551,
+	/*log(1.46069)*/ 1552,
+	/*log(1.46094)*/ 1552,
+	/*log(1.46118)*/ 1553,
+	/*log(1.46143)*/ 1554,
+	/*log(1.46167)*/ 1554,
+	/*log(1.46191)*/ 1555,
+	/*log(1.46216)*/ 1556,
+	/*log(1.46240)*/ 1556,
+	/*log(1.46265)*/ 1557,
+	/*log(1.46289)*/ 1558,
+	/*log(1.46313)*/ 1558,
+	/*log(1.46338)*/ 1559,
+	/*log(1.46362)*/ 1560,
+	/*log(1.46387)*/ 1560,
+	/*log(1.46411)*/ 1561,
+	/*log(1.46436)*/ 1562,
+	/*log(1.46460)*/ 1562,
+	/*log(1.46484)*/ 1563,
+	/*log(1.46509)*/ 1564,
+	/*log(1.46533)*/ 1565,
+	/*log(1.46558)*/ 1565,
+	/*log(1.46582)*/ 1566,
+	/*log(1.46606)*/ 1567,
+	/*log(1.46631)*/ 1567,
+	/*log(1.46655)*/ 1568,
+	/*log(1.46680)*/ 1569,
+	/*log(1.46704)*/ 1569,
+	/*log(1.46729)*/ 1570,
+	/*log(1.46753)*/ 1571,
+	/*log(1.46777)*/ 1571,
+	/*log(1.46802)*/ 1572,
+	/*log(1.46826)*/ 1573,
+	/*log(1.46851)*/ 1573,
+	/*log(1.46875)*/ 1574,
+	/*log(1.46899)*/ 1575,
+	/*log(1.46924)*/ 1575,
+	/*log(1.46948)*/ 1576,
+	/*log(1.46973)*/ 1577,
+	/*log(1.46997)*/ 1577,
+	/*log(1.47021)*/ 1578,
+	/*log(1.47046)*/ 1579,
+	/*log(1.47070)*/ 1579,
+	/*log(1.47095)*/ 1580,
+	/*log(1.47119)*/ 1581,
+	/*log(1.47144)*/ 1582,
+	/*log(1.47168)*/ 1582,
+	/*log(1.47192)*/ 1583,
+	/*log(1.47217)*/ 1584,
+	/*log(1.47241)*/ 1584,
+	/*log(1.47266)*/ 1585,
+	/*log(1.47290)*/ 1586,
+	/*log(1.47314)*/ 1586,
+	/*log(1.47339)*/ 1587,
+	/*log(1.47363)*/ 1588,
+	/*log(1.47388)*/ 1588,
+	/*log(1.47412)*/ 1589,
+	/*log(1.47437)*/ 1590,
+	/*log(1.47461)*/ 1590,
+	/*log(1.47485)*/ 1591,
+	/*log(1.47510)*/ 1592,
+	/*log(1.47534)*/ 1592,
+	/*log(1.47559)*/ 1593,
+	/*log(1.47583)*/ 1594,
+	/*log(1.47607)*/ 1594,
+	/*log(1.47632)*/ 1595,
+	/*log(1.47656)*/ 1596,
+	/*log(1.47681)*/ 1596,
+	/*log(1.47705)*/ 1597,
+	/*log(1.47729)*/ 1598,
+	/*log(1.47754)*/ 1598,
+	/*log(1.47778)*/ 1599,
+	/*log(1.47803)*/ 1600,
+	/*log(1.47827)*/ 1601,
+	/*log(1.47852)*/ 1601,
+	/*log(1.47876)*/ 1602,
+	/*log(1.47900)*/ 1603,
+	/*log(1.47925)*/ 1603,
+	/*log(1.47949)*/ 1604,
+	/*log(1.47974)*/ 1605,
+	/*log(1.47998)*/ 1605,
+	/*log(1.48022)*/ 1606,
+	/*log(1.48047)*/ 1607,
+	/*log(1.48071)*/ 1607,
+	/*log(1.48096)*/ 1608,
+	/*log(1.48120)*/ 1609,
+	/*log(1.48145)*/ 1609,
+	/*log(1.48169)*/ 1610,
+	/*log(1.48193)*/ 1611,
+	/*log(1.48218)*/ 1611,
+	/*log(1.48242)*/ 1612,
+	/*log(1.48267)*/ 1613,
+	/*log(1.48291)*/ 1613,
+	/*log(1.48315)*/ 1614,
+	/*log(1.48340)*/ 1615,
+	/*log(1.48364)*/ 1615,
+	/*log(1.48389)*/ 1616,
+	/*log(1.48413)*/ 1617,
+	/*log(1.48438)*/ 1617,
+	/*log(1.48462)*/ 1618,
+	/*log(1.48486)*/ 1619,
+	/*log(1.48511)*/ 1619,
+	/*log(1.48535)*/ 1620,
+	/*log(1.48560)*/ 1621,
+	/*log(1.48584)*/ 1621,
+	/*log(1.48608)*/ 1622,
+	/*log(1.48633)*/ 1623,
+	/*log(1.48657)*/ 1623,
+	/*log(1.48682)*/ 1624,
+	/*log(1.48706)*/ 1625,
+	/*log(1.48730)*/ 1625,
+	/*log(1.48755)*/ 1626,
+	/*log(1.48779)*/ 1627,
+	/*log(1.48804)*/ 1627,
+	/*log(1.48828)*/ 1628,
+	/*log(1.48853)*/ 1629,
+	/*log(1.48877)*/ 1630,
+	/*log(1.48901)*/ 1630,
+	/*log(1.48926)*/ 1631,
+	/*log(1.48950)*/ 1632,
+	/*log(1.48975)*/ 1632,
+	/*log(1.48999)*/ 1633,
+	/*log(1.49023)*/ 1634,
+	/*log(1.49048)*/ 1634,
+	/*log(1.49072)*/ 1635,
+	/*log(1.49097)*/ 1636,
+	/*log(1.49121)*/ 1636,
+	/*log(1.49146)*/ 1637,
+	/*log(1.49170)*/ 1638,
+	/*log(1.49194)*/ 1638,
+	/*log(1.49219)*/ 1639,
+	/*log(1.49243)*/ 1640,
+	/*log(1.49268)*/ 1640,
+	/*log(1.49292)*/ 1641,
+	/*log(1.49316)*/ 1642,
+	/*log(1.49341)*/ 1642,
+	/*log(1.49365)*/ 1643,
+	/*log(1.49390)*/ 1644,
+	/*log(1.49414)*/ 1644,
+	/*log(1.49438)*/ 1645,
+	/*log(1.49463)*/ 1646,
+	/*log(1.49487)*/ 1646,
+	/*log(1.49512)*/ 1647,
+	/*log(1.49536)*/ 1648,
+	/*log(1.49561)*/ 1648,
+	/*log(1.49585)*/ 1649,
+	/*log(1.49609)*/ 1650,
+	/*log(1.49634)*/ 1650,
+	/*log(1.49658)*/ 1651,
+	/*log(1.49683)*/ 1652,
+	/*log(1.49707)*/ 1652,
+	/*log(1.49731)*/ 1653,
+	/*log(1.49756)*/ 1654,
+	/*log(1.49780)*/ 1654,
+	/*log(1.49805)*/ 1655,
+	/*log(1.49829)*/ 1656,
+	/*log(1.49854)*/ 1656,
+	/*log(1.49878)*/ 1657,
+	/*log(1.49902)*/ 1658,
+	/*log(1.49927)*/ 1658,
+	/*log(1.49951)*/ 1659,
+	/*log(1.49976)*/ 1660,
+	/*log(1.50000)*/ 1660,
+	/*log(1.50024)*/ 1661,
+	/*log(1.50049)*/ 1662,
+	/*log(1.50073)*/ 1662,
+	/*log(1.50098)*/ 1663,
+	/*log(1.50122)*/ 1664,
+	/*log(1.50146)*/ 1664,
+	/*log(1.50171)*/ 1665,
+	/*log(1.50195)*/ 1666,
+	/*log(1.50220)*/ 1666,
+	/*log(1.50244)*/ 1667,
+	/*log(1.50269)*/ 1668,
+	/*log(1.50293)*/ 1668,
+	/*log(1.50317)*/ 1669,
+	/*log(1.50342)*/ 1670,
+	/*log(1.50366)*/ 1670,
+	/*log(1.50391)*/ 1671,
+	/*log(1.50415)*/ 1672,
+	/*log(1.50439)*/ 1672,
+	/*log(1.50464)*/ 1673,
+	/*log(1.50488)*/ 1674,
+	/*log(1.50513)*/ 1674,
+	/*log(1.50537)*/ 1675,
+	/*log(1.50562)*/ 1676,
+	/*log(1.50586)*/ 1676,
+	/*log(1.50610)*/ 1677,
+	/*log(1.50635)*/ 1678,
+	/*log(1.50659)*/ 1678,
+	/*log(1.50684)*/ 1679,
+	/*log(1.50708)*/ 1680,
+	/*log(1.50732)*/ 1680,
+	/*log(1.50757)*/ 1681,
+	/*log(1.50781)*/ 1682,
+	/*log(1.50806)*/ 1682,
+	/*log(1.50830)*/ 1683,
+	/*log(1.50854)*/ 1684,
+	/*log(1.50879)*/ 1684,
+	/*log(1.50903)*/ 1685,
+	/*log(1.50928)*/ 1686,
+	/*log(1.50952)*/ 1686,
+	/*log(1.50977)*/ 1687,
+	/*log(1.51001)*/ 1688,
+	/*log(1.51025)*/ 1688,
+	/*log(1.51050)*/ 1689,
+	/*log(1.51074)*/ 1690,
+	/*log(1.51099)*/ 1690,
+	/*log(1.51123)*/ 1691,
+	/*log(1.51147)*/ 1691,
+	/*log(1.51172)*/ 1692,
+	/*log(1.51196)*/ 1693,
+	/*log(1.51221)*/ 1693,
+	/*log(1.51245)*/ 1694,
+	/*log(1.51270)*/ 1695,
+	/*log(1.51294)*/ 1695,
+	/*log(1.51318)*/ 1696,
+	/*log(1.51343)*/ 1697,
+	/*log(1.51367)*/ 1697,
+	/*log(1.51392)*/ 1698,
+	/*log(1.51416)*/ 1699,
+	/*log(1.51440)*/ 1699,
+	/*log(1.51465)*/ 1700,
+	/*log(1.51489)*/ 1701,
+	/*log(1.51514)*/ 1701,
+	/*log(1.51538)*/ 1702,
+	/*log(1.51562)*/ 1703,
+	/*log(1.51587)*/ 1703,
+	/*log(1.51611)*/ 1704,
+	/*log(1.51636)*/ 1705,
+	/*log(1.51660)*/ 1705,
+	/*log(1.51685)*/ 1706,
+	/*log(1.51709)*/ 1707,
+	/*log(1.51733)*/ 1707,
+	/*log(1.51758)*/ 1708,
+	/*log(1.51782)*/ 1709,
+	/*log(1.51807)*/ 1709,
+	/*log(1.51831)*/ 1710,
+	/*log(1.51855)*/ 1711,
+	/*log(1.51880)*/ 1711,
+	/*log(1.51904)*/ 1712,
+	/*log(1.51929)*/ 1713,
+	/*log(1.51953)*/ 1713,
+	/*log(1.51978)*/ 1714,
+	/*log(1.52002)*/ 1715,
+	/*log(1.52026)*/ 1715,
+	/*log(1.52051)*/ 1716,
+	/*log(1.52075)*/ 1717,
+	/*log(1.52100)*/ 1717,
+	/*log(1.52124)*/ 1718,
+	/*log(1.52148)*/ 1719,
+	/*log(1.52173)*/ 1719,
+	/*log(1.52197)*/ 1720,
+	/*log(1.52222)*/ 1721,
+	/*log(1.52246)*/ 1721,
+	/*log(1.52271)*/ 1722,
+	/*log(1.52295)*/ 1722,
+	/*log(1.52319)*/ 1723,
+	/*log(1.52344)*/ 1724,
+	/*log(1.52368)*/ 1724,
+	/*log(1.52393)*/ 1725,
+	/*log(1.52417)*/ 1726,
+	/*log(1.52441)*/ 1726,
+	/*log(1.52466)*/ 1727,
+	/*log(1.52490)*/ 1728,
+	/*log(1.52515)*/ 1728,
+	/*log(1.52539)*/ 1729,
+	/*log(1.52563)*/ 1730,
+	/*log(1.52588)*/ 1730,
+	/*log(1.52612)*/ 1731,
+	/*log(1.52637)*/ 1732,
+	/*log(1.52661)*/ 1732,
+	/*log(1.52686)*/ 1733,
+	/*log(1.52710)*/ 1734,
+	/*log(1.52734)*/ 1734,
+	/*log(1.52759)*/ 1735,
+	/*log(1.52783)*/ 1736,
+	/*log(1.52808)*/ 1736,
+	/*log(1.52832)*/ 1737,
+	/*log(1.52856)*/ 1738,
+	/*log(1.52881)*/ 1738,
+	/*log(1.52905)*/ 1739,
+	/*log(1.52930)*/ 1740,
+	/*log(1.52954)*/ 1740,
+	/*log(1.52979)*/ 1741,
+	/*log(1.53003)*/ 1741,
+	/*log(1.53027)*/ 1742,
+	/*log(1.53052)*/ 1743,
+	/*log(1.53076)*/ 1743,
+	/*log(1.53101)*/ 1744,
+	/*log(1.53125)*/ 1745,
+	/*log(1.53149)*/ 1745,
+	/*log(1.53174)*/ 1746,
+	/*log(1.53198)*/ 1747,
+	/*log(1.53223)*/ 1747,
+	/*log(1.53247)*/ 1748,
+	/*log(1.53271)*/ 1749,
+	/*log(1.53296)*/ 1749,
+	/*log(1.53320)*/ 1750,
+	/*log(1.53345)*/ 1751,
+	/*log(1.53369)*/ 1751,
+	/*log(1.53394)*/ 1752,
+	/*log(1.53418)*/ 1753,
+	/*log(1.53442)*/ 1753,
+	/*log(1.53467)*/ 1754,
+	/*log(1.53491)*/ 1755,
+	/*log(1.53516)*/ 1755,
+	/*log(1.53540)*/ 1756,
+	/*log(1.53564)*/ 1756,
+	/*log(1.53589)*/ 1757,
+	/*log(1.53613)*/ 1758,
+	/*log(1.53638)*/ 1758,
+	/*log(1.53662)*/ 1759,
+	/*log(1.53687)*/ 1760,
+	/*log(1.53711)*/ 1760,
+	/*log(1.53735)*/ 1761,
+	/*log(1.53760)*/ 1762,
+	/*log(1.53784)*/ 1762,
+	/*log(1.53809)*/ 1763,
+	/*log(1.53833)*/ 1764,
+	/*log(1.53857)*/ 1764,
+	/*log(1.53882)*/ 1765,
+	/*log(1.53906)*/ 1766,
+	/*log(1.53931)*/ 1766,
+	/*log(1.53955)*/ 1767,
+	/*log(1.53979)*/ 1768,
+	/*log(1.54004)*/ 1768,
+	/*log(1.54028)*/ 1769,
+	/*log(1.54053)*/ 1769,
+	/*log(1.54077)*/ 1770,
+	/*log(1.54102)*/ 1771,
+	/*log(1.54126)*/ 1771,
+	/*log(1.54150)*/ 1772,
+	/*log(1.54175)*/ 1773,
+	/*log(1.54199)*/ 1773,
+	/*log(1.54224)*/ 1774,
+	/*log(1.54248)*/ 1775,
+	/*log(1.54272)*/ 1775,
+	/*log(1.54297)*/ 1776,
+	/*log(1.54321)*/ 1777,
+	/*log(1.54346)*/ 1777,
+	/*log(1.54370)*/ 1778,
+	/*log(1.54395)*/ 1779,
+	/*log(1.54419)*/ 1779,
+	/*log(1.54443)*/ 1780,
+	/*log(1.54468)*/ 1781,
+	/*log(1.54492)*/ 1781,
+	/*log(1.54517)*/ 1782,
+	/*log(1.54541)*/ 1782,
+	/*log(1.54565)*/ 1783,
+	/*log(1.54590)*/ 1784,
+	/*log(1.54614)*/ 1784,
+	/*log(1.54639)*/ 1785,
+	/*log(1.54663)*/ 1786,
+	/*log(1.54688)*/ 1786,
+	/*log(1.54712)*/ 1787,
+	/*log(1.54736)*/ 1788,
+	/*log(1.54761)*/ 1788,
+	/*log(1.54785)*/ 1789,
+	/*log(1.54810)*/ 1790,
+	/*log(1.54834)*/ 1790,
+	/*log(1.54858)*/ 1791,
+	/*log(1.54883)*/ 1791,
+	/*log(1.54907)*/ 1792,
+	/*log(1.54932)*/ 1793,
+	/*log(1.54956)*/ 1793,
+	/*log(1.54980)*/ 1794,
+	/*log(1.55005)*/ 1795,
+	/*log(1.55029)*/ 1795,
+	/*log(1.55054)*/ 1796,
+	/*log(1.55078)*/ 1797,
+	/*log(1.55103)*/ 1797,
+	/*log(1.55127)*/ 1798,
+	/*log(1.55151)*/ 1799,
+	/*log(1.55176)*/ 1799,
+	/*log(1.55200)*/ 1800,
+	/*log(1.55225)*/ 1801,
+	/*log(1.55249)*/ 1801,
+	/*log(1.55273)*/ 1802,
+	/*log(1.55298)*/ 1802,
+	/*log(1.55322)*/ 1803,
+	/*log(1.55347)*/ 1804,
+	/*log(1.55371)*/ 1804,
+	/*log(1.55396)*/ 1805,
+	/*log(1.55420)*/ 1806,
+	/*log(1.55444)*/ 1806,
+	/*log(1.55469)*/ 1807,
+	/*log(1.55493)*/ 1808,
+	/*log(1.55518)*/ 1808,
+	/*log(1.55542)*/ 1809,
+	/*log(1.55566)*/ 1810,
+	/*log(1.55591)*/ 1810,
+	/*log(1.55615)*/ 1811,
+	/*log(1.55640)*/ 1811,
+	/*log(1.55664)*/ 1812,
+	/*log(1.55688)*/ 1813,
+	/*log(1.55713)*/ 1813,
+	/*log(1.55737)*/ 1814,
+	/*log(1.55762)*/ 1815,
+	/*log(1.55786)*/ 1815,
+	/*log(1.55811)*/ 1816,
+	/*log(1.55835)*/ 1817,
+	/*log(1.55859)*/ 1817,
+	/*log(1.55884)*/ 1818,
+	/*log(1.55908)*/ 1819,
+	/*log(1.55933)*/ 1819,
+	/*log(1.55957)*/ 1820,
+	/*log(1.55981)*/ 1820,
+	/*log(1.56006)*/ 1821,
+	/*log(1.56030)*/ 1822,
+	/*log(1.56055)*/ 1822,
+	/*log(1.56079)*/ 1823,
+	/*log(1.56104)*/ 1824,
+	/*log(1.56128)*/ 1824,
+	/*log(1.56152)*/ 1825,
+	/*log(1.56177)*/ 1826,
+	/*log(1.56201)*/ 1826,
+	/*log(1.56226)*/ 1827,
+	/*log(1.56250)*/ 1827,
+	/*log(1.56274)*/ 1828,
+	/*log(1.56299)*/ 1829,
+	/*log(1.56323)*/ 1829,
+	/*log(1.56348)*/ 1830,
+	/*log(1.56372)*/ 1831,
+	/*log(1.56396)*/ 1831,
+	/*log(1.56421)*/ 1832,
+	/*log(1.56445)*/ 1833,
+	/*log(1.56470)*/ 1833,
+	/*log(1.56494)*/ 1834,
+	/*log(1.56519)*/ 1835,
+	/*log(1.56543)*/ 1835,
+	/*log(1.56567)*/ 1836,
+	/*log(1.56592)*/ 1836,
+	/*log(1.56616)*/ 1837,
+	/*log(1.56641)*/ 1838,
+	/*log(1.56665)*/ 1838,
+	/*log(1.56689)*/ 1839,
+	/*log(1.56714)*/ 1840,
+	/*log(1.56738)*/ 1840,
+	/*log(1.56763)*/ 1841,
+	/*log(1.56787)*/ 1842,
+	/*log(1.56812)*/ 1842,
+	/*log(1.56836)*/ 1843,
+	/*log(1.56860)*/ 1843,
+	/*log(1.56885)*/ 1844,
+	/*log(1.56909)*/ 1845,
+	/*log(1.56934)*/ 1845,
+	/*log(1.56958)*/ 1846,
+	/*log(1.56982)*/ 1847,
+	/*log(1.57007)*/ 1847,
+	/*log(1.57031)*/ 1848,
+	/*log(1.57056)*/ 1849,
+	/*log(1.57080)*/ 1849,
+	/*log(1.57104)*/ 1850,
+	/*log(1.57129)*/ 1850,
+	/*log(1.57153)*/ 1851,
+	/*log(1.57178)*/ 1852,
+	/*log(1.57202)*/ 1852,
+	/*log(1.57227)*/ 1853,
+	/*log(1.57251)*/ 1854,
+	/*log(1.57275)*/ 1854,
+	/*log(1.57300)*/ 1855,
+	/*log(1.57324)*/ 1856,
+	/*log(1.57349)*/ 1856,
+	/*log(1.57373)*/ 1857,
+	/*log(1.57397)*/ 1857,
+	/*log(1.57422)*/ 1858,
+	/*log(1.57446)*/ 1859,
+	/*log(1.57471)*/ 1859,
+	/*log(1.57495)*/ 1860,
+	/*log(1.57520)*/ 1861,
+	/*log(1.57544)*/ 1861,
+	/*log(1.57568)*/ 1862,
+	/*log(1.57593)*/ 1863,
+	/*log(1.57617)*/ 1863,
+	/*log(1.57642)*/ 1864,
+	/*log(1.57666)*/ 1864,
+	/*log(1.57690)*/ 1865,
+	/*log(1.57715)*/ 1866,
+	/*log(1.57739)*/ 1866,
+	/*log(1.57764)*/ 1867,
+	/*log(1.57788)*/ 1868,
+	/*log(1.57812)*/ 1868,
+	/*log(1.57837)*/ 1869,
+	/*log(1.57861)*/ 1870,
+	/*log(1.57886)*/ 1870,
+	/*log(1.57910)*/ 1871,
+	/*log(1.57935)*/ 1871,
+	/*log(1.57959)*/ 1872,
+	/*log(1.57983)*/ 1873,
+	/*log(1.58008)*/ 1873,
+	/*log(1.58032)*/ 1874,
+	/*log(1.58057)*/ 1875,
+	/*log(1.58081)*/ 1875,
+	/*log(1.58105)*/ 1876,
+	/*log(1.58130)*/ 1876,
+	/*log(1.58154)*/ 1877,
+	/*log(1.58179)*/ 1878,
+	/*log(1.58203)*/ 1878,
+	/*log(1.58228)*/ 1879,
+	/*log(1.58252)*/ 1880,
+	/*log(1.58276)*/ 1880,
+	/*log(1.58301)*/ 1881,
+	/*log(1.58325)*/ 1882,
+	/*log(1.58350)*/ 1882,
+	/*log(1.58374)*/ 1883,
+	/*log(1.58398)*/ 1883,
+	/*log(1.58423)*/ 1884,
+	/*log(1.58447)*/ 1885,
+	/*log(1.58472)*/ 1885,
+	/*log(1.58496)*/ 1886,
+	/*log(1.58521)*/ 1887,
+	/*log(1.58545)*/ 1887,
+	/*log(1.58569)*/ 1888,
+	/*log(1.58594)*/ 1888,
+	/*log(1.58618)*/ 1889,
+	/*log(1.58643)*/ 1890,
+	/*log(1.58667)*/ 1890,
+	/*log(1.58691)*/ 1891,
+	/*log(1.58716)*/ 1892,
+	/*log(1.58740)*/ 1892,
+	/*log(1.58765)*/ 1893,
+	/*log(1.58789)*/ 1894,
+	/*log(1.58813)*/ 1894,
+	/*log(1.58838)*/ 1895,
+	/*log(1.58862)*/ 1895,
+	/*log(1.58887)*/ 1896,
+	/*log(1.58911)*/ 1897,
+	/*log(1.58936)*/ 1897,
+	/*log(1.58960)*/ 1898,
+	/*log(1.58984)*/ 1899,
+	/*log(1.59009)*/ 1899,
+	/*log(1.59033)*/ 1900,
+	/*log(1.59058)*/ 1900,
+	/*log(1.59082)*/ 1901,
+	/*log(1.59106)*/ 1902,
+	/*log(1.59131)*/ 1902,
+	/*log(1.59155)*/ 1903,
+	/*log(1.59180)*/ 1904,
+	/*log(1.59204)*/ 1904,
+	/*log(1.59229)*/ 1905,
+	/*log(1.59253)*/ 1905,
+	/*log(1.59277)*/ 1906,
+	/*log(1.59302)*/ 1907,
+	/*log(1.59326)*/ 1907,
+	/*log(1.59351)*/ 1908,
+	/*log(1.59375)*/ 1909,
+	/*log(1.59399)*/ 1909,
+	/*log(1.59424)*/ 1910,
+	/*log(1.59448)*/ 1910,
+	/*log(1.59473)*/ 1911,
+	/*log(1.59497)*/ 1912,
+	/*log(1.59521)*/ 1912,
+	/*log(1.59546)*/ 1913,
+	/*log(1.59570)*/ 1914,
+	/*log(1.59595)*/ 1914,
+	/*log(1.59619)*/ 1915,
+	/*log(1.59644)*/ 1915,
+	/*log(1.59668)*/ 1916,
+	/*log(1.59692)*/ 1917,
+	/*log(1.59717)*/ 1917,
+	/*log(1.59741)*/ 1918,
+	/*log(1.59766)*/ 1919,
+	/*log(1.59790)*/ 1919,
+	/*log(1.59814)*/ 1920,
+	/*log(1.59839)*/ 1921,
+	/*log(1.59863)*/ 1921,
+	/*log(1.59888)*/ 1922,
+	/*log(1.59912)*/ 1922,
+	/*log(1.59937)*/ 1923,
+	/*log(1.59961)*/ 1924,
+	/*log(1.59985)*/ 1924,
+	/*log(1.60010)*/ 1925,
+	/*log(1.60034)*/ 1926,
+	/*log(1.60059)*/ 1926,
+	/*log(1.60083)*/ 1927,
+	/*log(1.60107)*/ 1927,
+	/*log(1.60132)*/ 1928,
+	/*log(1.60156)*/ 1929,
+	/*log(1.60181)*/ 1929,
+	/*log(1.60205)*/ 1930,
+	/*log(1.60229)*/ 1931,
+	/*log(1.60254)*/ 1931,
+	/*log(1.60278)*/ 1932,
+	/*log(1.60303)*/ 1932,
+	/*log(1.60327)*/ 1933,
+	/*log(1.60352)*/ 1934,
+	/*log(1.60376)*/ 1934,
+	/*log(1.60400)*/ 1935,
+	/*log(1.60425)*/ 1935,
+	/*log(1.60449)*/ 1936,
+	/*log(1.60474)*/ 1937,
+	/*log(1.60498)*/ 1937,
+	/*log(1.60522)*/ 1938,
+	/*log(1.60547)*/ 1939,
+	/*log(1.60571)*/ 1939,
+	/*log(1.60596)*/ 1940,
+	/*log(1.60620)*/ 1940,
+	/*log(1.60645)*/ 1941,
+	/*log(1.60669)*/ 1942,
+	/*log(1.60693)*/ 1942,
+	/*log(1.60718)*/ 1943,
+	/*log(1.60742)*/ 1944,
+	/*log(1.60767)*/ 1944,
+	/*log(1.60791)*/ 1945,
+	/*log(1.60815)*/ 1945,
+	/*log(1.60840)*/ 1946,
+	/*log(1.60864)*/ 1947,
+	/*log(1.60889)*/ 1947,
+	/*log(1.60913)*/ 1948,
+	/*log(1.60938)*/ 1949,
+	/*log(1.60962)*/ 1949,
+	/*log(1.60986)*/ 1950,
+	/*log(1.61011)*/ 1950,
+	/*log(1.61035)*/ 1951,
+	/*log(1.61060)*/ 1952,
+	/*log(1.61084)*/ 1952,
+	/*log(1.61108)*/ 1953,
+	/*log(1.61133)*/ 1954,
+	/*log(1.61157)*/ 1954,
+	/*log(1.61182)*/ 1955,
+	/*log(1.61206)*/ 1955,
+	/*log(1.61230)*/ 1956,
+	/*log(1.61255)*/ 1957,
+	/*log(1.61279)*/ 1957,
+	/*log(1.61304)*/ 1958,
+	/*log(1.61328)*/ 1958,
+	/*log(1.61353)*/ 1959,
+	/*log(1.61377)*/ 1960,
+	/*log(1.61401)*/ 1960,
+	/*log(1.61426)*/ 1961,
+	/*log(1.61450)*/ 1962,
+	/*log(1.61475)*/ 1962,
+	/*log(1.61499)*/ 1963,
+	/*log(1.61523)*/ 1963,
+	/*log(1.61548)*/ 1964,
+	/*log(1.61572)*/ 1965,
+	/*log(1.61597)*/ 1965,
+	/*log(1.61621)*/ 1966,
+	/*log(1.61646)*/ 1967,
+	/*log(1.61670)*/ 1967,
+	/*log(1.61694)*/ 1968,
+	/*log(1.61719)*/ 1968,
+	/*log(1.61743)*/ 1969,
+	/*log(1.61768)*/ 1970,
+	/*log(1.61792)*/ 1970,
+	/*log(1.61816)*/ 1971,
+	/*log(1.61841)*/ 1971,
+	/*log(1.61865)*/ 1972,
+	/*log(1.61890)*/ 1973,
+	/*log(1.61914)*/ 1973,
+	/*log(1.61938)*/ 1974,
+	/*log(1.61963)*/ 1975,
+	/*log(1.61987)*/ 1975,
+	/*log(1.62012)*/ 1976,
+	/*log(1.62036)*/ 1976,
+	/*log(1.62061)*/ 1977,
+	/*log(1.62085)*/ 1978,
+	/*log(1.62109)*/ 1978,
+	/*log(1.62134)*/ 1979,
+	/*log(1.62158)*/ 1980,
+	/*log(1.62183)*/ 1980,
+	/*log(1.62207)*/ 1981,
+	/*log(1.62231)*/ 1981,
+	/*log(1.62256)*/ 1982,
+	/*log(1.62280)*/ 1983,
+	/*log(1.62305)*/ 1983,
+	/*log(1.62329)*/ 1984,
+	/*log(1.62354)*/ 1984,
+	/*log(1.62378)*/ 1985,
+	/*log(1.62402)*/ 1986,
+	/*log(1.62427)*/ 1986,
+	/*log(1.62451)*/ 1987,
+	/*log(1.62476)*/ 1988,
+	/*log(1.62500)*/ 1988,
+	/*log(1.62524)*/ 1989,
+	/*log(1.62549)*/ 1989,
+	/*log(1.62573)*/ 1990,
+	/*log(1.62598)*/ 1991,
+	/*log(1.62622)*/ 1991,
+	/*log(1.62646)*/ 1992,
+	/*log(1.62671)*/ 1992,
+	/*log(1.62695)*/ 1993,
+	/*log(1.62720)*/ 1994,
+	/*log(1.62744)*/ 1994,
+	/*log(1.62769)*/ 1995,
+	/*log(1.62793)*/ 1996,
+	/*log(1.62817)*/ 1996,
+	/*log(1.62842)*/ 1997,
+	/*log(1.62866)*/ 1997,
+	/*log(1.62891)*/ 1998,
+	/*log(1.62915)*/ 1999,
+	/*log(1.62939)*/ 1999,
+	/*log(1.62964)*/ 2000,
+	/*log(1.62988)*/ 2000,
+	/*log(1.63013)*/ 2001,
+	/*log(1.63037)*/ 2002,
+	/*log(1.63062)*/ 2002,
+	/*log(1.63086)*/ 2003,
+	/*log(1.63110)*/ 2003,
+	/*log(1.63135)*/ 2004,
+	/*log(1.63159)*/ 2005,
+	/*log(1.63184)*/ 2005,
+	/*log(1.63208)*/ 2006,
+	/*log(1.63232)*/ 2007,
+	/*log(1.63257)*/ 2007,
+	/*log(1.63281)*/ 2008,
+	/*log(1.63306)*/ 2008,
+	/*log(1.63330)*/ 2009,
+	/*log(1.63354)*/ 2010,
+	/*log(1.63379)*/ 2010,
+	/*log(1.63403)*/ 2011,
+	/*log(1.63428)*/ 2011,
+	/*log(1.63452)*/ 2012,
+	/*log(1.63477)*/ 2013,
+	/*log(1.63501)*/ 2013,
+	/*log(1.63525)*/ 2014,
+	/*log(1.63550)*/ 2015,
+	/*log(1.63574)*/ 2015,
+	/*log(1.63599)*/ 2016,
+	/*log(1.63623)*/ 2016,
+	/*log(1.63647)*/ 2017,
+	/*log(1.63672)*/ 2018,
+	/*log(1.63696)*/ 2018,
+	/*log(1.63721)*/ 2019,
+	/*log(1.63745)*/ 2019,
+	/*log(1.63770)*/ 2020,
+	/*log(1.63794)*/ 2021,
+	/*log(1.63818)*/ 2021,
+	/*log(1.63843)*/ 2022,
+	/*log(1.63867)*/ 2022,
+	/*log(1.63892)*/ 2023,
+	/*log(1.63916)*/ 2024,
+	/*log(1.63940)*/ 2024,
+	/*log(1.63965)*/ 2025,
+	/*log(1.63989)*/ 2026,
+	/*log(1.64014)*/ 2026,
+	/*log(1.64038)*/ 2027,
+	/*log(1.64062)*/ 2027,
+	/*log(1.64087)*/ 2028,
+	/*log(1.64111)*/ 2029,
+	/*log(1.64136)*/ 2029,
+	/*log(1.64160)*/ 2030,
+	/*log(1.64185)*/ 2030,
+	/*log(1.64209)*/ 2031,
+	/*log(1.64233)*/ 2032,
+	/*log(1.64258)*/ 2032,
+	/*log(1.64282)*/ 2033,
+	/*log(1.64307)*/ 2033,
+	/*log(1.64331)*/ 2034,
+	/*log(1.64355)*/ 2035,
+	/*log(1.64380)*/ 2035,
+	/*log(1.64404)*/ 2036,
+	/*log(1.64429)*/ 2036,
+	/*log(1.64453)*/ 2037,
+	/*log(1.64478)*/ 2038,
+	/*log(1.64502)*/ 2038,
+	/*log(1.64526)*/ 2039,
+	/*log(1.64551)*/ 2040,
+	/*log(1.64575)*/ 2040,
+	/*log(1.64600)*/ 2041,
+	/*log(1.64624)*/ 2041,
+	/*log(1.64648)*/ 2042,
+	/*log(1.64673)*/ 2043,
+	/*log(1.64697)*/ 2043,
+	/*log(1.64722)*/ 2044,
+	/*log(1.64746)*/ 2044,
+	/*log(1.64771)*/ 2045,
+	/*log(1.64795)*/ 2046,
+	/*log(1.64819)*/ 2046,
+	/*log(1.64844)*/ 2047,
+	/*log(1.64868)*/ 2047,
+	/*log(1.64893)*/ 2048,
+	/*log(1.64917)*/ 2049,
+	/*log(1.64941)*/ 2049,
+	/*log(1.64966)*/ 2050,
+	/*log(1.64990)*/ 2050,
+	/*log(1.65015)*/ 2051,
+	/*log(1.65039)*/ 2052,
+	/*log(1.65063)*/ 2052,
+	/*log(1.65088)*/ 2053,
+	/*log(1.65112)*/ 2053,
+	/*log(1.65137)*/ 2054,
+	/*log(1.65161)*/ 2055,
+	/*log(1.65186)*/ 2055,
+	/*log(1.65210)*/ 2056,
+	/*log(1.65234)*/ 2056,
+	/*log(1.65259)*/ 2057,
+	/*log(1.65283)*/ 2058,
+	/*log(1.65308)*/ 2058,
+	/*log(1.65332)*/ 2059,
+	/*log(1.65356)*/ 2060,
+	/*log(1.65381)*/ 2060,
+	/*log(1.65405)*/ 2061,
+	/*log(1.65430)*/ 2061,
+	/*log(1.65454)*/ 2062,
+	/*log(1.65479)*/ 2063,
+	/*log(1.65503)*/ 2063,
+	/*log(1.65527)*/ 2064,
+	/*log(1.65552)*/ 2064,
+	/*log(1.65576)*/ 2065,
+	/*log(1.65601)*/ 2066,
+	/*log(1.65625)*/ 2066,
+	/*log(1.65649)*/ 2067,
+	/*log(1.65674)*/ 2067,
+	/*log(1.65698)*/ 2068,
+	/*log(1.65723)*/ 2069,
+	/*log(1.65747)*/ 2069,
+	/*log(1.65771)*/ 2070,
+	/*log(1.65796)*/ 2070,
+	/*log(1.65820)*/ 2071,
+	/*log(1.65845)*/ 2072,
+	/*log(1.65869)*/ 2072,
+	/*log(1.65894)*/ 2073,
+	/*log(1.65918)*/ 2073,
+	/*log(1.65942)*/ 2074,
+	/*log(1.65967)*/ 2075,
+	/*log(1.65991)*/ 2075,
+	/*log(1.66016)*/ 2076,
+	/*log(1.66040)*/ 2076,
+	/*log(1.66064)*/ 2077,
+	/*log(1.66089)*/ 2078,
+	/*log(1.66113)*/ 2078,
+	/*log(1.66138)*/ 2079,
+	/*log(1.66162)*/ 2079,
+	/*log(1.66187)*/ 2080,
+	/*log(1.66211)*/ 2081,
+	/*log(1.66235)*/ 2081,
+	/*log(1.66260)*/ 2082,
+	/*log(1.66284)*/ 2082,
+	/*log(1.66309)*/ 2083,
+	/*log(1.66333)*/ 2084,
+	/*log(1.66357)*/ 2084,
+	/*log(1.66382)*/ 2085,
+	/*log(1.66406)*/ 2085,
+	/*log(1.66431)*/ 2086,
+	/*log(1.66455)*/ 2087,
+	/*log(1.66479)*/ 2087,
+	/*log(1.66504)*/ 2088,
+	/*log(1.66528)*/ 2088,
+	/*log(1.66553)*/ 2089,
+	/*log(1.66577)*/ 2090,
+	/*log(1.66602)*/ 2090,
+	/*log(1.66626)*/ 2091,
+	/*log(1.66650)*/ 2091,
+	/*log(1.66675)*/ 2092,
+	/*log(1.66699)*/ 2093,
+	/*log(1.66724)*/ 2093,
+	/*log(1.66748)*/ 2094,
+	/*log(1.66772)*/ 2094,
+	/*log(1.66797)*/ 2095,
+	/*log(1.66821)*/ 2096,
+	/*log(1.66846)*/ 2096,
+	/*log(1.66870)*/ 2097,
+	/*log(1.66895)*/ 2097,
+	/*log(1.66919)*/ 2098,
+	/*log(1.66943)*/ 2099,
+	/*log(1.66968)*/ 2099,
+	/*log(1.66992)*/ 2100,
+	/*log(1.67017)*/ 2100,
+	/*log(1.67041)*/ 2101,
+	/*log(1.67065)*/ 2102,
+	/*log(1.67090)*/ 2102,
+	/*log(1.67114)*/ 2103,
+	/*log(1.67139)*/ 2103,
+	/*log(1.67163)*/ 2104,
+	/*log(1.67188)*/ 2105,
+	/*log(1.67212)*/ 2105,
+	/*log(1.67236)*/ 2106,
+	/*log(1.67261)*/ 2106,
+	/*log(1.67285)*/ 2107,
+	/*log(1.67310)*/ 2108,
+	/*log(1.67334)*/ 2108,
+	/*log(1.67358)*/ 2109,
+	/*log(1.67383)*/ 2109,
+	/*log(1.67407)*/ 2110,
+	/*log(1.67432)*/ 2111,
+	/*log(1.67456)*/ 2111,
+	/*log(1.67480)*/ 2112,
+	/*log(1.67505)*/ 2112,
+	/*log(1.67529)*/ 2113,
+	/*log(1.67554)*/ 2114,
+	/*log(1.67578)*/ 2114,
+	/*log(1.67603)*/ 2115,
+	/*log(1.67627)*/ 2115,
+	/*log(1.67651)*/ 2116,
+	/*log(1.67676)*/ 2117,
+	/*log(1.67700)*/ 2117,
+	/*log(1.67725)*/ 2118,
+	/*log(1.67749)*/ 2118,
+	/*log(1.67773)*/ 2119,
+	/*log(1.67798)*/ 2120,
+	/*log(1.67822)*/ 2120,
+	/*log(1.67847)*/ 2121,
+	/*log(1.67871)*/ 2121,
+	/*log(1.67896)*/ 2122,
+	/*log(1.67920)*/ 2123,
+	/*log(1.67944)*/ 2123,
+	/*log(1.67969)*/ 2124,
+	/*log(1.67993)*/ 2124,
+	/*log(1.68018)*/ 2125,
+	/*log(1.68042)*/ 2126,
+	/*log(1.68066)*/ 2126,
+	/*log(1.68091)*/ 2127,
+	/*log(1.68115)*/ 2127,
+	/*log(1.68140)*/ 2128,
+	/*log(1.68164)*/ 2128,
+	/*log(1.68188)*/ 2129,
+	/*log(1.68213)*/ 2130,
+	/*log(1.68237)*/ 2130,
+	/*log(1.68262)*/ 2131,
+	/*log(1.68286)*/ 2131,
+	/*log(1.68311)*/ 2132,
+	/*log(1.68335)*/ 2133,
+	/*log(1.68359)*/ 2133,
+	/*log(1.68384)*/ 2134,
+	/*log(1.68408)*/ 2134,
+	/*log(1.68433)*/ 2135,
+	/*log(1.68457)*/ 2136,
+	/*log(1.68481)*/ 2136,
+	/*log(1.68506)*/ 2137,
+	/*log(1.68530)*/ 2137,
+	/*log(1.68555)*/ 2138,
+	/*log(1.68579)*/ 2139,
+	/*log(1.68604)*/ 2139,
+	/*log(1.68628)*/ 2140,
+	/*log(1.68652)*/ 2140,
+	/*log(1.68677)*/ 2141,
+	/*log(1.68701)*/ 2142,
+	/*log(1.68726)*/ 2142,
+	/*log(1.68750)*/ 2143,
+	/*log(1.68774)*/ 2143,
+	/*log(1.68799)*/ 2144,
+	/*log(1.68823)*/ 2145,
+	/*log(1.68848)*/ 2145,
+	/*log(1.68872)*/ 2146,
+	/*log(1.68896)*/ 2146,
+	/*log(1.68921)*/ 2147,
+	/*log(1.68945)*/ 2147,
+	/*log(1.68970)*/ 2148,
+	/*log(1.68994)*/ 2149,
+	/*log(1.69019)*/ 2149,
+	/*log(1.69043)*/ 2150,
+	/*log(1.69067)*/ 2150,
+	/*log(1.69092)*/ 2151,
+	/*log(1.69116)*/ 2152,
+	/*log(1.69141)*/ 2152,
+	/*log(1.69165)*/ 2153,
+	/*log(1.69189)*/ 2153,
+	/*log(1.69214)*/ 2154,
+	/*log(1.69238)*/ 2155,
+	/*log(1.69263)*/ 2155,
+	/*log(1.69287)*/ 2156,
+	/*log(1.69312)*/ 2156,
+	/*log(1.69336)*/ 2157,
+	/*log(1.69360)*/ 2158,
+	/*log(1.69385)*/ 2158,
+	/*log(1.69409)*/ 2159,
+	/*log(1.69434)*/ 2159,
+	/*log(1.69458)*/ 2160,
+	/*log(1.69482)*/ 2160,
+	/*log(1.69507)*/ 2161,
+	/*log(1.69531)*/ 2162,
+	/*log(1.69556)*/ 2162,
+	/*log(1.69580)*/ 2163,
+	/*log(1.69604)*/ 2163,
+	/*log(1.69629)*/ 2164,
+	/*log(1.69653)*/ 2165,
+	/*log(1.69678)*/ 2165,
+	/*log(1.69702)*/ 2166,
+	/*log(1.69727)*/ 2166,
+	/*log(1.69751)*/ 2167,
+	/*log(1.69775)*/ 2168,
+	/*log(1.69800)*/ 2168,
+	/*log(1.69824)*/ 2169,
+	/*log(1.69849)*/ 2169,
+	/*log(1.69873)*/ 2170,
+	/*log(1.69897)*/ 2170,
+	/*log(1.69922)*/ 2171,
+	/*log(1.69946)*/ 2172,
+	/*log(1.69971)*/ 2172,
+	/*log(1.69995)*/ 2173,
+	/*log(1.70020)*/ 2173,
+	/*log(1.70044)*/ 2174,
+	/*log(1.70068)*/ 2175,
+	/*log(1.70093)*/ 2175,
+	/*log(1.70117)*/ 2176,
+	/*log(1.70142)*/ 2176,
+	/*log(1.70166)*/ 2177,
+	/*log(1.70190)*/ 2178,
+	/*log(1.70215)*/ 2178,
+	/*log(1.70239)*/ 2179,
+	/*log(1.70264)*/ 2179,
+	/*log(1.70288)*/ 2180,
+	/*log(1.70312)*/ 2180,
+	/*log(1.70337)*/ 2181,
+	/*log(1.70361)*/ 2182,
+	/*log(1.70386)*/ 2182,
+	/*log(1.70410)*/ 2183,
+	/*log(1.70435)*/ 2183,
+	/*log(1.70459)*/ 2184,
+	/*log(1.70483)*/ 2185,
+	/*log(1.70508)*/ 2185,
+	/*log(1.70532)*/ 2186,
+	/*log(1.70557)*/ 2186,
+	/*log(1.70581)*/ 2187,
+	/*log(1.70605)*/ 2188,
+	/*log(1.70630)*/ 2188,
+	/*log(1.70654)*/ 2189,
+	/*log(1.70679)*/ 2189,
+	/*log(1.70703)*/ 2190,
+	/*log(1.70728)*/ 2190,
+	/*log(1.70752)*/ 2191,
+	/*log(1.70776)*/ 2192,
+	/*log(1.70801)*/ 2192,
+	/*log(1.70825)*/ 2193,
+	/*log(1.70850)*/ 2193,
+	/*log(1.70874)*/ 2194,
+	/*log(1.70898)*/ 2195,
+	/*log(1.70923)*/ 2195,
+	/*log(1.70947)*/ 2196,
+	/*log(1.70972)*/ 2196,
+	/*log(1.70996)*/ 2197,
+	/*log(1.71021)*/ 2197,
+	/*log(1.71045)*/ 2198,
+	/*log(1.71069)*/ 2199,
+	/*log(1.71094)*/ 2199,
+	/*log(1.71118)*/ 2200,
+	/*log(1.71143)*/ 2200,
+	/*log(1.71167)*/ 2201,
+	/*log(1.71191)*/ 2202,
+	/*log(1.71216)*/ 2202,
+	/*log(1.71240)*/ 2203,
+	/*log(1.71265)*/ 2203,
+	/*log(1.71289)*/ 2204,
+	/*log(1.71313)*/ 2204,
+	/*log(1.71338)*/ 2205,
+	/*log(1.71362)*/ 2206,
+	/*log(1.71387)*/ 2206,
+	/*log(1.71411)*/ 2207,
+	/*log(1.71436)*/ 2207,
+	/*log(1.71460)*/ 2208,
+	/*log(1.71484)*/ 2209,
+	/*log(1.71509)*/ 2209,
+	/*log(1.71533)*/ 2210,
+	/*log(1.71558)*/ 2210,
+	/*log(1.71582)*/ 2211,
+	/*log(1.71606)*/ 2211,
+	/*log(1.71631)*/ 2212,
+	/*log(1.71655)*/ 2213,
+	/*log(1.71680)*/ 2213,
+	/*log(1.71704)*/ 2214,
+	/*log(1.71729)*/ 2214,
+	/*log(1.71753)*/ 2215,
+	/*log(1.71777)*/ 2216,
+	/*log(1.71802)*/ 2216,
+	/*log(1.71826)*/ 2217,
+	/*log(1.71851)*/ 2217,
+	/*log(1.71875)*/ 2218,
+	/*log(1.71899)*/ 2218,
+	/*log(1.71924)*/ 2219,
+	/*log(1.71948)*/ 2220,
+	/*log(1.71973)*/ 2220,
+	/*log(1.71997)*/ 2221,
+	/*log(1.72021)*/ 2221,
+	/*log(1.72046)*/ 2222,
+	/*log(1.72070)*/ 2223,
+	/*log(1.72095)*/ 2223,
+	/*log(1.72119)*/ 2224,
+	/*log(1.72144)*/ 2224,
+	/*log(1.72168)*/ 2225,
+	/*log(1.72192)*/ 2225,
+	/*log(1.72217)*/ 2226,
+	/*log(1.72241)*/ 2227,
+	/*log(1.72266)*/ 2227,
+	/*log(1.72290)*/ 2228,
+	/*log(1.72314)*/ 2228,
+	/*log(1.72339)*/ 2229,
+	/*log(1.72363)*/ 2230,
+	/*log(1.72388)*/ 2230,
+	/*log(1.72412)*/ 2231,
+	/*log(1.72437)*/ 2231,
+	/*log(1.72461)*/ 2232,
+	/*log(1.72485)*/ 2232,
+	/*log(1.72510)*/ 2233,
+	/*log(1.72534)*/ 2234,
+	/*log(1.72559)*/ 2234,
+	/*log(1.72583)*/ 2235,
+	/*log(1.72607)*/ 2235,
+	/*log(1.72632)*/ 2236,
+	/*log(1.72656)*/ 2236,
+	/*log(1.72681)*/ 2237,
+	/*log(1.72705)*/ 2238,
+	/*log(1.72729)*/ 2238,
+	/*log(1.72754)*/ 2239,
+	/*log(1.72778)*/ 2239,
+	/*log(1.72803)*/ 2240,
+	/*log(1.72827)*/ 2241,
+	/*log(1.72852)*/ 2241,
+	/*log(1.72876)*/ 2242,
+	/*log(1.72900)*/ 2242,
+	/*log(1.72925)*/ 2243,
+	/*log(1.72949)*/ 2243,
+	/*log(1.72974)*/ 2244,
+	/*log(1.72998)*/ 2245,
+	/*log(1.73022)*/ 2245,
+	/*log(1.73047)*/ 2246,
+	/*log(1.73071)*/ 2246,
+	/*log(1.73096)*/ 2247,
+	/*log(1.73120)*/ 2247,
+	/*log(1.73145)*/ 2248,
+	/*log(1.73169)*/ 2249,
+	/*log(1.73193)*/ 2249,
+	/*log(1.73218)*/ 2250,
+	/*log(1.73242)*/ 2250,
+	/*log(1.73267)*/ 2251,
+	/*log(1.73291)*/ 2251,
+	/*log(1.73315)*/ 2252,
+	/*log(1.73340)*/ 2253,
+	/*log(1.73364)*/ 2253,
+	/*log(1.73389)*/ 2254,
+	/*log(1.73413)*/ 2254,
+	/*log(1.73438)*/ 2255,
+	/*log(1.73462)*/ 2256,
+	/*log(1.73486)*/ 2256,
+	/*log(1.73511)*/ 2257,
+	/*log(1.73535)*/ 2257,
+	/*log(1.73560)*/ 2258,
+	/*log(1.73584)*/ 2258,
+	/*log(1.73608)*/ 2259,
+	/*log(1.73633)*/ 2260,
+	/*log(1.73657)*/ 2260,
+	/*log(1.73682)*/ 2261,
+	/*log(1.73706)*/ 2261,
+	/*log(1.73730)*/ 2262,
+	/*log(1.73755)*/ 2262,
+	/*log(1.73779)*/ 2263,
+	/*log(1.73804)*/ 2264,
+	/*log(1.73828)*/ 2264,
+	/*log(1.73853)*/ 2265,
+	/*log(1.73877)*/ 2265,
+	/*log(1.73901)*/ 2266,
+	/*log(1.73926)*/ 2266,
+	/*log(1.73950)*/ 2267,
+	/*log(1.73975)*/ 2268,
+	/*log(1.73999)*/ 2268,
+	/*log(1.74023)*/ 2269,
+	/*log(1.74048)*/ 2269,
+	/*log(1.74072)*/ 2270,
+	/*log(1.74097)*/ 2270,
+	/*log(1.74121)*/ 2271,
+	/*log(1.74146)*/ 2272,
+	/*log(1.74170)*/ 2272,
+	/*log(1.74194)*/ 2273,
+	/*log(1.74219)*/ 2273,
+	/*log(1.74243)*/ 2274,
+	/*log(1.74268)*/ 2275,
+	/*log(1.74292)*/ 2275,
+	/*log(1.74316)*/ 2276,
+	/*log(1.74341)*/ 2276,
+	/*log(1.74365)*/ 2277,
+	/*log(1.74390)*/ 2277,
+	/*log(1.74414)*/ 2278,
+	/*log(1.74438)*/ 2279,
+	/*log(1.74463)*/ 2279,
+	/*log(1.74487)*/ 2280,
+	/*log(1.74512)*/ 2280,
+	/*log(1.74536)*/ 2281,
+	/*log(1.74561)*/ 2281,
+	/*log(1.74585)*/ 2282,
+	/*log(1.74609)*/ 2283,
+	/*log(1.74634)*/ 2283,
+	/*log(1.74658)*/ 2284,
+	/*log(1.74683)*/ 2284,
+	/*log(1.74707)*/ 2285,
+	/*log(1.74731)*/ 2285,
+	/*log(1.74756)*/ 2286,
+	/*log(1.74780)*/ 2287,
+	/*log(1.74805)*/ 2287,
+	/*log(1.74829)*/ 2288,
+	/*log(1.74854)*/ 2288,
+	/*log(1.74878)*/ 2289,
+	/*log(1.74902)*/ 2289,
+	/*log(1.74927)*/ 2290,
+	/*log(1.74951)*/ 2291,
+	/*log(1.74976)*/ 2291,
+	/*log(1.75000)*/ 2292,
+	/*log(1.75024)*/ 2292,
+	/*log(1.75049)*/ 2293,
+	/*log(1.75073)*/ 2293,
+	/*log(1.75098)*/ 2294,
+	/*log(1.75122)*/ 2295,
+	/*log(1.75146)*/ 2295,
+	/*log(1.75171)*/ 2296,
+	/*log(1.75195)*/ 2296,
+	/*log(1.75220)*/ 2297,
+	/*log(1.75244)*/ 2297,
+	/*log(1.75269)*/ 2298,
+	/*log(1.75293)*/ 2299,
+	/*log(1.75317)*/ 2299,
+	/*log(1.75342)*/ 2300,
+	/*log(1.75366)*/ 2300,
+	/*log(1.75391)*/ 2301,
+	/*log(1.75415)*/ 2301,
+	/*log(1.75439)*/ 2302,
+	/*log(1.75464)*/ 2303,
+	/*log(1.75488)*/ 2303,
+	/*log(1.75513)*/ 2304,
+	/*log(1.75537)*/ 2304,
+	/*log(1.75562)*/ 2305,
+	/*log(1.75586)*/ 2305,
+	/*log(1.75610)*/ 2306,
+	/*log(1.75635)*/ 2307,
+	/*log(1.75659)*/ 2307,
+	/*log(1.75684)*/ 2308,
+	/*log(1.75708)*/ 2308,
+	/*log(1.75732)*/ 2309,
+	/*log(1.75757)*/ 2309,
+	/*log(1.75781)*/ 2310,
+	/*log(1.75806)*/ 2311,
+	/*log(1.75830)*/ 2311,
+	/*log(1.75854)*/ 2312,
+	/*log(1.75879)*/ 2312,
+	/*log(1.75903)*/ 2313,
+	/*log(1.75928)*/ 2313,
+	/*log(1.75952)*/ 2314,
+	/*log(1.75977)*/ 2314,
+	/*log(1.76001)*/ 2315,
+	/*log(1.76025)*/ 2316,
+	/*log(1.76050)*/ 2316,
+	/*log(1.76074)*/ 2317,
+	/*log(1.76099)*/ 2317,
+	/*log(1.76123)*/ 2318,
+	/*log(1.76147)*/ 2318,
+	/*log(1.76172)*/ 2319,
+	/*log(1.76196)*/ 2320,
+	/*log(1.76221)*/ 2320,
+	/*log(1.76245)*/ 2321,
+	/*log(1.76270)*/ 2321,
+	/*log(1.76294)*/ 2322,
+	/*log(1.76318)*/ 2322,
+	/*log(1.76343)*/ 2323,
+	/*log(1.76367)*/ 2324,
+	/*log(1.76392)*/ 2324,
+	/*log(1.76416)*/ 2325,
+	/*log(1.76440)*/ 2325,
+	/*log(1.76465)*/ 2326,
+	/*log(1.76489)*/ 2326,
+	/*log(1.76514)*/ 2327,
+	/*log(1.76538)*/ 2328,
+	/*log(1.76562)*/ 2328,
+	/*log(1.76587)*/ 2329,
+	/*log(1.76611)*/ 2329,
+	/*log(1.76636)*/ 2330,
+	/*log(1.76660)*/ 2330,
+	/*log(1.76685)*/ 2331,
+	/*log(1.76709)*/ 2331,
+	/*log(1.76733)*/ 2332,
+	/*log(1.76758)*/ 2333,
+	/*log(1.76782)*/ 2333,
+	/*log(1.76807)*/ 2334,
+	/*log(1.76831)*/ 2334,
+	/*log(1.76855)*/ 2335,
+	/*log(1.76880)*/ 2335,
+	/*log(1.76904)*/ 2336,
+	/*log(1.76929)*/ 2337,
+	/*log(1.76953)*/ 2337,
+	/*log(1.76978)*/ 2338,
+	/*log(1.77002)*/ 2338,
+	/*log(1.77026)*/ 2339,
+	/*log(1.77051)*/ 2339,
+	/*log(1.77075)*/ 2340,
+	/*log(1.77100)*/ 2341,
+	/*log(1.77124)*/ 2341,
+	/*log(1.77148)*/ 2342,
+	/*log(1.77173)*/ 2342,
+	/*log(1.77197)*/ 2343,
+	/*log(1.77222)*/ 2343,
+	/*log(1.77246)*/ 2344,
+	/*log(1.77271)*/ 2344,
+	/*log(1.77295)*/ 2345,
+	/*log(1.77319)*/ 2346,
+	/*log(1.77344)*/ 2346,
+	/*log(1.77368)*/ 2347,
+	/*log(1.77393)*/ 2347,
+	/*log(1.77417)*/ 2348,
+	/*log(1.77441)*/ 2348,
+	/*log(1.77466)*/ 2349,
+	/*log(1.77490)*/ 2350,
+	/*log(1.77515)*/ 2350,
+	/*log(1.77539)*/ 2351,
+	/*log(1.77563)*/ 2351,
+	/*log(1.77588)*/ 2352,
+	/*log(1.77612)*/ 2352,
+	/*log(1.77637)*/ 2353,
+	/*log(1.77661)*/ 2354,
+	/*log(1.77686)*/ 2354,
+	/*log(1.77710)*/ 2355,
+	/*log(1.77734)*/ 2355,
+	/*log(1.77759)*/ 2356,
+	/*log(1.77783)*/ 2356,
+	/*log(1.77808)*/ 2357,
+	/*log(1.77832)*/ 2357,
+	/*log(1.77856)*/ 2358,
+	/*log(1.77881)*/ 2359,
+	/*log(1.77905)*/ 2359,
+	/*log(1.77930)*/ 2360,
+	/*log(1.77954)*/ 2360,
+	/*log(1.77979)*/ 2361,
+	/*log(1.78003)*/ 2361,
+	/*log(1.78027)*/ 2362,
+	/*log(1.78052)*/ 2362,
+	/*log(1.78076)*/ 2363,
+	/*log(1.78101)*/ 2364,
+	/*log(1.78125)*/ 2364,
+	/*log(1.78149)*/ 2365,
+	/*log(1.78174)*/ 2365,
+	/*log(1.78198)*/ 2366,
+	/*log(1.78223)*/ 2366,
+	/*log(1.78247)*/ 2367,
+	/*log(1.78271)*/ 2368,
+	/*log(1.78296)*/ 2368,
+	/*log(1.78320)*/ 2369,
+	/*log(1.78345)*/ 2369,
+	/*log(1.78369)*/ 2370,
+	/*log(1.78394)*/ 2370,
+	/*log(1.78418)*/ 2371,
+	/*log(1.78442)*/ 2371,
+	/*log(1.78467)*/ 2372,
+	/*log(1.78491)*/ 2373,
+	/*log(1.78516)*/ 2373,
+	/*log(1.78540)*/ 2374,
+	/*log(1.78564)*/ 2374,
+	/*log(1.78589)*/ 2375,
+	/*log(1.78613)*/ 2375,
+	/*log(1.78638)*/ 2376,
+	/*log(1.78662)*/ 2377,
+	/*log(1.78687)*/ 2377,
+	/*log(1.78711)*/ 2378,
+	/*log(1.78735)*/ 2378,
+	/*log(1.78760)*/ 2379,
+	/*log(1.78784)*/ 2379,
+	/*log(1.78809)*/ 2380,
+	/*log(1.78833)*/ 2380,
+	/*log(1.78857)*/ 2381,
+	/*log(1.78882)*/ 2382,
+	/*log(1.78906)*/ 2382,
+	/*log(1.78931)*/ 2383,
+	/*log(1.78955)*/ 2383,
+	/*log(1.78979)*/ 2384,
+	/*log(1.79004)*/ 2384,
+	/*log(1.79028)*/ 2385,
+	/*log(1.79053)*/ 2385,
+	/*log(1.79077)*/ 2386,
+	/*log(1.79102)*/ 2387,
+	/*log(1.79126)*/ 2387,
+	/*log(1.79150)*/ 2388,
+	/*log(1.79175)*/ 2388,
+	/*log(1.79199)*/ 2389,
+	/*log(1.79224)*/ 2389,
+	/*log(1.79248)*/ 2390,
+	/*log(1.79272)*/ 2390,
+	/*log(1.79297)*/ 2391,
+	/*log(1.79321)*/ 2392,
+	/*log(1.79346)*/ 2392,
+	/*log(1.79370)*/ 2393,
+	/*log(1.79395)*/ 2393,
+	/*log(1.79419)*/ 2394,
+	/*log(1.79443)*/ 2394,
+	/*log(1.79468)*/ 2395,
+	/*log(1.79492)*/ 2396,
+	/*log(1.79517)*/ 2396,
+	/*log(1.79541)*/ 2397,
+	/*log(1.79565)*/ 2397,
+	/*log(1.79590)*/ 2398,
+	/*log(1.79614)*/ 2398,
+	/*log(1.79639)*/ 2399,
+	/*log(1.79663)*/ 2399,
+	/*log(1.79688)*/ 2400,
+	/*log(1.79712)*/ 2401,
+	/*log(1.79736)*/ 2401,
+	/*log(1.79761)*/ 2402,
+	/*log(1.79785)*/ 2402,
+	/*log(1.79810)*/ 2403,
+	/*log(1.79834)*/ 2403,
+	/*log(1.79858)*/ 2404,
+	/*log(1.79883)*/ 2404,
+	/*log(1.79907)*/ 2405,
+	/*log(1.79932)*/ 2406,
+	/*log(1.79956)*/ 2406,
+	/*log(1.79980)*/ 2407,
+	/*log(1.80005)*/ 2407,
+	/*log(1.80029)*/ 2408,
+	/*log(1.80054)*/ 2408,
+	/*log(1.80078)*/ 2409,
+	/*log(1.80103)*/ 2409,
+	/*log(1.80127)*/ 2410,
+	/*log(1.80151)*/ 2411,
+	/*log(1.80176)*/ 2411,
+	/*log(1.80200)*/ 2412,
+	/*log(1.80225)*/ 2412,
+	/*log(1.80249)*/ 2413,
+	/*log(1.80273)*/ 2413,
+	/*log(1.80298)*/ 2414,
+	/*log(1.80322)*/ 2414,
+	/*log(1.80347)*/ 2415,
+	/*log(1.80371)*/ 2416,
+	/*log(1.80396)*/ 2416,
+	/*log(1.80420)*/ 2417,
+	/*log(1.80444)*/ 2417,
+	/*log(1.80469)*/ 2418,
+	/*log(1.80493)*/ 2418,
+	/*log(1.80518)*/ 2419,
+	/*log(1.80542)*/ 2419,
+	/*log(1.80566)*/ 2420,
+	/*log(1.80591)*/ 2420,
+	/*log(1.80615)*/ 2421,
+	/*log(1.80640)*/ 2422,
+	/*log(1.80664)*/ 2422,
+	/*log(1.80688)*/ 2423,
+	/*log(1.80713)*/ 2423,
+	/*log(1.80737)*/ 2424,
+	/*log(1.80762)*/ 2424,
+	/*log(1.80786)*/ 2425,
+	/*log(1.80811)*/ 2425,
+	/*log(1.80835)*/ 2426,
+	/*log(1.80859)*/ 2427,
+	/*log(1.80884)*/ 2427,
+	/*log(1.80908)*/ 2428,
+	/*log(1.80933)*/ 2428,
+	/*log(1.80957)*/ 2429,
+	/*log(1.80981)*/ 2429,
+	/*log(1.81006)*/ 2430,
+	/*log(1.81030)*/ 2430,
+	/*log(1.81055)*/ 2431,
+	/*log(1.81079)*/ 2432,
+	/*log(1.81104)*/ 2432,
+	/*log(1.81128)*/ 2433,
+	/*log(1.81152)*/ 2433,
+	/*log(1.81177)*/ 2434,
+	/*log(1.81201)*/ 2434,
+	/*log(1.81226)*/ 2435,
+	/*log(1.81250)*/ 2435,
+	/*log(1.81274)*/ 2436,
+	/*log(1.81299)*/ 2437,
+	/*log(1.81323)*/ 2437,
+	/*log(1.81348)*/ 2438,
+	/*log(1.81372)*/ 2438,
+	/*log(1.81396)*/ 2439,
+	/*log(1.81421)*/ 2439,
+	/*log(1.81445)*/ 2440,
+	/*log(1.81470)*/ 2440,
+	/*log(1.81494)*/ 2441,
+	/*log(1.81519)*/ 2441,
+	/*log(1.81543)*/ 2442,
+	/*log(1.81567)*/ 2443,
+	/*log(1.81592)*/ 2443,
+	/*log(1.81616)*/ 2444,
+	/*log(1.81641)*/ 2444,
+	/*log(1.81665)*/ 2445,
+	/*log(1.81689)*/ 2445,
+	/*log(1.81714)*/ 2446,
+	/*log(1.81738)*/ 2446,
+	/*log(1.81763)*/ 2447,
+	/*log(1.81787)*/ 2448,
+	/*log(1.81812)*/ 2448,
+	/*log(1.81836)*/ 2449,
+	/*log(1.81860)*/ 2449,
+	/*log(1.81885)*/ 2450,
+	/*log(1.81909)*/ 2450,
+	/*log(1.81934)*/ 2451,
+	/*log(1.81958)*/ 2451,
+	/*log(1.81982)*/ 2452,
+	/*log(1.82007)*/ 2452,
+	/*log(1.82031)*/ 2453,
+	/*log(1.82056)*/ 2454,
+	/*log(1.82080)*/ 2454,
+	/*log(1.82104)*/ 2455,
+	/*log(1.82129)*/ 2455,
+	/*log(1.82153)*/ 2456,
+	/*log(1.82178)*/ 2456,
+	/*log(1.82202)*/ 2457,
+	/*log(1.82227)*/ 2457,
+	/*log(1.82251)*/ 2458,
+	/*log(1.82275)*/ 2459,
+	/*log(1.82300)*/ 2459,
+	/*log(1.82324)*/ 2460,
+	/*log(1.82349)*/ 2460,
+	/*log(1.82373)*/ 2461,
+	/*log(1.82397)*/ 2461,
+	/*log(1.82422)*/ 2462,
+	/*log(1.82446)*/ 2462,
+	/*log(1.82471)*/ 2463,
+	/*log(1.82495)*/ 2463,
+	/*log(1.82520)*/ 2464,
+	/*log(1.82544)*/ 2465,
+	/*log(1.82568)*/ 2465,
+	/*log(1.82593)*/ 2466,
+	/*log(1.82617)*/ 2466,
+	/*log(1.82642)*/ 2467,
+	/*log(1.82666)*/ 2467,
+	/*log(1.82690)*/ 2468,
+	/*log(1.82715)*/ 2468,
+	/*log(1.82739)*/ 2469,
+	/*log(1.82764)*/ 2469,
+	/*log(1.82788)*/ 2470,
+	/*log(1.82812)*/ 2471,
+	/*log(1.82837)*/ 2471,
+	/*log(1.82861)*/ 2472,
+	/*log(1.82886)*/ 2472,
+	/*log(1.82910)*/ 2473,
+	/*log(1.82935)*/ 2473,
+	/*log(1.82959)*/ 2474,
+	/*log(1.82983)*/ 2474,
+	/*log(1.83008)*/ 2475,
+	/*log(1.83032)*/ 2475,
+	/*log(1.83057)*/ 2476,
+	/*log(1.83081)*/ 2477,
+	/*log(1.83105)*/ 2477,
+	/*log(1.83130)*/ 2478,
+	/*log(1.83154)*/ 2478,
+	/*log(1.83179)*/ 2479,
+	/*log(1.83203)*/ 2479,
+	/*log(1.83228)*/ 2480,
+	/*log(1.83252)*/ 2480,
+	/*log(1.83276)*/ 2481,
+	/*log(1.83301)*/ 2482,
+	/*log(1.83325)*/ 2482,
+	/*log(1.83350)*/ 2483,
+	/*log(1.83374)*/ 2483,
+	/*log(1.83398)*/ 2484,
+	/*log(1.83423)*/ 2484,
+	/*log(1.83447)*/ 2485,
+	/*log(1.83472)*/ 2485,
+	/*log(1.83496)*/ 2486,
+	/*log(1.83521)*/ 2486,
+	/*log(1.83545)*/ 2487,
+	/*log(1.83569)*/ 2488,
+	/*log(1.83594)*/ 2488,
+	/*log(1.83618)*/ 2489,
+	/*log(1.83643)*/ 2489,
+	/*log(1.83667)*/ 2490,
+	/*log(1.83691)*/ 2490,
+	/*log(1.83716)*/ 2491,
+	/*log(1.83740)*/ 2491,
+	/*log(1.83765)*/ 2492,
+	/*log(1.83789)*/ 2492,
+	/*log(1.83813)*/ 2493,
+	/*log(1.83838)*/ 2493,
+	/*log(1.83862)*/ 2494,
+	/*log(1.83887)*/ 2495,
+	/*log(1.83911)*/ 2495,
+	/*log(1.83936)*/ 2496,
+	/*log(1.83960)*/ 2496,
+	/*log(1.83984)*/ 2497,
+	/*log(1.84009)*/ 2497,
+	/*log(1.84033)*/ 2498,
+	/*log(1.84058)*/ 2498,
+	/*log(1.84082)*/ 2499,
+	/*log(1.84106)*/ 2499,
+	/*log(1.84131)*/ 2500,
+	/*log(1.84155)*/ 2501,
+	/*log(1.84180)*/ 2501,
+	/*log(1.84204)*/ 2502,
+	/*log(1.84229)*/ 2502,
+	/*log(1.84253)*/ 2503,
+	/*log(1.84277)*/ 2503,
+	/*log(1.84302)*/ 2504,
+	/*log(1.84326)*/ 2504,
+	/*log(1.84351)*/ 2505,
+	/*log(1.84375)*/ 2505,
+	/*log(1.84399)*/ 2506,
+	/*log(1.84424)*/ 2507,
+	/*log(1.84448)*/ 2507,
+	/*log(1.84473)*/ 2508,
+	/*log(1.84497)*/ 2508,
+	/*log(1.84521)*/ 2509,
+	/*log(1.84546)*/ 2509,
+	/*log(1.84570)*/ 2510,
+	/*log(1.84595)*/ 2510,
+	/*log(1.84619)*/ 2511,
+	/*log(1.84644)*/ 2511,
+	/*log(1.84668)*/ 2512,
+	/*log(1.84692)*/ 2512,
+	/*log(1.84717)*/ 2513,
+	/*log(1.84741)*/ 2514,
+	/*log(1.84766)*/ 2514,
+	/*log(1.84790)*/ 2515,
+	/*log(1.84814)*/ 2515,
+	/*log(1.84839)*/ 2516,
+	/*log(1.84863)*/ 2516,
+	/*log(1.84888)*/ 2517,
+	/*log(1.84912)*/ 2517,
+	/*log(1.84937)*/ 2518,
+	/*log(1.84961)*/ 2518,
+	/*log(1.84985)*/ 2519,
+	/*log(1.85010)*/ 2520,
+	/*log(1.85034)*/ 2520,
+	/*log(1.85059)*/ 2521,
+	/*log(1.85083)*/ 2521,
+	/*log(1.85107)*/ 2522,
+	/*log(1.85132)*/ 2522,
+	/*log(1.85156)*/ 2523,
+	/*log(1.85181)*/ 2523,
+	/*log(1.85205)*/ 2524,
+	/*log(1.85229)*/ 2524,
+	/*log(1.85254)*/ 2525,
+	/*log(1.85278)*/ 2525,
+	/*log(1.85303)*/ 2526,
+	/*log(1.85327)*/ 2527,
+	/*log(1.85352)*/ 2527,
+	/*log(1.85376)*/ 2528,
+	/*log(1.85400)*/ 2528,
+	/*log(1.85425)*/ 2529,
+	/*log(1.85449)*/ 2529,
+	/*log(1.85474)*/ 2530,
+	/*log(1.85498)*/ 2530,
+	/*log(1.85522)*/ 2531,
+	/*log(1.85547)*/ 2531,
+	/*log(1.85571)*/ 2532,
+	/*log(1.85596)*/ 2532,
+	/*log(1.85620)*/ 2533,
+	/*log(1.85645)*/ 2534,
+	/*log(1.85669)*/ 2534,
+	/*log(1.85693)*/ 2535,
+	/*log(1.85718)*/ 2535,
+	/*log(1.85742)*/ 2536,
+	/*log(1.85767)*/ 2536,
+	/*log(1.85791)*/ 2537,
+	/*log(1.85815)*/ 2537,
+	/*log(1.85840)*/ 2538,
+	/*log(1.85864)*/ 2538,
+	/*log(1.85889)*/ 2539,
+	/*log(1.85913)*/ 2539,
+	/*log(1.85938)*/ 2540,
+	/*log(1.85962)*/ 2541,
+	/*log(1.85986)*/ 2541,
+	/*log(1.86011)*/ 2542,
+	/*log(1.86035)*/ 2542,
+	/*log(1.86060)*/ 2543,
+	/*log(1.86084)*/ 2543,
+	/*log(1.86108)*/ 2544,
+	/*log(1.86133)*/ 2544,
+	/*log(1.86157)*/ 2545,
+	/*log(1.86182)*/ 2545,
+	/*log(1.86206)*/ 2546,
+	/*log(1.86230)*/ 2546,
+	/*log(1.86255)*/ 2547,
+	/*log(1.86279)*/ 2548,
+	/*log(1.86304)*/ 2548,
+	/*log(1.86328)*/ 2549,
+	/*log(1.86353)*/ 2549,
+	/*log(1.86377)*/ 2550,
+	/*log(1.86401)*/ 2550,
+	/*log(1.86426)*/ 2551,
+	/*log(1.86450)*/ 2551,
+	/*log(1.86475)*/ 2552,
+	/*log(1.86499)*/ 2552,
+	/*log(1.86523)*/ 2553,
+	/*log(1.86548)*/ 2553,
+	/*log(1.86572)*/ 2554,
+	/*log(1.86597)*/ 2555,
+	/*log(1.86621)*/ 2555,
+	/*log(1.86646)*/ 2556,
+	/*log(1.86670)*/ 2556,
+	/*log(1.86694)*/ 2557,
+	/*log(1.86719)*/ 2557,
+	/*log(1.86743)*/ 2558,
+	/*log(1.86768)*/ 2558,
+	/*log(1.86792)*/ 2559,
+	/*log(1.86816)*/ 2559,
+	/*log(1.86841)*/ 2560,
+	/*log(1.86865)*/ 2560,
+	/*log(1.86890)*/ 2561,
+	/*log(1.86914)*/ 2561,
+	/*log(1.86938)*/ 2562,
+	/*log(1.86963)*/ 2563,
+	/*log(1.86987)*/ 2563,
+	/*log(1.87012)*/ 2564,
+	/*log(1.87036)*/ 2564,
+	/*log(1.87061)*/ 2565,
+	/*log(1.87085)*/ 2565,
+	/*log(1.87109)*/ 2566,
+	/*log(1.87134)*/ 2566,
+	/*log(1.87158)*/ 2567,
+	/*log(1.87183)*/ 2567,
+	/*log(1.87207)*/ 2568,
+	/*log(1.87231)*/ 2568,
+	/*log(1.87256)*/ 2569,
+	/*log(1.87280)*/ 2569,
+	/*log(1.87305)*/ 2570,
+	/*log(1.87329)*/ 2571,
+	/*log(1.87354)*/ 2571,
+	/*log(1.87378)*/ 2572,
+	/*log(1.87402)*/ 2572,
+	/*log(1.87427)*/ 2573,
+	/*log(1.87451)*/ 2573,
+	/*log(1.87476)*/ 2574,
+	/*log(1.87500)*/ 2574,
+	/*log(1.87524)*/ 2575,
+	/*log(1.87549)*/ 2575,
+	/*log(1.87573)*/ 2576,
+	/*log(1.87598)*/ 2576,
+	/*log(1.87622)*/ 2577,
+	/*log(1.87646)*/ 2577,
+	/*log(1.87671)*/ 2578,
+	/*log(1.87695)*/ 2579,
+	/*log(1.87720)*/ 2579,
+	/*log(1.87744)*/ 2580,
+	/*log(1.87769)*/ 2580,
+	/*log(1.87793)*/ 2581,
+	/*log(1.87817)*/ 2581,
+	/*log(1.87842)*/ 2582,
+	/*log(1.87866)*/ 2582,
+	/*log(1.87891)*/ 2583,
+	/*log(1.87915)*/ 2583,
+	/*log(1.87939)*/ 2584,
+	/*log(1.87964)*/ 2584,
+	/*log(1.87988)*/ 2585,
+	/*log(1.88013)*/ 2585,
+	/*log(1.88037)*/ 2586,
+	/*log(1.88062)*/ 2587,
+	/*log(1.88086)*/ 2587,
+	/*log(1.88110)*/ 2588,
+	/*log(1.88135)*/ 2588,
+	/*log(1.88159)*/ 2589,
+	/*log(1.88184)*/ 2589,
+	/*log(1.88208)*/ 2590,
+	/*log(1.88232)*/ 2590,
+	/*log(1.88257)*/ 2591,
+	/*log(1.88281)*/ 2591,
+	/*log(1.88306)*/ 2592,
+	/*log(1.88330)*/ 2592,
+	/*log(1.88354)*/ 2593,
+	/*log(1.88379)*/ 2593,
+	/*log(1.88403)*/ 2594,
+	/*log(1.88428)*/ 2594,
+	/*log(1.88452)*/ 2595,
+	/*log(1.88477)*/ 2596,
+	/*log(1.88501)*/ 2596,
+	/*log(1.88525)*/ 2597,
+	/*log(1.88550)*/ 2597,
+	/*log(1.88574)*/ 2598,
+	/*log(1.88599)*/ 2598,
+	/*log(1.88623)*/ 2599,
+	/*log(1.88647)*/ 2599,
+	/*log(1.88672)*/ 2600,
+	/*log(1.88696)*/ 2600,
+	/*log(1.88721)*/ 2601,
+	/*log(1.88745)*/ 2601,
+	/*log(1.88770)*/ 2602,
+	/*log(1.88794)*/ 2602,
+	/*log(1.88818)*/ 2603,
+	/*log(1.88843)*/ 2604,
+	/*log(1.88867)*/ 2604,
+	/*log(1.88892)*/ 2605,
+	/*log(1.88916)*/ 2605,
+	/*log(1.88940)*/ 2606,
+	/*log(1.88965)*/ 2606,
+	/*log(1.88989)*/ 2607,
+	/*log(1.89014)*/ 2607,
+	/*log(1.89038)*/ 2608,
+	/*log(1.89062)*/ 2608,
+	/*log(1.89087)*/ 2609,
+	/*log(1.89111)*/ 2609,
+	/*log(1.89136)*/ 2610,
+	/*log(1.89160)*/ 2610,
+	/*log(1.89185)*/ 2611,
+	/*log(1.89209)*/ 2611,
+	/*log(1.89233)*/ 2612,
+	/*log(1.89258)*/ 2613,
+	/*log(1.89282)*/ 2613,
+	/*log(1.89307)*/ 2614,
+	/*log(1.89331)*/ 2614,
+	/*log(1.89355)*/ 2615,
+	/*log(1.89380)*/ 2615,
+	/*log(1.89404)*/ 2616,
+	/*log(1.89429)*/ 2616,
+	/*log(1.89453)*/ 2617,
+	/*log(1.89478)*/ 2617,
+	/*log(1.89502)*/ 2618,
+	/*log(1.89526)*/ 2618,
+	/*log(1.89551)*/ 2619,
+	/*log(1.89575)*/ 2619,
+	/*log(1.89600)*/ 2620,
+	/*log(1.89624)*/ 2620,
+	/*log(1.89648)*/ 2621,
+	/*log(1.89673)*/ 2621,
+	/*log(1.89697)*/ 2622,
+	/*log(1.89722)*/ 2623,
+	/*log(1.89746)*/ 2623,
+	/*log(1.89771)*/ 2624,
+	/*log(1.89795)*/ 2624,
+	/*log(1.89819)*/ 2625,
+	/*log(1.89844)*/ 2625,
+	/*log(1.89868)*/ 2626,
+	/*log(1.89893)*/ 2626,
+	/*log(1.89917)*/ 2627,
+	/*log(1.89941)*/ 2627,
+	/*log(1.89966)*/ 2628,
+	/*log(1.89990)*/ 2628,
+	/*log(1.90015)*/ 2629,
+	/*log(1.90039)*/ 2629,
+	/*log(1.90063)*/ 2630,
+	/*log(1.90088)*/ 2630,
+	/*log(1.90112)*/ 2631,
+	/*log(1.90137)*/ 2631,
+	/*log(1.90161)*/ 2632,
+	/*log(1.90186)*/ 2633,
+	/*log(1.90210)*/ 2633,
+	/*log(1.90234)*/ 2634,
+	/*log(1.90259)*/ 2634,
+	/*log(1.90283)*/ 2635,
+	/*log(1.90308)*/ 2635,
+	/*log(1.90332)*/ 2636,
+	/*log(1.90356)*/ 2636,
+	/*log(1.90381)*/ 2637,
+	/*log(1.90405)*/ 2637,
+	/*log(1.90430)*/ 2638,
+	/*log(1.90454)*/ 2638,
+	/*log(1.90479)*/ 2639,
+	/*log(1.90503)*/ 2639,
+	/*log(1.90527)*/ 2640,
+	/*log(1.90552)*/ 2640,
+	/*log(1.90576)*/ 2641,
+	/*log(1.90601)*/ 2641,
+	/*log(1.90625)*/ 2642,
+	/*log(1.90649)*/ 2643,
+	/*log(1.90674)*/ 2643,
+	/*log(1.90698)*/ 2644,
+	/*log(1.90723)*/ 2644,
+	/*log(1.90747)*/ 2645,
+	/*log(1.90771)*/ 2645,
+	/*log(1.90796)*/ 2646,
+	/*log(1.90820)*/ 2646,
+	/*log(1.90845)*/ 2647,
+	/*log(1.90869)*/ 2647,
+	/*log(1.90894)*/ 2648,
+	/*log(1.90918)*/ 2648,
+	/*log(1.90942)*/ 2649,
+	/*log(1.90967)*/ 2649,
+	/*log(1.90991)*/ 2650,
+	/*log(1.91016)*/ 2650,
+	/*log(1.91040)*/ 2651,
+	/*log(1.91064)*/ 2651,
+	/*log(1.91089)*/ 2652,
+	/*log(1.91113)*/ 2652,
+	/*log(1.91138)*/ 2653,
+	/*log(1.91162)*/ 2654,
+	/*log(1.91187)*/ 2654,
+	/*log(1.91211)*/ 2655,
+	/*log(1.91235)*/ 2655,
+	/*log(1.91260)*/ 2656,
+	/*log(1.91284)*/ 2656,
+	/*log(1.91309)*/ 2657,
+	/*log(1.91333)*/ 2657,
+	/*log(1.91357)*/ 2658,
+	/*log(1.91382)*/ 2658,
+	/*log(1.91406)*/ 2659,
+	/*log(1.91431)*/ 2659,
+	/*log(1.91455)*/ 2660,
+	/*log(1.91479)*/ 2660,
+	/*log(1.91504)*/ 2661,
+	/*log(1.91528)*/ 2661,
+	/*log(1.91553)*/ 2662,
+	/*log(1.91577)*/ 2662,
+	/*log(1.91602)*/ 2663,
+	/*log(1.91626)*/ 2663,
+	/*log(1.91650)*/ 2664,
+	/*log(1.91675)*/ 2664,
+	/*log(1.91699)*/ 2665,
+	/*log(1.91724)*/ 2666,
+	/*log(1.91748)*/ 2666,
+	/*log(1.91772)*/ 2667,
+	/*log(1.91797)*/ 2667,
+	/*log(1.91821)*/ 2668,
+	/*log(1.91846)*/ 2668,
+	/*log(1.91870)*/ 2669,
+	/*log(1.91895)*/ 2669,
+	/*log(1.91919)*/ 2670,
+	/*log(1.91943)*/ 2670,
+	/*log(1.91968)*/ 2671,
+	/*log(1.91992)*/ 2671,
+	/*log(1.92017)*/ 2672,
+	/*log(1.92041)*/ 2672,
+	/*log(1.92065)*/ 2673,
+	/*log(1.92090)*/ 2673,
+	/*log(1.92114)*/ 2674,
+	/*log(1.92139)*/ 2674,
+	/*log(1.92163)*/ 2675,
+	/*log(1.92188)*/ 2675,
+	/*log(1.92212)*/ 2676,
+	/*log(1.92236)*/ 2676,
+	/*log(1.92261)*/ 2677,
+	/*log(1.92285)*/ 2678,
+	/*log(1.92310)*/ 2678,
+	/*log(1.92334)*/ 2679,
+	/*log(1.92358)*/ 2679,
+	/*log(1.92383)*/ 2680,
+	/*log(1.92407)*/ 2680,
+	/*log(1.92432)*/ 2681,
+	/*log(1.92456)*/ 2681,
+	/*log(1.92480)*/ 2682,
+	/*log(1.92505)*/ 2682,
+	/*log(1.92529)*/ 2683,
+	/*log(1.92554)*/ 2683,
+	/*log(1.92578)*/ 2684,
+	/*log(1.92603)*/ 2684,
+	/*log(1.92627)*/ 2685,
+	/*log(1.92651)*/ 2685,
+	/*log(1.92676)*/ 2686,
+	/*log(1.92700)*/ 2686,
+	/*log(1.92725)*/ 2687,
+	/*log(1.92749)*/ 2687,
+	/*log(1.92773)*/ 2688,
+	/*log(1.92798)*/ 2688,
+	/*log(1.92822)*/ 2689,
+	/*log(1.92847)*/ 2689,
+	/*log(1.92871)*/ 2690,
+	/*log(1.92896)*/ 2690,
+	/*log(1.92920)*/ 2691,
+	/*log(1.92944)*/ 2692,
+	/*log(1.92969)*/ 2692,
+	/*log(1.92993)*/ 2693,
+	/*log(1.93018)*/ 2693,
+	/*log(1.93042)*/ 2694,
+	/*log(1.93066)*/ 2694,
+	/*log(1.93091)*/ 2695,
+	/*log(1.93115)*/ 2695,
+	/*log(1.93140)*/ 2696,
+	/*log(1.93164)*/ 2696,
+	/*log(1.93188)*/ 2697,
+	/*log(1.93213)*/ 2697,
+	/*log(1.93237)*/ 2698,
+	/*log(1.93262)*/ 2698,
+	/*log(1.93286)*/ 2699,
+	/*log(1.93311)*/ 2699,
+	/*log(1.93335)*/ 2700,
+	/*log(1.93359)*/ 2700,
+	/*log(1.93384)*/ 2701,
+	/*log(1.93408)*/ 2701,
+	/*log(1.93433)*/ 2702,
+	/*log(1.93457)*/ 2702,
+	/*log(1.93481)*/ 2703,
+	/*log(1.93506)*/ 2703,
+	/*log(1.93530)*/ 2704,
+	/*log(1.93555)*/ 2704,
+	/*log(1.93579)*/ 2705,
+	/*log(1.93604)*/ 2705,
+	/*log(1.93628)*/ 2706,
+	/*log(1.93652)*/ 2707,
+	/*log(1.93677)*/ 2707,
+	/*log(1.93701)*/ 2708,
+	/*log(1.93726)*/ 2708,
+	/*log(1.93750)*/ 2709,
+	/*log(1.93774)*/ 2709,
+	/*log(1.93799)*/ 2710,
+	/*log(1.93823)*/ 2710,
+	/*log(1.93848)*/ 2711,
+	/*log(1.93872)*/ 2711,
+	/*log(1.93896)*/ 2712,
+	/*log(1.93921)*/ 2712,
+	/*log(1.93945)*/ 2713,
+	/*log(1.93970)*/ 2713,
+	/*log(1.93994)*/ 2714,
+	/*log(1.94019)*/ 2714,
+	/*log(1.94043)*/ 2715,
+	/*log(1.94067)*/ 2715,
+	/*log(1.94092)*/ 2716,
+	/*log(1.94116)*/ 2716,
+	/*log(1.94141)*/ 2717,
+	/*log(1.94165)*/ 2717,
+	/*log(1.94189)*/ 2718,
+	/*log(1.94214)*/ 2718,
+	/*log(1.94238)*/ 2719,
+	/*log(1.94263)*/ 2719,
+	/*log(1.94287)*/ 2720,
+	/*log(1.94312)*/ 2720,
+	/*log(1.94336)*/ 2721,
+	/*log(1.94360)*/ 2721,
+	/*log(1.94385)*/ 2722,
+	/*log(1.94409)*/ 2723,
+	/*log(1.94434)*/ 2723,
+	/*log(1.94458)*/ 2724,
+	/*log(1.94482)*/ 2724,
+	/*log(1.94507)*/ 2725,
+	/*log(1.94531)*/ 2725,
+	/*log(1.94556)*/ 2726,
+	/*log(1.94580)*/ 2726,
+	/*log(1.94604)*/ 2727,
+	/*log(1.94629)*/ 2727,
+	/*log(1.94653)*/ 2728,
+	/*log(1.94678)*/ 2728,
+	/*log(1.94702)*/ 2729,
+	/*log(1.94727)*/ 2729,
+	/*log(1.94751)*/ 2730,
+	/*log(1.94775)*/ 2730,
+	/*log(1.94800)*/ 2731,
+	/*log(1.94824)*/ 2731,
+	/*log(1.94849)*/ 2732,
+	/*log(1.94873)*/ 2732,
+	/*log(1.94897)*/ 2733,
+	/*log(1.94922)*/ 2733,
+	/*log(1.94946)*/ 2734,
+	/*log(1.94971)*/ 2734,
+	/*log(1.94995)*/ 2735,
+	/*log(1.95020)*/ 2735,
+	/*log(1.95044)*/ 2736,
+	/*log(1.95068)*/ 2736,
+	/*log(1.95093)*/ 2737,
+	/*log(1.95117)*/ 2737,
+	/*log(1.95142)*/ 2738,
+	/*log(1.95166)*/ 2738,
+	/*log(1.95190)*/ 2739,
+	/*log(1.95215)*/ 2739,
+	/*log(1.95239)*/ 2740,
+	/*log(1.95264)*/ 2740,
+	/*log(1.95288)*/ 2741,
+	/*log(1.95312)*/ 2741,
+	/*log(1.95337)*/ 2742,
+	/*log(1.95361)*/ 2743,
+	/*log(1.95386)*/ 2743,
+	/*log(1.95410)*/ 2744,
+	/*log(1.95435)*/ 2744,
+	/*log(1.95459)*/ 2745,
+	/*log(1.95483)*/ 2745,
+	/*log(1.95508)*/ 2746,
+	/*log(1.95532)*/ 2746,
+	/*log(1.95557)*/ 2747,
+	/*log(1.95581)*/ 2747,
+	/*log(1.95605)*/ 2748,
+	/*log(1.95630)*/ 2748,
+	/*log(1.95654)*/ 2749,
+	/*log(1.95679)*/ 2749,
+	/*log(1.95703)*/ 2750,
+	/*log(1.95728)*/ 2750,
+	/*log(1.95752)*/ 2751,
+	/*log(1.95776)*/ 2751,
+	/*log(1.95801)*/ 2752,
+	/*log(1.95825)*/ 2752,
+	/*log(1.95850)*/ 2753,
+	/*log(1.95874)*/ 2753,
+	/*log(1.95898)*/ 2754,
+	/*log(1.95923)*/ 2754,
+	/*log(1.95947)*/ 2755,
+	/*log(1.95972)*/ 2755,
+	/*log(1.95996)*/ 2756,
+	/*log(1.96021)*/ 2756,
+	/*log(1.96045)*/ 2757,
+	/*log(1.96069)*/ 2757,
+	/*log(1.96094)*/ 2758,
+	/*log(1.96118)*/ 2758,
+	/*log(1.96143)*/ 2759,
+	/*log(1.96167)*/ 2759,
+	/*log(1.96191)*/ 2760,
+	/*log(1.96216)*/ 2760,
+	/*log(1.96240)*/ 2761,
+	/*log(1.96265)*/ 2761,
+	/*log(1.96289)*/ 2762,
+	/*log(1.96313)*/ 2762,
+	/*log(1.96338)*/ 2763,
+	/*log(1.96362)*/ 2763,
+	/*log(1.96387)*/ 2764,
+	/*log(1.96411)*/ 2764,
+	/*log(1.96436)*/ 2765,
+	/*log(1.96460)*/ 2765,
+	/*log(1.96484)*/ 2766,
+	/*log(1.96509)*/ 2766,
+	/*log(1.96533)*/ 2767,
+	/*log(1.96558)*/ 2768,
+	/*log(1.96582)*/ 2768,
+	/*log(1.96606)*/ 2769,
+	/*log(1.96631)*/ 2769,
+	/*log(1.96655)*/ 2770,
+	/*log(1.96680)*/ 2770,
+	/*log(1.96704)*/ 2771,
+	/*log(1.96729)*/ 2771,
+	/*log(1.96753)*/ 2772,
+	/*log(1.96777)*/ 2772,
+	/*log(1.96802)*/ 2773,
+	/*log(1.96826)*/ 2773,
+	/*log(1.96851)*/ 2774,
+	/*log(1.96875)*/ 2774,
+	/*log(1.96899)*/ 2775,
+	/*log(1.96924)*/ 2775,
+	/*log(1.96948)*/ 2776,
+	/*log(1.96973)*/ 2776,
+	/*log(1.96997)*/ 2777,
+	/*log(1.97021)*/ 2777,
+	/*log(1.97046)*/ 2778,
+	/*log(1.97070)*/ 2778,
+	/*log(1.97095)*/ 2779,
+	/*log(1.97119)*/ 2779,
+	/*log(1.97144)*/ 2780,
+	/*log(1.97168)*/ 2780,
+	/*log(1.97192)*/ 2781,
+	/*log(1.97217)*/ 2781,
+	/*log(1.97241)*/ 2782,
+	/*log(1.97266)*/ 2782,
+	/*log(1.97290)*/ 2783,
+	/*log(1.97314)*/ 2783,
+	/*log(1.97339)*/ 2784,
+	/*log(1.97363)*/ 2784,
+	/*log(1.97388)*/ 2785,
+	/*log(1.97412)*/ 2785,
+	/*log(1.97437)*/ 2786,
+	/*log(1.97461)*/ 2786,
+	/*log(1.97485)*/ 2787,
+	/*log(1.97510)*/ 2787,
+	/*log(1.97534)*/ 2788,
+	/*log(1.97559)*/ 2788,
+	/*log(1.97583)*/ 2789,
+	/*log(1.97607)*/ 2789,
+	/*log(1.97632)*/ 2790,
+	/*log(1.97656)*/ 2790,
+	/*log(1.97681)*/ 2791,
+	/*log(1.97705)*/ 2791,
+	/*log(1.97729)*/ 2792,
+	/*log(1.97754)*/ 2792,
+	/*log(1.97778)*/ 2793,
+	/*log(1.97803)*/ 2793,
+	/*log(1.97827)*/ 2794,
+	/*log(1.97852)*/ 2794,
+	/*log(1.97876)*/ 2795,
+	/*log(1.97900)*/ 2795,
+	/*log(1.97925)*/ 2796,
+	/*log(1.97949)*/ 2796,
+	/*log(1.97974)*/ 2797,
+	/*log(1.97998)*/ 2797,
+	/*log(1.98022)*/ 2798,
+	/*log(1.98047)*/ 2798,
+	/*log(1.98071)*/ 2799,
+	/*log(1.98096)*/ 2799,
+	/*log(1.98120)*/ 2800,
+	/*log(1.98145)*/ 2800,
+	/*log(1.98169)*/ 2801,
+	/*log(1.98193)*/ 2801,
+	/*log(1.98218)*/ 2802,
+	/*log(1.98242)*/ 2802,
+	/*log(1.98267)*/ 2803,
+	/*log(1.98291)*/ 2803,
+	/*log(1.98315)*/ 2804,
+	/*log(1.98340)*/ 2804,
+	/*log(1.98364)*/ 2805,
+	/*log(1.98389)*/ 2805,
+	/*log(1.98413)*/ 2806,
+	/*log(1.98438)*/ 2807,
+	/*log(1.98462)*/ 2807,
+	/*log(1.98486)*/ 2808,
+	/*log(1.98511)*/ 2808,
+	/*log(1.98535)*/ 2809,
+	/*log(1.98560)*/ 2809,
+	/*log(1.98584)*/ 2810,
+	/*log(1.98608)*/ 2810,
+	/*log(1.98633)*/ 2811,
+	/*log(1.98657)*/ 2811,
+	/*log(1.98682)*/ 2812,
+	/*log(1.98706)*/ 2812,
+	/*log(1.98730)*/ 2813,
+	/*log(1.98755)*/ 2813,
+	/*log(1.98779)*/ 2814,
+	/*log(1.98804)*/ 2814,
+	/*log(1.98828)*/ 2815,
+	/*log(1.98853)*/ 2815,
+	/*log(1.98877)*/ 2816,
+	/*log(1.98901)*/ 2816,
+	/*log(1.98926)*/ 2817,
+	/*log(1.98950)*/ 2817,
+	/*log(1.98975)*/ 2818,
+	/*log(1.98999)*/ 2818,
+	/*log(1.99023)*/ 2819,
+	/*log(1.99048)*/ 2819,
+	/*log(1.99072)*/ 2820,
+	/*log(1.99097)*/ 2820,
+	/*log(1.99121)*/ 2821,
+	/*log(1.99146)*/ 2821,
+	/*log(1.99170)*/ 2822,
+	/*log(1.99194)*/ 2822,
+	/*log(1.99219)*/ 2823,
+	/*log(1.99243)*/ 2823,
+	/*log(1.99268)*/ 2824,
+	/*log(1.99292)*/ 2824,
+	/*log(1.99316)*/ 2825,
+	/*log(1.99341)*/ 2825,
+	/*log(1.99365)*/ 2826,
+	/*log(1.99390)*/ 2826,
+	/*log(1.99414)*/ 2827,
+	/*log(1.99438)*/ 2827,
+	/*log(1.99463)*/ 2828,
+	/*log(1.99487)*/ 2828,
+	/*log(1.99512)*/ 2829,
+	/*log(1.99536)*/ 2829,
+	/*log(1.99561)*/ 2830,
+	/*log(1.99585)*/ 2830,
+	/*log(1.99609)*/ 2831,
+	/*log(1.99634)*/ 2831,
+	/*log(1.99658)*/ 2832,
+	/*log(1.99683)*/ 2832,
+	/*log(1.99707)*/ 2833,
+	/*log(1.99731)*/ 2833,
+	/*log(1.99756)*/ 2834,
+	/*log(1.99780)*/ 2834,
+	/*log(1.99805)*/ 2835,
+	/*log(1.99829)*/ 2835,
+	/*log(1.99854)*/ 2836,
+	/*log(1.99878)*/ 2836,
+	/*log(1.99902)*/ 2837,
+	/*log(1.99927)*/ 2837,
+	/*log(1.99951)*/ 2838,
+	/*log(1.99976)*/ 2838,
+	/*log(2.00000)*/ 2839,
 }
 
 const logApproxLen = 4097
