+++ log_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 )
 
-const logEpsilon = 0.01
+const logEpsilon = 0.001
 
 func TestRange_Log(t *testing.T) {
 	for i := 100; i < 1000000; i++ {
@@ -21,13 +21,13 @@ func TestRange_Log(t *testing.T) {
 }
 
 func BenchmarkFixed_Log(b *testing.B) {
-	for i := 1; i < 100000; i++ {
+	for i := 1; i < b.N+1; i++ {
 		Result = Log(Fixed{int64(i)})
 	}
 }
 
-func BenchmarkFixed_RefLog(b *testing.B) {
-	for i := 1; i < 100000; i++ {
+func BenchmarkFixed_LogRef(b *testing.B) {
+	for i := 1; i < b.N+1; i++ {
 		Result = From(math.Log(Fixed{int64(i)}.Float()))
 	}
 }
